Differential Expression Analysis Using DESeq2

A wrapper function for DESeq2 analysis tailored for RNA-seq exon count data. The function executes the DESeq2 workflow, including size factor estimation, normalization, and differential expression testing between two conditions, such as wild type (WT) and knockout (KO). It produces a series of diagnostic and exploratory plots to visualize data quality and expression differences, including boxplots, principal component analysis (PCA) plots, multi-dimensional scaling (MDS) plots, MA plots, histograms of p-values, and heatmaps for top differentially expressed genes.

DESeqCalculation(dat, genotypes, fc = 1.15)

Arguments

dat

A matrix with RNAseq exon count data where rows are genes and columns are samples.

genotypes

A factor vector indicating the genotype of each sample, used in the design formula for DESeq2.

fc

The fold change threshold for determining significant differential expression. Default is 1.15.

Value

A list containing the DESeq2 results object, various ggplot and base R plot objects for visualization, and a pheatmap object for heatmap visualization.

Examples

if (FALSE) {
# Simulate RNA-seq count data
set.seed(123) # for reproducibility
counts <- matrix(rpois(2000, lambda = 10), ncol = 10)
colnames(counts) <- paste0("Sample", 1:10)
rownames(counts) <- paste0("Gene", 1:200)
genotypes <- factor(rep(c("WT", "KO"), each = 5))

# Perform the DESeq2 analysis
deseq_results <- DESeqCalculation(dat = counts, genotypes = genotypes)

# Explore the results
print(deseq_results$plots$PCAplot)
}