Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeans.between.ABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeans.between.ABC(dat = dat, A.samples = 1:3, B.samples = 4:6,
 C.samples = 7:9)
#>              V1           V2           V3           V4            V5
#> 1   0.278337419  0.071787007 -0.188408946  0.099779624 -0.4071761667
#> 2   0.738444430  0.393048321  0.302855273 -0.022088108 -0.0014620628
#> 3   0.091558421  0.076382408  0.532534821 -0.044560970  0.3684405822
#> 4  -0.323070043 -0.104231899 -0.126668842 -0.099204907  0.0008049337
#> 5  -0.108797862 -0.095129538 -0.054498853 -0.164928179 -0.1360997849
#> 6  -0.361258619 -0.354101139 -0.290433019 -0.327844989 -0.2274910956
#> 7   0.257826772  0.235637114 -0.119719847  0.233095064 -0.2239947350
#> 8  -0.059650136 -0.195015061  0.309691198  0.019592286 -0.7095818686
#> 9   0.293826484 -0.279761784  0.098390232 -0.009954467  0.5535104645
#> 10  0.331481988 -0.126656571 -0.170528488  0.298331052  0.1088325445
#> 11 -0.017568635  0.035391079 -0.076762664 -0.074525861 -0.0333863562
#> 12  0.034187777  0.290467914  0.595196802 -0.202401551 -0.2275867513
#> 13 -0.416964954 -0.083556392  0.077950875  0.173301098  0.4497504038
#> 14 -0.276372041 -0.173335445 -0.357790292 -0.010849587 -0.0326717751
#> 15  0.022762523  0.130881967 -0.085201182 -0.059613905  0.3604154942
#> 16  0.211078751 -0.550714724 -0.040945234 -0.184442354  0.1243959407
#> 17 -0.382488384  0.126966405  0.146227406  0.511938158  0.4389206925
#> 18 -0.001055016  0.342885306 -0.123092919 -0.273366596  0.1824839623
#> 19 -0.069153394  0.228276994 -0.061278803 -0.202429360  0.0769912993
#> 20 -0.087469305 -0.084352998  0.119242340  0.272363820  0.1041946594
#> 21  0.283436829  0.030342153  0.108252496  0.412924876 -0.1045082148
#> 22  0.147500003 -0.232306897 -0.045752932 -0.205226811  0.2638553656
#> 23  0.175089737 -0.202771089  0.380193775  0.099307436  0.2821986490
#> 24  0.082849376 -0.072611294 -0.184618835 -0.492720277  0.3384513519
#> 25  0.194751418 -0.040107654  0.300820012 -0.259593098  0.3187429062
#> 26 -0.122385373 -0.442251304 -0.644009426 -0.042073426 -0.6600202502
#> 27 -0.242655999 -0.340755284 -0.160289633 -0.010085024  0.0686160175
#> 28  0.048967585 -0.314091522  0.083793721  0.222472385 -0.5831127283
#> 29  0.202609379  0.277599896 -0.008386139 -0.083246271 -0.0201104391
#> 30  0.393734686 -0.561564125 -0.310099458  0.390302581 -0.1050045387
#> 31 -0.239051987  0.394780854 -0.205700498 -0.215220919  0.3708878253
#> 32 -0.743088345 -0.221975786 -0.062360932  0.014086346  0.0372554020
#> 33  0.387476723  0.177126606 -0.153483995 -0.042761759  0.7118180756
#> 34  0.167225334 -0.359633647 -0.176536442 -0.058337308 -0.0039713272
#> 35 -0.335788652 -0.046084443 -0.260671978 -0.429937489  0.2559220463
#> 36 -0.105206244 -0.024860580  0.541799302  0.236722052  0.3689062967
#> 37 -0.322117524  0.016326756  0.408393842  0.253138306 -0.3684419856
#> 38  0.030511479  0.026041377 -0.103767623  0.346625835  0.0102505220
#> 39  0.523526444  0.003498961 -0.426848772  0.139254647 -0.0381421697
#> 40  0.334057264  0.244176079  0.201847491 -0.366624983  0.1713788463
#> 41  0.466913976  0.396010359 -0.501089312  0.082265108  0.6105124066
#> 42  0.507686856  0.047879575 -0.337767265 -0.633603927  0.4544124517
#> 43  0.257199620 -0.389491208 -0.294061689 -0.162630185  0.0750325281
#> 44  0.236201535  0.731051206 -0.302968938 -0.481693206 -0.1925464674
#> 45 -0.077637621  0.171894881  0.649254531 -0.144490377  0.5693019137
#> 46 -0.308414128 -0.245077899 -0.096678635 -0.178172307 -0.3621727158
#> 47  0.235948819  0.036735380  0.437064889 -0.184338103  0.0464822115
#> 48 -0.010479974  0.348177064  0.236273972  0.359646021 -0.1274544225
#> 49 -0.043601241 -0.187129246  0.083470156 -0.672123706  0.5866397365
#> 50  0.034212630 -0.009931511  0.033463208  0.006753624  0.1881577383
#> 51 -0.100291325  0.193604750  0.161167474 -0.347596580  0.2559313299
#> 52  0.141572610  0.289303524 -0.156968030  0.017201845 -0.4157956451
#> 53 -0.387065754  0.374626623  0.488725549 -0.108733853  0.6801585335
#> 54 -0.261910444  0.254371962  0.173260205  0.437154846  0.1862485989
#> 55  0.406849870  0.657790980 -0.126445375 -0.389744546  0.1877608998
#> 56  0.036660112 -0.896900349  0.055633950 -0.237858878 -0.4835206911
#> 57  0.388817593  0.041816298  0.002172366  0.261575125  0.1219998642
#> 58 -0.803019979 -0.275682678 -0.466523057  0.365305037  0.0456448803
#> 59 -0.271248148 -0.373084057  0.205818174 -0.030775158 -0.0851353834
#> 60  0.758505575  0.032985727  0.387837800  0.253973496 -0.4950296275
#> 61 -0.497250486 -0.054582802 -0.465647402 -0.443635580  0.1876879042
#> 62  0.016303848 -0.117589632  0.464523227  0.166313445  0.0421690782
#> 63 -0.312049135  0.333923920  0.169853733 -0.240762121 -0.2707993686
#> 64  0.448229921 -0.256734908 -0.124559184  0.154623268 -0.2040822828
#> 65  0.651234598 -0.067349544 -0.271711355  0.206328826 -0.0053753011
#> 66 -0.241724445  0.624748164  0.334313790  0.047028138  0.1924893918
#> 67 -0.300207725 -0.631271152  0.028500180 -0.002921422 -0.4601888556
#>               V6          V7           V8           V9      Mean.A       Mean.B
#> 1  -0.0927101153  0.03696020  0.217856939  0.694155599  0.05390516 -0.133368886
#> 2   0.0948994716 -0.34414334 -0.173592899 -0.211163345  0.47811601  0.023783100
#> 3   0.2949281444  0.46640333 -0.218989489  0.175709889  0.23349188  0.206269252
#> 4   0.2505522252  0.09202550 -0.046568417 -0.221454173 -0.18465693  0.050717417
#> 5  -0.1031499765  0.16541434 -0.354284449 -0.177651016 -0.08614208 -0.134725980
#> 6   0.2806982062  0.38024963 -0.044579985  0.544825331 -0.33526426 -0.091545959
#> 7  -0.4049906773  0.03312619  0.047499800  0.113361922  0.12458135 -0.131963450
#> 8   0.1067367976  0.25144652  0.355130908  0.123512175  0.01834200 -0.194417595
#> 9  -0.0190539212 -0.08717085 -0.250673094 -0.233708859  0.03748498  0.174834025
#> 10  0.2169806917 -0.20714871  0.326067391  0.297313307  0.01143231  0.208048096
#> 11  0.1963851596  0.03120017  0.487735989  0.343901313 -0.01964674  0.029490981
#> 12 -0.0459241206 -0.22187784  0.112032758  0.091687935  0.30661750 -0.158637474
#> 13 -0.3546772401 -0.05386214 -0.315591867  0.302576799 -0.14085682  0.089458087
#> 14 -0.2670413645 -0.08519226  0.133340428 -0.563394400 -0.26916593 -0.103520909
#> 15  0.1206968670 -0.10066455  0.100690640  0.617630621  0.02281444  0.140499485
#> 16  0.0737999786 -0.54851003 -0.045900733 -0.650590303 -0.12686040  0.004584522
#> 17 -0.3395042616 -0.48275339  0.061439815 -0.173865630 -0.03643152  0.203784863
#> 18 -0.2454875024  0.49263746  0.374457206  0.175462401  0.07291246 -0.112123379
#> 19 -0.0457427907  0.21401441  0.088668539  0.009800545  0.03261493 -0.057060284
#> 20 -0.3395259190 -0.45016375 -0.165822998  0.129178444 -0.01752665  0.012344187
#> 21  0.2981840174  0.10124700 -0.095832643 -0.057325811  0.14067716  0.202200226
#> 22 -0.1598057519  0.03595538  0.334288565 -0.346542220 -0.04351994 -0.033725733
#> 23 -0.0393387481  0.54644559  0.303094653  0.164973022  0.11750414  0.114055779
#> 24 -0.1328470470 -0.02496533 -0.048021438  0.849698747 -0.05812692 -0.095705324
#> 25 -0.0128089347 -0.29755537  0.393300396  0.107657897  0.15182126  0.015446958
#> 26  0.1628288208 -0.48754759 -0.267054381  0.203616949 -0.40288203 -0.179754952
#> 27  0.0534283419 -0.13157454 -0.619621858 -0.165557325 -0.24790031  0.037319778
#> 28 -0.2771381434 -0.53194287  0.391401550 -0.262443963 -0.06044341 -0.212592829
#> 29 -0.0373822391  0.10485517  0.153527750 -0.139376194  0.15727438 -0.046912983
#> 30  0.2708740947  0.03427758  0.099574358 -0.159670984 -0.15930963  0.185390712
#> 31 -0.3961157996  0.38376706  0.261075833  0.494677600 -0.01665721 -0.080149631
#> 32  0.0241498435  0.21356142 -0.487366385 -0.053428870 -0.34247502  0.025163864
#> 33 -0.1831470653 -0.36038559 -0.046479858 -0.489646274  0.13703978  0.161969750
#> 34  0.4520677499 -0.07140337  0.173009216  0.321321014 -0.12298159  0.129919705
#> 35  0.1428940033 -0.66934819  0.316187026  0.061567496 -0.21418169 -0.010373813
#> 36 -0.5601750113 -0.56664554 -0.227470922  0.175460312  0.13724416  0.015151112
#> 37  0.3303241081  0.09753483 -0.094899867 -0.091937999  0.03420102  0.071673476
#> 38  0.2191643625 -0.38602932  0.440910121  0.260297069 -0.01573826  0.192013573
#> 39  0.1462889191 -0.27436629  0.200957438 -0.409031169  0.03339221  0.082467132
#> 40 -0.0492669773  0.31283364 -0.049730606  0.250619338  0.26002694 -0.081504371
#> 41 -0.3304384362  0.21212301  0.142172123 -0.497354448  0.12061167  0.120779693
#> 42 -0.1110113437 -0.14561885 -0.072164848 -0.022106432  0.07259972 -0.096734273
#> 43  0.0932247329  0.19080714 -0.117374212 -0.432475579 -0.14211776  0.001875692
#> 44  0.4944763084 -0.44360401 -0.330308381  0.346152242  0.22142793 -0.059921122
#> 45  0.1268128213  0.11682072  0.069948268  0.601597291  0.24783726  0.183874786
#> 46 -0.0503305260  0.12393403 -0.171094516 -0.208032568 -0.21672355 -0.196891850
#> 47  0.3514987919  0.23760401 -0.137461572  0.619742029  0.23658303  0.071214300
#> 48 -0.2696241323 -0.19520276  0.347692296 -0.328902077  0.19132369 -0.012477511
#> 49  0.4754054331  0.16161646  0.158791862  0.096256861 -0.04908678  0.129973821
#> 50  0.1561016685  0.37094877 -0.180310500  0.307917556  0.01924811  0.117004343
#> 51 -0.3952251341  0.01709686  0.048014152 -0.107946418  0.08482697 -0.162296795
#> 52  0.2988917503 -0.19763620  0.343055981  0.227120488  0.09130270 -0.033234017
#> 53  0.0576267661  0.09585308 -0.220086985  0.296885260  0.15876214  0.209683815
#> 54 -0.2717187577 -0.12233314  0.027447568 -0.105495240  0.05524057  0.117228229
#> 55 -0.1610051700  0.08200893  0.005006185  0.225190890  0.31273183 -0.120996272
#> 56 -0.1402091607  0.71606320  0.742328286  0.247615825 -0.26820210 -0.287196243
#> 57  0.6118610772  0.36081344  0.016958822  0.386294661  0.14426875  0.331812022
#> 58 -0.1567409593  0.06127502 -0.142559866 -0.004062544 -0.51507524  0.084736319
#> 59  0.1678239644 -0.23107105  0.095115950 -0.479912895 -0.14617134  0.017304474
#> 60  0.0006578945  0.32418739  0.473008368 -0.162291875  0.39310970 -0.080132746
#> 61 -0.0243125432 -0.55722051 -0.012409596 -0.334152665 -0.33916023 -0.093420073
#> 62  0.1352556218 -0.51653662  0.135934965  0.064503374  0.12107915  0.114579382
#> 63 -0.0027417702  0.46570916 -0.469706205  0.146260485  0.06390951 -0.171434420
#> 64 -0.3462956263  0.19597620 -0.032024318 -0.266050210  0.02231194 -0.131918214
#> 65  0.1237611791 -0.18752037 -0.150734476  0.278337419  0.10405790  0.108238235
#> 66  0.4535285736  0.24214634  0.408849950  0.738444430  0.23911250  0.231015368
#> 67 -0.1293727094  0.12834310 -0.217306862  0.091558421 -0.30099290 -0.197494329
#>           Mean.C      FC.crude  logFC.crude
#> 1   0.3163242456 -7.136449e-01 -0.635350050
#> 2  -0.2429665275  6.300706e-01  0.968182770
#> 3   0.1410412430  2.944558e-01  0.100133555
#> 4  -0.0586656960  1.868180e+00  0.133752683
#> 5  -0.1221737092  1.348368e+00 -0.018909372
#> 6   0.2934983242  3.876158e-01 -0.389116570
#> 7   0.0646626362  4.281547e+00 -0.338539697
#> 8   0.2433632012 -9.455091e-01 -0.637930544
#> 9  -0.1905175995 -6.024013e-01  0.559007143
#> 10  0.1387439957  1.544366e+00  0.086073545
#> 11  0.2876124899  1.599227e-01 -0.317341183
#> 12 -0.0060523807  1.488007e+00 -0.362152022
#> 13 -0.0222924015  1.942530e+00  0.137379286
#> 14 -0.1717487426  1.700368e+00  0.087016429
#> 15  0.2058855688  6.428378e-01 -0.082023108
#> 16 -0.4150003546 -4.561843e-01  0.668500820
#> 17 -0.1983930675 -1.483169e+00  0.565503503
#> 18  0.3475190234 -6.738216e-01 -0.645243471
#> 19  0.1041611631 -1.253388e+00 -0.235240831
#> 20 -0.1622694350 -2.063719e-01  0.268033134
#> 21 -0.0173038188 -3.894334e-01  0.334210991
#> 22  0.0079005753  1.904728e-01 -0.058278483
#> 23  0.3381710888 -1.562700e-02 -0.292653155
#> 24  0.2589039921 -1.185323e-01 -0.452548287
#> 25  0.0678009732  1.623112e+00 -0.084909474
#> 26 -0.1836616749  1.017821e+00  0.004615408
#> 27 -0.3055845728 -4.944504e+00  0.447732995
#> 28 -0.1343284272  2.059273e+00 -0.127381287
#> 29  0.0396689080  1.736206e+00 -0.148995036
#> 30 -0.0086063480  2.287278e+00  0.224768842
#> 31  0.3798401648 -1.601333e-01 -0.576450287
#> 32 -0.1090779451  1.575165e+00  0.149050016
#> 33 -0.2988372419 -5.719497e-02  0.861443729
#> 34  0.1409756205  9.581147e-01 -0.012674862
#> 35 -0.0971978890  1.742189e+00  0.107996898
#> 36 -0.2062187153  3.554767e-01  0.419191447
#> 37 -0.0297676787 -5.857935e-01  0.148444355
#> 38  0.1050592888  1.719835e+00  0.107798337
#> 39 -0.1608133394 -2.526958e-01  0.380633941
#> 40  0.1712407916  3.846673e+00 -0.468674841
#> 41 -0.0476864388 -9.983378e-04  0.266103969
#> 42 -0.0799633784  1.109928e+00 -0.028837435
#> 43 -0.1196808826  6.417714e+00  0.162067017
#> 44 -0.1425867176  7.729059e-01  0.176297681
#> 45  0.2627887600 -4.277998e+00 -0.116772516
#> 46 -0.0850643517  1.506291e-01 -0.150108464
#> 47  0.2399614905 -4.894794e+01 -0.265654994
#> 48 -0.0588041793  8.147881e-01  0.086484088
#> 49  0.1388883932  9.525758e-01 -0.010866815
#> 50  0.1661852765  6.652928e-01 -0.063228631
#> 51 -0.0142784697  2.493544e+00 -0.258945542
#> 52  0.1241800901 -3.787914e+00 -0.238550428
#> 53  0.0575504514 -5.031205e-01  0.225601882
#> 54 -0.0667936040 -5.079532e-01  0.274530312
#> 55  0.1040686683  2.078604e+00 -0.482796980
#> 56  0.5686691048 -2.269662e-02 -0.904916821
#> 57  0.2546889760  1.698450e+00  0.096874332
#> 58 -0.0284491310  1.232592e+00  0.105860130
#> 59 -0.2052893330 -2.765247e+00  0.306355504
#> 60  0.2116346284  2.607755e+00 -0.635887256
#> 61 -0.3012609237  6.484028e+00  0.263336690
#> 62 -0.1053660925  2.870348e-02  0.361016884
#> 63  0.0474211459  1.427334e+01 -0.363131126
#> 64 -0.0340327751  2.737260e+00 -0.157994817
#> 65 -0.0199724754 -3.370412e-02  0.197065629
#> 66  0.4631469082 -3.614237e-02 -0.303373358
#> 67  0.0008648864  3.428720e-01 -0.238487097