Calculate Log2 Fold Change Between Two Sets of Samples

The function calculates the average expression values for two sets of samples (A and B) and then computes the fold change and log2 fold change between them. This is typically used to analyze the differential expression of genes between two conditions in RNA-seq data.

logofMeansBetweenAB(dat, A.samples, B.samples)

Arguments

dat

A dataframe or matrix where rows represent genes and columns represent samples. The differential expression results from `DESeqCalculation` can be used as input.

A.samples

A vector of column indices or names for the first sample type (e.g., control group).

B.samples

A vector of column indices or names for the second sample type (e.g., treatment group).

Value

A dataframe with additional columns for the mean expression of samples A and B, crude fold change, and log2 fold change. Rows with `NA` values in the log2 fold change are removed from the output.

Examples

# Simulate RNA-seq count data for 100 genes and 6 samples
set.seed(123) # for reproducibility
dat <- as.data.frame(matrix(rnbinom(n=600, mu=100, size=0.5), ncol=6))
colnames(dat) <- c(paste("Sample_A", 1:3, sep=""), paste("Sample_B", 1:3, sep=""))
rownames(dat) <- paste("Gene", 1:100, sep="")

# Calculate log2 fold change between the sample groups
logFC_results <- logofMeansBetweenAB(dat = dat, A.samples = 1:3, B.samples = 4:6)
head(logFC_results)
#>       Sample_A1 Sample_A2 Sample_A3 Sample_B1 Sample_B2 Sample_B3    Mean.A
#> Gene1        22       174         4        42       121       190  66.66667
#> Gene2       393         0       151       195         0       268 181.33333
#> Gene3        89       170       173        58        21        67 144.00000
#> Gene4         2        44        48         5        51        26  31.33333
#> Gene5       498       420         9       199       161       131 309.00000
#> Gene6        88        64        65       412        24       132  72.33333
#>          Mean.B  FC.crude logFC.crude
#> Gene1 117.66667 1.7650000   0.8103975
#> Gene2 154.33333 0.8511029  -0.2312109
#> Gene3  48.66667 0.3379630  -1.5457031
#> Gene4  27.33333 0.8723404  -0.1905219
#> Gene5 163.66667 0.5296656  -0.9127197
#> Gene6 189.33333 2.6175115   1.3759872