Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeansBetweenABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeansBetweenABC(dat = dat, A.samples = 1:3, B.samples = 4:6,
 C.samples = 7:9)
#>              V1           V2          V3            V4           V5
#> 1   0.065642249  0.038567357  0.48616140  0.1991035650  0.025782863
#> 2  -0.097226542  0.192954043  0.25764938  0.1945209568  0.280274066
#> 3  -0.278957239  0.185431597  0.10869907 -0.6262258746 -0.340523679
#> 4   0.103799095  0.216805259 -0.44516590 -0.5293846659  0.200879098
#> 5   0.321349880  0.253100464  0.12550070  0.0248505293  0.293554384
#> 6   0.041091186  0.575841961 -0.40216391  1.0410126940 -0.153045450
#> 7   0.468380370  0.165583680 -0.05957384 -0.2024006312  0.236827787
#> 8   0.119696964 -0.099265372  0.04916165 -0.1421227474  1.135345654
#> 9   0.341600187 -0.376296542 -0.50984036  0.3721211361 -0.733372237
#> 10 -0.517602781  0.420196703 -0.13075408  0.0022724734  0.437365341
#> 11 -0.043117530 -0.085703282  0.33631827 -0.2467026887  0.272562597
#> 12 -0.282197258 -0.335078128  0.78370625  0.1249984096 -0.204617595
#> 13  0.288172476 -0.444935644 -0.38230012 -0.0095430617  0.268872860
#> 14  0.550141494 -0.342150852  0.11020453  0.1561071754  0.321918018
#> 15 -0.133703128 -0.033111572  0.25114553 -0.4485670624 -0.222732274
#> 16 -0.229529516 -0.167107234 -0.12423142  0.3043566669  0.070932144
#> 17 -0.517353865  0.132489963 -0.33004638 -0.0816822698  0.460285153
#> 18 -0.133901198  0.140723836  0.26579075 -0.2619598598 -0.595356792
#> 19 -0.100173618  0.014164362 -0.68735870 -0.0279234862  0.376257526
#> 20 -0.200523767 -0.268608071  0.18783015  0.1589720450 -0.418900313
#> 21 -0.632628490 -0.031112426  0.38444530 -0.1411692369 -0.041919595
#> 22 -0.005385756  0.037270917  0.35450924 -0.2654137914  0.092274594
#> 23 -0.255097049 -0.551319703  0.24460699  0.5654685105  0.026476680
#> 24 -0.523536071 -0.228964307 -0.11841128 -0.2595713490  0.129812068
#> 25  0.456703887  0.456759803 -0.10255805  0.2127472812  0.005675926
#> 26  0.466047393  0.220808309  0.08465097  0.3604323832  0.314151067
#> 27  0.342629520  0.026488025  0.10583021 -0.1295649025 -0.690711798
#> 28  0.431686985 -0.634197144  0.13704566  0.0613047210 -0.421508004
#> 29  0.252916766  0.252584115 -0.10967945  0.1094917977  0.693942999
#> 30 -0.229462009 -0.104051199  0.13522039  0.3176903165 -0.276127266
#> 31 -0.095414516  0.133910973 -0.37516255  0.0809807594  0.074901000
#> 32  0.315966308  0.071193497 -0.00757344  0.4835584697  0.060512983
#> 33 -0.676017563  0.303076314 -0.30361105  0.0126128597  0.162673350
#> 34 -0.584926270 -0.086767504  0.15307029  0.4300729976 -0.381762456
#> 35 -0.145471924 -0.112852496  0.26454634 -0.0638027746 -0.013684077
#> 36 -0.274988708 -0.205958703  0.12164054 -0.1352264382 -0.216780088
#> 37 -0.246183661 -0.096148052 -0.03496997  0.1236523968  0.059099934
#> 38 -0.407683861 -0.058532705 -0.18913413 -0.0707792702  0.526231738
#> 39 -0.086444144 -0.206767117 -0.25807736 -0.3944801276 -0.151539186
#> 40  0.459368647  0.124950326  0.36961693 -0.0055367277  0.079662609
#> 41 -0.279320605  0.047263996  0.27786245 -0.0740508844  0.172178613
#> 42  0.138720746  0.882181239 -0.09924124  0.0965206600  0.144714036
#> 43  0.039736900 -0.364210829  0.09095292 -0.4309048531  0.037831759
#> 44  0.041430961 -0.033491871 -0.23798037  0.2856731276 -0.038709747
#> 45  0.320863015  0.301027352  0.23321918  0.3949824477  0.050170785
#> 46 -0.325214579 -0.502276777 -0.06256003 -0.0965456795 -0.342839158
#> 47 -0.151147377 -0.183222071  0.27463172  0.1354301341  0.196978213
#> 48  0.222799776  0.117727789  0.21472151 -0.5605983794 -0.095585530
#> 49 -0.521635331 -0.497279659 -0.37495719  0.1244568654  0.035792514
#> 50 -0.110499327  0.330583771 -0.28475155 -0.0516535001  0.025760461
#> 51  0.415021805 -0.391379567 -0.49184263  0.2299998002  0.125649513
#> 52  0.029844659 -0.334704095 -0.04508369 -0.0181582484  0.059487455
#> 53 -0.229863848  0.509023563 -0.22802288 -0.2880028048  0.144929687
#> 54 -0.184750790 -0.458308089 -0.23186436 -0.0266351649  0.241153625
#> 55 -0.485628043  0.009466434 -0.30832771 -0.0003078786  0.368866810
#> 56  0.200633090  0.003984856  0.14092648  0.5553842837 -0.125899883
#> 57 -0.338153090 -0.296813509  0.05636322 -0.2459701612 -0.331028632
#> 58  0.548297083  0.019097138 -0.73026775  0.1776893628  0.232426371
#> 59  0.048415331  0.430918879 -0.28760255  0.3262080514  0.073872225
#> 60 -0.031524077  0.135850369  0.02398801  0.3286076798  0.589435822
#> 61  0.456118738 -0.405990685 -0.18311940 -0.2085236750  0.206098036
#> 62  0.582170659  0.077324613  0.02149973 -0.1529159587 -0.059312489
#> 63  0.253066831  0.196874153 -0.13797999 -0.1564282776  0.198945708
#> 64 -0.165330686 -0.097199006  0.08893333  0.1452429238 -0.069957406
#> 65  0.466548914 -0.198808863  0.32842011  0.2413523441 -0.421604191
#> 66 -0.152405149  0.236637094  0.21868486 -0.5872338684  0.133694794
#> 67  0.272086155 -0.136623245 -0.48545594 -0.0272847246 -0.080551444
#>              V6           V7          V8           V9       Mean.A       Mean.B
#> 1   0.366700203 -0.421108059  0.04204554 -0.231502580  0.196790336  0.197195544
#> 2  -0.520403586  0.602352785  0.13680287 -0.344963031  0.117792293 -0.015202855
#> 3   0.237559070 -0.070543696 -0.18155784 -0.418070358  0.005057811 -0.243063494
#> 4   0.372176933  0.362608193 -0.27512841 -0.257997203 -0.041520514  0.014557122
#> 5  -0.189764668  0.336203006  0.08751127 -0.486113189  0.233317015  0.042880082
#> 6  -0.040750651  0.042785596  0.13278847  0.102846285  0.071589745  0.282405531
#> 7  -0.231946110  0.088836757  0.31732541  0.284583619  0.191463404 -0.065839651
#> 8  -0.537821599  0.479119405  0.01857549  0.404366588  0.023197747  0.151800436
#> 9   0.336691738  0.013551785  0.32119050  0.390247436 -0.181512238 -0.008186454
#> 10  0.045395353 -0.553828055 -0.33572116 -0.486243113 -0.076053385  0.161677723
#> 11 -0.011843567 -0.053387945 -0.16188286 -0.563981183  0.069165818  0.004672114
#> 12  0.195063025  0.191765475  0.10358820 -0.322899490  0.055476954  0.038481280
#> 13 -0.033847097  0.432126890  0.13542730 -0.086571005 -0.179687764  0.075160900
#> 14 -0.286440063  0.138135525 -0.29931391  0.378334056  0.106065059  0.063861710
#> 15  0.290747115 -0.444647534  0.38821932 -0.059864615  0.028110276 -0.126850740
#> 16 -0.527488975  0.035920910  0.20045284 -0.799327314 -0.173622723 -0.050733388
#> 17 -0.022393121  0.575516304 -0.01552755 -0.457543129 -0.238303427  0.118736587
#> 18 -0.457931534  0.510849832 -0.40352611 -0.288622075  0.090871128 -0.438416062
#> 19  0.281050249 -0.409509099  0.23844958 -0.350902362 -0.257789319  0.209794763
#> 20 -0.227991630 -0.448674076 -0.19521582  0.197179552 -0.093767231 -0.162639966
#> 21  0.022576446  0.467312102 -0.19101478  0.253010009 -0.093098539 -0.053504129
#> 22  0.506381673 -0.028105147 -0.06454547  0.159347199  0.128798132  0.111080825
#> 23  0.002748493  0.102619082 -0.01294285 -0.117916543 -0.187269920  0.198231228
#> 24  0.478031286 -0.172527459 -0.13894383  0.054168324 -0.290303886  0.116090668
#> 25  0.150716428  0.130574760 -0.63621441  0.337934625  0.270301880  0.123046545
#> 26  0.028125365  0.032913929 -0.34892908 -0.018055595  0.257168892  0.234236272
#> 27 -0.225472768 -0.071180798 -0.10715199 -0.239208150  0.158315920 -0.348583156
#> 28  0.306072687  0.343559051  0.22185790 -0.054923589 -0.021821501 -0.018043532
#> 29 -0.391936995 -0.474743155 -0.22655086  0.132391646  0.131940476  0.137165934
#> 30 -0.227142516 -0.676073894 -0.33335486  0.502476790 -0.066097605 -0.061859822
#> 31 -0.330015029  0.154983107 -0.32693329  0.166374179 -0.112222032 -0.058044423
#> 32  0.161236671  0.107100667  0.26302603  0.308894414  0.126528788  0.235102708
#> 33  0.305266886 -0.030609088  0.69055778 -0.007815330 -0.225517432  0.160184365
#> 34 -0.706117885 -0.112647948  0.13800811  0.426906960 -0.172874496 -0.219269114
#> 35 -0.269483781  0.213966880 -0.06678413  0.405639987  0.002073973 -0.115656878
#> 36 -0.244749825  0.382582259  0.39637148  0.203476046 -0.119768956 -0.198918784
#> 37 -0.056422487  0.382841762 -0.16125169  0.514841619 -0.125767227  0.042109948
#> 38  0.379692466 -0.517776263  0.49186494 -0.060619173 -0.218450233  0.278381645
#> 39  0.062498346 -0.409273429 -0.24756268  0.142000885 -0.183762873 -0.161173656
#> 40 -0.114808451 -0.242592149 -0.21886685 -0.229318798  0.317978635 -0.013560857
#> 41  0.553389985 -0.419408435  0.23053521  0.480344187  0.015268613  0.217172571
#> 42 -0.186073350  0.207372857  0.02412271  0.195970484  0.307220249  0.018387116
#> 43 -0.112016463 -0.609148610 -0.22571791 -0.046940005 -0.077840336 -0.168363186
#> 44 -0.432099836 -0.157281050 -0.18808483  0.219663022 -0.076680426 -0.061712152
#> 45 -0.231184755 -0.006774796  0.33323609 -0.036678542  0.285036515  0.071322826
#> 46  0.174032322  0.509201399  0.07762426 -0.257867093 -0.296683794 -0.088450839
#> 47 -0.024891286 -0.395281247  0.62601223 -0.031799577 -0.019912576  0.102505687
#> 48 -0.222047280 -0.178688076 -0.18859896 -0.396306855  0.185083024 -0.292743730
#> 49 -0.303847306 -0.138410830  0.08316775  0.082581086 -0.464624060 -0.047865976
#> 50 -0.377282820 -0.491007326  0.09548567 -0.004868864 -0.021555702 -0.134391953
#> 51 -0.191796285  0.264024278 -0.13817319  0.439494279 -0.156066799  0.054617676
#> 52  0.060347686 -0.197699096  0.22000205 -0.243283307 -0.116647710  0.033892297
#> 53 -0.196967747 -0.363200305  0.07787897 -0.248042352  0.017045613 -0.113346955
#> 54  0.215377992 -0.278449231  0.01070831  0.257671309 -0.291641081  0.143298818
#> 55  0.116854472 -0.265521586  0.10997369 -0.927418846 -0.261496439  0.161804468
#> 56 -0.005076019  0.098085385  0.10598383 -0.009245223  0.115181475  0.141469461
#> 57 -0.059537032 -0.552905515 -0.41910674  0.362413718 -0.192867793 -0.212178609
#> 58  0.584022938  0.054035871 -0.25162410  0.256031918 -0.054291175  0.331379557
#> 59  0.277640325 -0.116527602 -0.07920072  0.489981954  0.063910555  0.225906867
#> 60 -0.345277885  0.378228731  0.20802274  0.070820770  0.042771435  0.190921872
#> 61  0.141581016 -0.171061519 -0.07795703  0.415321465 -0.044330449  0.046385126
#> 62  0.443986359 -0.322206373  0.11751995 -0.346108039  0.226998336  0.077252637
#> 63  0.125420932 -0.249089450 -0.16096768  0.667256169  0.103986999  0.055979454
#> 64 -0.312807901 -0.287227244 -0.17196344 -0.537482457 -0.057865452 -0.079174128
#> 65 -0.037086317  0.105659271  0.55844759  0.065642249  0.198720055 -0.072446055
#> 66  0.341903174 -0.216150156  0.31124798 -0.097226542  0.100972267 -0.037211967
#> 67 -0.108062001 -0.150827614 -0.24978470 -0.278957239 -0.116664344 -0.071966057
#>           Mean.C     FC.crude  logFC.crude
#> 1  -0.2035217006 -0.001012229  0.738300552
#> 2   0.1313975426 -9.775281907 -0.225383929
#> 3  -0.2233906302  1.086115116 -0.037262646
#> 4  -0.0568391406 -3.660748194  0.100987023
#> 5  -0.0207996382  0.749407529  0.118193458
#> 6   0.0928067853  9.936154391  0.245689873
#> 7   0.2302485962 -6.634053792 -0.484051805
#> 8   0.3006871621  0.463450790 -0.178773646
#> 9   0.2416632410  0.409583713 -0.278509913
#> 10 -0.4585974431 -0.621447653  1.003289826
#> 11 -0.2597506612  0.196079275  0.479255032
#> 12 -0.0091819385  0.262851298  0.071705171
#> 13  0.1603277297  0.749520740 -0.094736297
#> 14  0.0723852244  1.253074713 -0.012781877
#> 15 -0.0387642769  2.317189561 -0.143053149
#> 16 -0.1876511873 -8.759998808  0.187597849
#> 17  0.0341485402  1.310469578  0.092852063
#> 18 -0.0604327838  3.498172543 -0.850400912
#> 19 -0.1739872935  5.579627439  0.437341913
#> 20 -0.1489034495  1.249137804 -0.021127988
#> 21  0.1764357787  0.146899330 -0.288278612
#> 22  0.0222321928  0.166256749  0.136777024
#> 23 -0.0094134386  2.167484392  0.234244145
#> 24 -0.0857676560  1.986907428  0.223523602
#> 25 -0.0559016747  0.451421613  0.339801007
#> 26 -0.1113569151  0.062227991  0.629734281
#> 27 -0.1391803120  1.703884023 -0.510623008
#> 28  0.1701644523  0.019678361 -0.247927048
#> 29 -0.1896341223 -0.016249598  0.567257020
#> 30 -0.1689839886 -0.041188963  0.162738297
#> 31 -0.0018586668  0.490902110 -0.074913920
#> 32  0.2263403687  1.087788803  0.011448590
#> 33  0.2173777857  0.870864669 -0.058349713
#> 34  0.1507557056 -0.143356889 -0.473035833
#> 35  0.1842742449 -0.646161777 -0.422183714
#> 36  0.3274765952 -0.176971751 -0.652271377
#> 37  0.2454772302  0.452201162 -0.231597234
#> 38 -0.0288434986  2.620328221  0.331427469
#> 39 -0.1716117406  1.859021582  0.014801992
#> 40 -0.2302592673  0.604736539  0.565279095
#> 41  0.0971569864  2.465599812  0.151769965
#> 42  0.1424886829  1.753356326 -0.232051807
#> 43 -0.2939355068  0.418902697  0.214358879
#> 44 -0.0419009527  0.430376677 -0.027888709
#> 45  0.0965942507  1.134106989 -0.045638997
#> 46  0.1096528565  0.512464124 -0.219043347
#> 47  0.0663104673  1.419785920  0.047290014
#> 48 -0.2545312965  1.086922630 -0.101891546
#> 49  0.0091126678  0.879725088 -0.056885386
#> 50 -0.1334635063  1.008296530 -0.001509039
#> 51  0.1884484570  0.611538884 -0.151263202
#> 52 -0.0736601161  3.501940787  0.141589153
#> 53 -0.1777878954  0.669251245  0.111077137
#> 54 -0.0033565367  1.508717367  0.155539034
#> 55 -0.3609889153 -4.254602158  0.660430466
#> 56  0.0649413294 -0.523246616  0.111800940
#> 57 -0.2031995107  1.869080824 -0.013149075
#> 58  0.0194812316  5.227845372  0.367896244
#> 59  0.0980845425  4.740339763  0.168126568
#> 60  0.2190240810  0.840557237 -0.034886284
#> 61  0.0554343064  0.909294812 -0.011920018
#> 62 -0.1835981547  0.364702822  0.528638748
#> 63  0.0857330141  2.629976132 -0.044399698
#> 64 -0.3322243799  0.077667148  0.431597795
#> 65  0.2432497033 -6.089563167 -0.519191493
#> 66 -0.0007095712  1.358986385 -0.059847010
#> 67 -0.2265231844 -0.406870185  0.230980304