Overlay Plots for Methylation and Gene Expression Analysis

This function generates overlay plots to analyze the relationship between methylation levels (mCA/CA) and gene expression across different gene lengths. It uses binning to aggregate data points and visualize trends in methylation and expression changes.

overlaymC(
  mat,
  bin.size = 200,
  shift.size = 40,
  length.type = "Gene",
  methyl.type = "",
  comp.between1 = "",
  comp.between2 = ""
)

Arguments

mat

A dataframe with columns corresponding to comparison matrix values (comp.mat), log fold change (logFC.crude), gene length, methylation values (mCA.CA), and gene names.

bin.size

Numeric, the size of the bin for aggregating genes.

shift.size

Numeric, the shift size for the moving window across the genes.

length.type

Descriptive string of the length type being analyzed, typically "Gene".

methyl.type

Descriptive string of the methylation type, typically something like "mCA/CA".

comp.between1

Descriptive string of the first condition or state being compared.

comp.between2

Descriptive string of the second condition or state being compared.

Value

A list containing ggplot objects for the overlay plots.

Examples

# Generating a toy dataset
set.seed(123) # For reproducibility
a <- runif(1000, min=-2, max=2)
b <- runif(1000, min=-2, max=2)
d <- runif(1000, min=0, max=0.05)
c <- sample(2000:1000000, 1000, replace=TRUE)
genes <- replicate(1000, paste(sample(letters, 8, replace = TRUE), collapse = ""))
df <- data.frame(comp.mat = a, logFC.crude = b, gene.length = c, mCA.CA = d, gene.name = genes)
# Running the overlaymC function
overlay_plots <- overlaymC(mat = df, bin.size = 60, shift.size = 6, methyl.type = "mCA/CA")
#>        comp.mat logFC.crude gene.length       mCA.CA gene.name         lg sublg
#> 869  0.03073145 -0.01762625     382.521 0.0001757976  aqcnpfbd  long gene  long
#> 20   1.81801460 -0.88310476     781.547 0.0001767747  zctlpawb  long gene  long
#> 312 -0.14133623 -0.53607945     122.230 0.0002495589  gbpdalrg  long gene  long
#> 9    0.20574006 -1.73748755      95.353 0.0002698155  orrkgnxr short gene short
#> 382  1.04329450 -0.58108152     128.589 0.0004020768  qaoghquf  long gene  long
#> 183  0.83316129  0.71609124     346.914 0.0004358710  qwcpiwli  long gene  long
#> Total number of bins =  158 
#> Total number of Short Gene bins =  0 
#> Total number of Long Gene bins =  158 
#> Total number of bins that are statistically significant =  0 
#> Total number of Short Gene bins that are statistically significant =  0 
#> Total number of Long Gene bins that are statistically significant =  0 
#> Bin with max mean gene length: 130 
#> Bin with max num of long genes (>= 100 Kb): 147 
#> Using bin as id variables
# To view the plots, you would call each element from the list, e.g., print(overlay_plots$plot1)