Calculate P-Values from a Two-Sample T-Test

This function performs a two-sample t-test to compare the means of two groups. It calculates the t-statistic and the corresponding p-value, allowing for unequal variances between the groups by default, a scenario often referred to as Welch's t-test.

studentTest2(m1, m2, s1, s2, n1, n2 = n1, m0 = 0, equal.variance = FALSE)

Arguments

m1

The mean of the first sample.

m2

The mean of the second sample.

s1

The standard deviation of the first sample.

s2

The standard deviation of the second sample.

n1

The size of the first sample.

n2

The size of the second sample, defaulting to the size of the first sample (`n1`) if not specified.

m0

The hypothesized difference between the two means under the null hypothesis, default is 0.

equal.variance

A logical flag indicating whether to assume equal variances for the two samples, default is `FALSE`.

Value

The p-value resulting from the two-sample t-test.

Examples

# Generate random data for two samples
a <- runif(30, min=-0.2, max=0.2)
b <- runif(30, min=-0.2, max=0.2)
a_std <- runif(30, min=1, max=1.2)
b_std <- runif(30, min=1.3, max=1.5)
n <- rep(30, 30)
# Calculate p-values using the studentTest2 function
p_values <- mapply(studentTest2, m1 = a, m2 = b, s1 = a_std, s2 = b_std, n1 = n)
print(p_values)
#>  [1] 0.9199738 0.8114198 0.5501029 0.9660185 0.6470425 0.5804570 0.6236858
#>  [8] 0.9441062 0.4558250 0.9135324 0.8182590 0.8551891 0.9621863 0.9805997
#> [15] 0.9251608 0.9136996 0.4911980 0.4305298 0.2619970 0.9936578 0.8893505
#> [22] 0.6845603 0.5775615 0.9390004 0.9842703 0.7063366 0.7206298 0.4318245
#> [29] 0.6990497 0.3111045