Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeans.between.ABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeans.between.ABC(dat = dat, A.samples = 1:3, B.samples = 4:6, C.samples = 7:9)
#>              V1           V2           V3            V4          V5
#> 1   0.154442508  0.624950528  0.004434717 -0.1345943531  0.33588700
#> 2  -0.263189775  0.729257156  0.063234240 -0.0961375091  0.04718981
#> 3   0.139486957 -0.011930982  0.060883931  0.0345761734 -0.29279352
#> 4  -0.694323787  0.295789090  0.181329289  0.0658711929 -0.05193623
#> 5   0.326102435  0.059612711  0.632525761 -0.0116148399  0.20164608
#> 6   0.122346269  0.422991256 -0.026309828  0.0983083094 -0.03357204
#> 7  -0.172467767 -0.057107625  0.005626631 -0.0605548777  0.03716419
#> 8  -0.478148501  0.005647068  0.480928537  0.1910589324 -0.12852298
#> 9  -0.039940665 -0.167276596  0.071419159  0.1473031382 -0.07487419
#> 10 -0.127656862  0.499057643 -0.063709297  0.0571876545  0.13711517
#> 11  0.493253615 -0.059088052 -0.309627561 -0.0294951706  0.33942440
#> 12 -0.270455079 -0.149406670 -0.312502119  0.1360036301  0.49421912
#> 13 -0.245199686 -0.323378744 -0.646839390  0.1429044141 -0.09367505
#> 14 -0.112255111  1.033894329 -0.058269177 -0.2272053790 -0.36876476
#> 15 -0.149366926 -0.254030897 -0.586077927 -0.2472186580 -0.30628920
#> 16 -0.053572596  0.304381773 -0.148343706  0.3847275608 -0.10156769
#> 17  0.214318136 -0.434640297 -0.138369815 -0.0111378099  0.59382937
#> 18 -0.007793524 -0.295905954 -0.210783678  0.2637487307  0.14329996
#> 19  0.134973934 -0.313029436 -0.594240635 -0.2870371092  0.01097123
#> 20  0.404541644  0.429905347  0.063880622 -0.3500243413 -0.01693971
#> 21 -0.264322290  0.804513439 -0.623238718  0.2658152157 -0.06928643
#> 22 -0.457198619 -0.035222463 -0.016784413  0.2756396240 -0.32052708
#> 23 -0.103539853  0.407290792 -0.012048279 -0.4218182720  0.16440883
#> 24 -0.032311215 -0.253252498  0.135899823  0.0007081162 -0.29842146
#> 25 -0.019535472 -0.491501680  0.013412951 -0.4315680466  0.13000729
#> 26  0.057856687  0.111808821  0.080709585  0.3600460862 -0.21081023
#> 27 -0.358958498  0.111839065 -0.538118732 -0.5319428453 -0.21628385
#> 28  0.258628510  0.045928121 -0.286305738 -0.3505465038 -0.05304190
#> 29 -0.116894194 -0.264142692  0.038682619  0.0607795139 -0.10634416
#> 30  0.003577393  0.394828072 -0.035653769 -0.8508042203 -0.24815395
#> 31 -0.152934792 -0.428641916  0.525597138  0.5912968815 -0.69952538
#> 32  0.304756287 -0.502762536 -0.124035379  0.7119865676  0.32491808
#> 33  0.255755459  0.043243420 -0.024848587 -0.5138872002  0.08555727
#> 34 -0.290690942  0.057323642  0.082693142 -0.4100752992 -0.18972772
#> 35  0.002282534 -0.200066003 -0.215025542  0.1182871234 -0.51509349
#> 36 -0.008516783  0.035484096 -0.466448316 -0.0861672667  0.22916675
#> 37  0.047042436  0.198458876 -0.150181734  0.0014495943  0.28412815
#> 38  0.446927532 -0.050425880  0.490301457 -0.0407960998 -0.22129112
#> 39  0.295577552  0.011240486 -0.052005808 -0.0949111356 -0.52804468
#> 40 -0.441037954 -0.162445602 -0.070583797  0.0482771074 -0.18709504
#> 41 -0.135574163  0.399621525  0.335191470  0.4646472805 -0.57818474
#> 42 -0.035193611 -0.144188635 -0.336371064  0.1368590461  0.01145475
#> 43  0.225538152 -0.615691880 -0.134212890 -0.2969097397 -0.14445453
#> 44  0.061460556  0.149254470 -0.574580289 -0.3188686915  0.55414104
#> 45  0.408196776 -0.111729030 -0.463387026  0.0557344126  0.13066231
#> 46 -0.057731772  0.406929109 -0.305990207  0.0357504640  0.76648209
#> 47  0.294812541 -0.217170475 -0.144823112  0.1125170736 -0.16829942
#> 48  0.401535967  0.142154293  0.235771418 -0.4071169091  0.11355896
#> 49 -0.201539105 -0.176875520  0.132206635 -0.5744214702 -0.01809780
#> 50  0.604042869 -0.383383511  0.279656631  0.1263080270  0.21865273
#> 51  0.034157418 -0.175293715 -0.153282212 -0.1766735423 -0.20479804
#> 52  0.603495112 -0.286203106  0.152374368  0.0969832331 -0.03589177
#> 53  0.197977848  0.378602008  0.005331787 -1.0653974024 -0.37665724
#> 54  0.263428767  0.580647506 -0.180905871 -0.1946319384 -0.08502153
#> 55  0.280957939 -0.006374516 -0.236456856  0.4216011836  0.02333700
#> 56 -0.477820100 -0.209728554 -0.004740149 -0.6428783701  0.14643045
#> 57 -0.157622962 -0.490581947  0.526897185 -0.0224460296 -0.21590284
#> 58 -0.649416324 -0.807234770 -0.256962880  0.2885961822  0.09258373
#> 59  0.025120151 -0.044685293 -0.300809177 -0.3520239790 -0.36240624
#> 60  0.064292715 -0.270498278  0.057208152  0.0461990645 -0.29814575
#> 61 -0.863543057 -0.285717283 -0.229330318  0.0495728519 -0.36463421
#> 62  0.054209692 -0.384266251  0.462285089 -0.1978780203  0.27544416
#> 63  0.068653457  0.425960470 -0.006262629  0.3297382137 -0.17694085
#> 64  0.540898480 -0.162233444  0.452320434 -0.6713621604  0.68857122
#> 65 -0.053456559 -0.096113232  0.609921433  0.2755665340  0.14436395
#> 66  0.015970081  0.125904344 -0.559636838  0.0578621964  0.13636642
#> 67  0.452187730  0.112612940  0.175454607  0.2408949870  0.32483100
#>              V6           V7           V8           V9       Mean.A
#> 1  -0.410027108  0.056211539  0.125908152 -0.429229861  0.261275918
#> 2  -0.043773307 -0.178615869 -0.125073100 -0.298142216  0.176433874
#> 3  -0.300750507  0.757434089  0.383109821  0.286303385  0.062813302
#> 4   0.029420059  0.218199735 -0.432489408  0.248990086 -0.072401802
#> 5  -0.378307828 -0.129185058  0.091926603  0.028494424  0.339413636
#> 6   0.228057310 -0.250658226  0.154928258 -0.039959261  0.173009232
#> 7   0.132840088  0.219371748  0.190067432  0.148788773 -0.074649587
#> 8  -0.128423129  0.569411552  0.489450089  0.231232278  0.002809035
#> 9  -0.140258894 -0.055697088 -0.099798578  0.413203518 -0.045266034
#> 10  0.291295217 -0.057934024 -0.017116735 -0.103880530  0.102563828
#> 11 -0.084452146 -0.172594385  0.295586843 -0.547323007  0.041512667
#> 12  0.448659363  0.072714744 -0.254298909  0.222282101 -0.244121289
#> 13  0.042440712 -0.318828719  0.483511808 -0.082865658 -0.405139274
#> 14  0.299415942  0.194884730 -0.290082305  0.090905365  0.287790014
#> 15  0.243196736  0.124480930  0.063814362  0.596134642 -0.329825250
#> 16  0.135487328 -0.071821945 -0.103187111  0.336492789  0.034155157
#> 17 -0.071395471 -0.292423478 -0.290703738  0.358256801 -0.119563992
#> 18 -0.217710364  0.077156947  0.248953650 -0.412497488 -0.171494385
#> 19 -0.004781147  0.178936790  0.144576660 -0.001495397 -0.257432046
#> 20  0.442870207 -0.254195217 -0.017635457  0.141209543  0.299442538
#> 21 -0.116698632 -0.212972724  0.415825667 -0.069695671 -0.027682523
#> 22  0.153852964 -0.284522574 -0.111988225 -0.022740932 -0.169735165
#> 23 -0.267027762 -0.422865520 -0.261230536  0.570446935  0.097234220
#> 24 -0.077383880  0.070401803 -0.032824286 -0.402049931 -0.049887963
#> 25 -0.729724105  0.644970979  0.104757698 -0.040570603 -0.165874734
#> 26 -0.156398105 -0.067817717 -0.063663905 -0.167572552  0.083458364
#> 27 -0.212585270  0.305337093  0.051202610  0.039525553 -0.261746055
#> 28 -0.169436245  0.520393606  0.074326026 -0.001022739  0.006083631
#> 29 -0.243983086 -0.083663722 -0.062914368 -0.283054655 -0.114118089
#> 30 -0.043131250  0.054099721  0.303493916 -0.276776163  0.120917232
#> 31  0.464873331  0.663786508  0.456145529  0.022207969 -0.018659857
#> 32  0.342153699  0.094809502 -0.092187209  0.071365687 -0.107347209
#> 33  0.256910483  0.066212476 -0.895443926 -0.232922330  0.091383431
#> 34  0.163903492 -0.329231766 -0.448512724  0.255821428 -0.050224719
#> 35 -0.031878989  0.292507637  0.206425779 -0.054472179 -0.137603004
#> 36  0.017600877 -0.007163728  0.090144610 -0.481717905 -0.146493668
#> 37 -0.166169162 -0.009375363 -0.611634690 -0.330270000  0.031773193
#> 38  0.519744683  0.250643106  0.257705018 -0.439028138  0.295601036
#> 39  0.539557071  0.268778382  0.369083001  0.142275124  0.084937410
#> 40 -0.101524544  0.402206605  0.108845628  0.094720354 -0.224689118
#> 41  0.107104658 -0.179325028  0.012853875 -0.066435645  0.199746278
#> 42  0.390607531  0.152601250  0.508505638  0.443537152 -0.171917770
#> 43 -0.383057174  0.471955287  0.261131626  0.024870765 -0.174788873
#> 44 -0.112952830  0.139104798 -0.015065448 -0.240493530 -0.121288421
#> 45  0.261251856  0.003758203 -0.211573391  0.234385367 -0.055639760
#> 46  0.350077663 -0.328217438 -0.822853086 -0.450987279  0.014402377
#> 47 -0.011384719  0.114464659  0.148455175 -0.312663300 -0.022393682
#> 48 -0.262786003  0.039559031 -0.383993598 -0.078036538  0.259820559
#> 49  0.127219442 -0.232022251 -0.210328916  0.173788637 -0.082069330
#> 50  0.665147990 -0.402236037  0.053666171 -0.233600958  0.166771996
#> 51 -0.159712985 -0.063829176 -0.530365833  0.281716110 -0.098139503
#> 52  0.244253683 -0.178627000  0.002382421 -0.106113434  0.156555458
#> 53 -0.060689884 -0.005962914  0.392776698 -0.071311083  0.193970548
#> 54 -0.204626598  0.056835019 -0.343008062  0.196096981  0.221056801
#> 55 -0.028308519 -0.285804453 -0.035844293  0.012646123  0.012708856
#> 56  0.282884941 -0.429693501 -0.032344612 -0.673158070 -0.230762935
#> 57  0.084000763  0.492223733  0.420495333 -0.292961955 -0.040435908
#> 58  0.461910862  0.108821136 -0.361075408  0.121938498 -0.571204658
#> 59  0.066059835  0.050965726 -0.140729148 -0.100734652 -0.106791440
#> 60  0.325584158 -0.116218936  0.215754228  0.329838928 -0.049665804
#> 61  0.226984108  0.206618619  0.137739253 -0.267952317 -0.459530219
#> 62  0.268599695 -0.046068206 -0.526512188  0.078987163  0.044076177
#> 63  0.232480042 -0.562166454  0.372861438 -0.032330963  0.162783766
#> 64  0.163229545  0.074541779  0.364242940 -0.493209876  0.276995157
#> 65 -0.025235286  0.546351546 -0.367677522  0.154442508  0.153450547
#> 66  0.328830599  0.169081505  0.246920752 -0.263189775 -0.139254138
#> 67  0.328424775  0.099313825 -0.229699771  0.139486957  0.246751759
#>          Mean.B       Mean.C      FC.crude  logFC.crude
#> 1  -0.069578152 -0.082370057  0.9627759235  0.027846684
#> 2  -0.030907003 -0.200610395  0.5499112268  0.347570935
#> 3  -0.186322619  0.475615765 -0.6035233409 -0.911936084
#> 4   0.014451674  0.011566804  1.0343565251  0.003834482
#> 5  -0.062758862 -0.002921344  1.1747922987 -0.137623671
#> 6   0.097597860 -0.045229743  0.3455449334  0.242083958
#> 7   0.036483133  0.186075985  0.4262440383 -0.182223120
#> 8  -0.021962392  0.430031307 -0.0579825272 -0.549397733
#> 9  -0.022609980  0.085902618  0.1727246065 -0.145493034
#> 10  0.161866012 -0.059643763 -0.3655943860  0.338449484
#> 11  0.075159026 -0.141443516 -0.1839039188  0.339257327
#> 12  0.359627372  0.013565979  2.3429510735  0.350674829
#> 13  0.030556693  0.027272477  1.0075951131  0.003304007
#> 14 -0.098851400 -0.001430737  1.3368384294 -0.212670804
#> 15 -0.103437041  0.261476645  0.3828639999 -0.375791832
#> 16  0.139549066  0.053827911  5.3573541515  0.116454349
#> 17  0.170432029 -0.074956805  6.5011053530  0.304406081
#> 18  0.063112776 -0.028795630  1.6440729386  0.111606943
#> 19 -0.093615676  0.107339351  0.4490932422 -0.229795873
#> 20  0.025302053 -0.043540377  0.7992832094  0.143759465
#> 21  0.026610051  0.044385757  0.7533491008 -0.024121513
#> 22  0.036321837 -0.139750577  6.8720971834  0.227675346
#> 23 -0.174812400 -0.037883040  2.0134113082 -0.248658486
#> 24 -0.125032409 -0.121490805  1.0494617919 -0.005514048
#> 25 -0.343761620  0.236386025 -0.4422178452 -0.770345839
#> 26 -0.002387417 -0.099684724  0.4687361249  0.162354185
#> 27 -0.320270657  0.132021752 -0.1486271877 -0.565994936
#> 28 -0.191008216  0.197898964 -1.0275083017 -0.569853846
#> 29 -0.096515912 -0.143210915 -0.6050349559  0.067768387
#> 30 -0.380696473  0.026939158  5.3375610097 -0.862281568
#> 31  0.118881610  0.380713335  0.3443933380 -0.298861622
#> 32  0.459686115  0.024662660  4.2953858414  0.469149324
#> 33 -0.057139817 -0.354051260  0.3334343977  0.618609769
#> 34 -0.145299844 -0.173974354  0.7682861043  0.046454887
#> 35 -0.142895119  0.148153746 -0.0185196499 -0.370272909
#> 36  0.053533453 -0.132912341 14.7280988230  0.243605161
#> 37  0.039802862 -0.317093351 -0.0230164493  0.630512929
#> 38  0.085885822  0.023106662  0.7696130049  0.119414230
#> 39 -0.027799582  0.260045502 -0.6438137198 -0.405359744
#> 40 -0.080114157  0.201924196  0.3388899406 -0.317782391
#> 41 -0.002144267 -0.077635600  0.7278433140  0.143353179
#> 42  0.179640442  0.368214680  0.6508740806 -0.188430776
#> 43 -0.274807148  0.252652559 -0.2339929357 -0.665463942
#> 44  0.040773172 -0.038818060  1.9650889212  0.102359015
#> 45  0.149216194  0.008856726  3.1762343316  0.178689488
#> 46  0.384103405 -0.534019268 -0.6741182294  1.600812780
#> 47 -0.022389021 -0.016581155  0.0008018526 -0.008354702
#> 48 -0.185447984 -0.140823702  1.1113813087 -0.111623010
#> 49 -0.155099943 -0.089520843  9.8007758754 -0.098615885
#> 50  0.336702914 -0.194056941 -0.4709459251  0.872149345
#> 51 -0.180394856 -0.104159633 13.6633840841 -0.115123840
#> 52  0.101781716 -0.094119338  0.2185051811  0.335067729
#> 53 -0.500914843  0.105167567  7.8250232924 -1.578411769
#> 54 -0.161426688 -0.030025354  1.5233399966 -0.278329754
#> 55  0.138876555 -0.103000875 -1.0903810684  0.348829758
#> 56 -0.071187661 -0.378398727 -1.0808711829  0.444054559
#> 57 -0.051449368  0.206585704 -0.0445850056 -0.334463677
#> 58  0.281030258 -0.043438591  1.6147967253  0.277843432
#> 59 -0.216123461 -0.063499358 -2.5254507859 -0.228183482
#> 60  0.024545824  0.143124740  0.3849339624 -0.151062482
#> 61 -0.029359083  0.025468518  0.8869531050 -0.054273912
#> 62  0.115388612 -0.164531077 -0.3418502201  0.436913539
#> 63  0.128425802 -0.073878660  0.1451770964  0.339167273
#> 64  0.060146200 -0.018141719  0.7347402992  0.151947451
#> 65  0.131565066  0.111038844  0.5160245703  0.030597868
#> 66  0.174353072  0.050937494  1.6489011964  0.142340070
#> 67  0.298050256  0.003033670 -0.2104829281  0.475176335