Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeans.between.ABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeans.between.ABC(dat = dat, A.samples = 1:3, B.samples = 4:6,
 C.samples = 7:9)
#>             V1           V2          V3           V4           V5           V6
#> 1  -0.20279907 -0.299823871 -0.26046725 -0.275690243 -0.139749310  0.209696130
#> 2   0.30600515 -0.070984774  0.70282009 -0.444689080  0.525539496  0.192769517
#> 3   0.20147959 -0.254033854 -0.23896666  0.180073035  0.304892123 -0.182481544
#> 4   0.11001241  0.109897549  0.83500804  0.288813069  0.453029727 -0.481245082
#> 5  -0.36879514 -0.299597488  0.39840395 -0.217785446  0.316232085  0.109493111
#> 6   0.12948925  0.306253740 -0.13494704  0.205904982 -0.049280157  0.128228852
#> 7  -0.06927900  0.592664365 -0.07739669 -0.213538737  0.376462738  0.558924603
#> 8  -0.23848309  0.363048124  0.22236061  0.154762772  0.108217583  0.086193369
#> 9   0.62012123 -0.255410981  0.62313785 -0.203229679  0.022226648  0.291567207
#> 10 -0.12778975 -0.489158462  0.17041050  0.326748534 -0.079491609  0.386877231
#> 11  0.34138465  0.315374747 -0.36954015 -0.078384881  0.017439565 -0.061152003
#> 12 -0.33399347  0.245904729  0.36298192 -0.541296483  0.170114408 -0.012983415
#> 13 -0.28103978  0.136232080 -0.49977998 -0.064422418 -0.357270833 -0.209859441
#> 14 -0.19829829 -0.022860518  0.09658264 -0.468719112 -0.168878468 -0.120503395
#> 15 -0.23634514  0.228989100 -0.20454780 -0.252959653  0.466908809  0.278770766
#> 16  0.27259221  0.046718767 -0.45628589  0.284112545 -0.616092843 -0.347865022
#> 17  0.20730701 -0.392054985  0.29855942 -0.035741791 -0.573248985  0.204573249
#> 18  0.17999544  0.330263853 -0.19153998 -0.193575983  0.109863075  0.158749100
#> 19  0.50692142  0.420568064 -0.06557873 -0.008609323 -0.060214850 -0.096695644
#> 20  0.50684615 -0.477523893  0.49044176  0.028285306 -0.521722188  0.100176543
#> 21  0.42977830 -0.558523728  0.14116868 -0.182990567  0.704184017 -0.240193494
#> 22  0.35222278 -0.127580550  0.20383548  0.278788992 -0.468461837 -0.028318056
#> 23  0.22485156 -0.548855806 -0.41216561  0.141029642  0.022119278  0.393108354
#> 24  0.30577555 -0.295218169  0.40141132  0.659757804 -0.041549489 -0.416170939
#> 25  0.25060544 -0.025079704  0.25727266 -0.438434419 -0.250202848  0.259883333
#> 26 -0.65646117 -0.237650220 -0.08746651  0.292742026  0.391759236 -0.228651244
#> 27  0.21835248  0.226186934  0.21862924 -0.017079630  0.221254499 -0.635931169
#> 28 -0.01752210  0.076263905  0.38880095 -0.020697245 -0.120445345 -0.123999622
#> 29 -0.62280838  0.518312951  0.07396989  0.085246764  0.673388756 -0.311061402
#> 30  0.39273923 -0.256990836  0.03936516  0.248562288  0.521389500 -0.059341993
#> 31 -0.30684352  0.259383147  0.54931265 -0.373078088  0.655384422 -0.376149404
#> 32  0.30433179  0.124098438  0.73305842  0.037587887  0.268204491 -0.565705327
#> 33  0.42974784 -0.131841104 -0.06876531 -0.473069983 -0.555402584  0.198191797
#> 34 -0.58116477  0.065227864 -0.05354765 -0.306374257 -0.049739231  0.376871647
#> 35 -0.57334126 -0.371978689  0.01193375 -0.150601437  0.093795944  0.248069979
#> 36  0.48394718 -0.293559938  0.14970227 -0.109466108 -0.038124262  0.192146938
#> 37  0.29051651  0.030708620 -0.55353372  0.264218855 -0.419706886 -0.290209703
#> 38  0.13802593 -0.424083567 -0.17198710 -0.454291434 -0.040642644 -0.121629071
#> 39  0.37581836  0.236347741  0.04576433  0.003316215 -0.178724499  0.030634890
#> 40 -0.61685416 -0.088066791 -0.30650884 -0.026851039 -0.016450870 -0.300067528
#> 41 -0.28624127  0.450024419  0.63280805  0.182192603  0.211158532 -0.456351237
#> 42 -0.65763706 -0.070651379  0.20780041 -0.365967875 -0.409056700  0.184326339
#> 43 -0.29965481 -0.024602563 -0.39742094  0.158671185  0.436705879  0.046715527
#> 44  0.17173964 -0.316543571 -0.15098862  0.560547724  0.007585594 -0.034949619
#> 45  0.44817641  0.180132621 -0.48641190 -0.455085777  0.096892540 -0.434413317
#> 46 -0.16874690 -0.196231607 -0.02902133 -0.103334579 -0.300750930  0.159049988
#> 47  0.08739616  0.075838460  0.01279954 -0.719828754  0.119079065  0.609873093
#> 48 -0.05361063 -0.561207928 -0.30742807 -0.090230502  0.205891958  0.122361679
#> 49  0.20963235 -0.107465622  0.09376105 -0.030352078 -0.457154712  0.141039818
#> 50  0.42168232  0.017142515 -0.28809352 -0.068891756  0.718228860  0.085362760
#> 51 -0.10239133  0.415033186  0.23825159  0.340200512  0.173145352 -0.057838736
#> 52  0.01844080  0.083474290  0.21101965  0.203751012  0.337688641  0.875428062
#> 53  0.25933211 -0.120886563 -0.03599340 -0.117980476  0.220934843  0.390367066
#> 54 -0.26667924  0.291260104 -0.19926641  0.516612494  0.276449242 -0.273647389
#> 55  0.35083079 -0.064640910  0.38304649 -0.029537908 -0.473358981 -0.432512521
#> 56  0.03503800 -0.057151118 -0.23463686 -0.021694655 -0.094418249 -0.352204244
#> 57 -0.41313827 -0.370331461 -0.62636405 -0.255361634 -0.090006947 -0.245612512
#> 58 -0.24616022  0.740813287 -0.02597604 -0.057655326 -0.207477414  0.033841965
#> 59 -0.09557437  0.221870335  0.10711943 -0.394573456  0.076995868  0.209251158
#> 60 -0.08917655  0.192157830  0.23247047 -0.410762061 -0.295654253  0.080637588
#> 61 -0.40371801 -0.200268305  0.46703784 -0.511409699 -0.378636702  0.004925565
#> 62 -0.44069976  0.070901142  0.51139936  0.171031785  0.160635166 -0.265089581
#> 63  0.37967391 -0.298934416  0.49984796  0.910965194 -0.213568793  0.445126513
#> 64  0.24516018 -0.420621526 -0.04272808 -0.321495192  0.567119427  0.007192656
#> 65  0.25553335  0.171584578  0.31327737  0.723391126 -0.272917026 -0.279872192
#> 66  0.39950271  0.392760373 -0.11207473 -0.186142367  0.399801459  0.153959583
#> 67 -0.18686192 -0.008426929 -0.68903581 -0.181003751  0.068778290 -0.591635419
#>               V7          V8           V9       Mean.A       Mean.B
#> 1   0.1006686181 -0.25297298 -0.136506979 -0.254363399 -0.068581141
#> 2  -0.2487650464 -0.07162721 -0.024307711  0.312613488  0.091206644
#> 3   0.3111369507 -0.28119297  0.035860771 -0.097173644  0.100827872
#> 4   0.4981126093  0.29265749  0.069840539  0.351639335  0.086865905
#> 5  -0.4590715555 -0.06933577 -0.500916061 -0.089996227  0.069313250
#> 6  -0.0024031000 -0.03359813 -0.581861476  0.100265316  0.094951226
#> 7   0.4840565185  0.31707554  0.157009591  0.148662889  0.240616201
#> 8  -0.3372389485  0.27967514  0.188941054  0.115641882  0.116391241
#> 9  -0.2021897521 -0.03805841 -0.160907196  0.329282699  0.036854725
#> 10  0.4746993554 -0.03391793 -0.511840881 -0.148845904  0.211378052
#> 11 -0.3527690264 -0.31368247  0.168414811  0.095739749 -0.040699106
#> 12 -0.3291271328  0.18705968  0.100102357  0.091631059 -0.128055164
#> 13 -0.0263598534  0.06610867 -0.136745854 -0.214862560 -0.210517564
#> 14 -0.3085092294 -0.31054832 -0.421066030 -0.041525391 -0.252700325
#> 15  0.2784726351  0.05765718 -0.364134721 -0.070634611  0.164239974
#> 16  0.0402211394 -0.18554524 -0.037562137 -0.045658304 -0.226615107
#> 17  0.1953854083  0.14873426  0.037774373  0.037937146 -0.134805842
#> 18 -0.4242374679  0.11074423 -0.099026911  0.106239770  0.025012064
#> 19  0.5446076772 -0.17615253  0.086085468  0.287303588 -0.055173273
#> 20  0.3576649262 -0.42251896 -0.060615089  0.173254672 -0.131086779
#> 21  0.1790992565  0.54391597  0.013414717  0.004141082  0.093666652
#> 22 -0.0325283380  0.11533657  0.441323631  0.142825906 -0.072663634
#> 23  0.1547126114 -0.16782029 -0.111147777 -0.245389951  0.185419092
#> 24 -0.1598591558  0.31541799 -0.184541416  0.137322902  0.067345792
#> 25 -0.3861400194  0.11852226 -0.320395426  0.160932796 -0.142917978
#> 26 -0.6230984265  0.30714918  0.414280927 -0.327192634  0.151950006
#> 27 -0.3088216563 -0.35221946 -0.026170447  0.221056219 -0.143918767
#> 28 -0.0642251233 -0.77921711 -0.024105940  0.149180918 -0.088380737
#> 29  0.3717664287 -0.38171679 -0.293144559 -0.010175180  0.149191373
#> 30  0.5074354368  0.55776821  0.004646101  0.058371186  0.236869931
#> 31  0.2276329362 -0.18854721 -0.231968730  0.167284094 -0.031281024
#> 32  0.1641256558  0.42326597  0.085536560  0.387162882 -0.086637650
#> 33  0.5259470119 -0.47724631  0.031484373  0.076380476 -0.276760257
#> 34 -0.2975817398  0.06434782 -0.244973936 -0.189828185  0.006919387
#> 35 -0.0246759617  0.05560148  0.028135405 -0.311128734  0.063754829
#> 36 -0.6260171783 -0.38113798 -0.338119429  0.113363171  0.014852189
#> 37  0.0169226266 -0.59728471 -0.399707611 -0.077436196 -0.148565911
#> 38  0.4879590857 -0.40145797 -0.247743329 -0.152681578 -0.205521050
#> 39 -0.5415079114 -0.27286566  0.096366791  0.219310143 -0.048257798
#> 40 -0.0032554532 -0.16115807 -0.171890541 -0.337143262 -0.114456479
#> 41 -0.0879953818 -0.04072481  0.513538880  0.265530400 -0.021000034
#> 42  0.1281732410  0.29793928 -0.007457361 -0.173496010 -0.196899412
#> 43 -0.1307769878  0.03782629 -0.554720912 -0.240559438  0.214030864
#> 44 -0.0548797389  0.24786481  0.340944709 -0.098597518  0.177727899
#> 45 -0.2944651752  0.58537967  0.476226982  0.047299044 -0.264202185
#> 46 -0.2133394183  0.03059382  0.425925687 -0.131333277 -0.081678507
#> 47  0.0541920900 -0.20623786 -0.257999386  0.058678055  0.003041135
#> 48 -0.0235176759  0.29955745  0.159412990 -0.307415545  0.079341045
#> 49  0.2278052439  0.27754651  0.229872634  0.065309260 -0.115488991
#> 50  0.0780650123 -0.04217503  0.278141297  0.050243772  0.244899955
#> 51  0.1441549886  0.17522620  0.108911221  0.183631151  0.151835709
#> 52  0.6358698704 -0.25257127  0.004656934  0.104311580  0.472289238
#> 53  0.0574799738  0.46989160 -0.401982683  0.034150717  0.164440478
#> 54 -0.3071546630  0.32272699 -0.319976298 -0.058228518  0.173138115
#> 55  0.4268560306 -0.54137725  0.206915564  0.223078791 -0.311803137
#> 56  0.8577728373 -0.21745853  0.106495670 -0.085583326 -0.156105716
#> 57 -0.2199196677 -0.52202348  0.176181085 -0.469944593 -0.196993698
#> 58 -0.2381990063 -0.35858299 -0.141858280  0.156225677 -0.077096925
#> 59  0.0405259496  0.61753050  0.102448234  0.077805134 -0.036108810
#> 60  0.1541951056  0.20983564 -0.582113821  0.111817250 -0.208592909
#> 61 -0.1149448621  0.48923965  0.425479906 -0.045649491 -0.295040278
#> 62 -0.0001406826 -0.09595933 -0.208360022  0.047200246  0.022192457
#> 63 -0.2993314105 -0.34544478  0.004970247  0.193529151  0.380840971
#> 64  0.7433080295 -0.11041949  0.019627828 -0.072729809  0.084272297
#> 65  0.3332897844  0.26127135 -0.202799075  0.246798434  0.056867303
#> 66 -0.1264012471 -0.29832441  0.306005150  0.226729450  0.122539558
#> 67 -0.3155792334  0.07432260  0.201479586 -0.294774888 -0.234620293
#>          Mean.C    FC.crude  logFC.crude
#> 1  -0.096270448  1.17514574  0.034088064
#> 2  -0.114899990  0.51789442  0.443627936
#> 3   0.021934918  1.66236174  0.098279738
#> 4   0.286870213  4.08795772 -0.347133660
#> 5  -0.343107796 -0.62940417  0.634301088
#> 6  -0.205954236  0.01735386  0.519761857
#> 7   0.319380551  0.53862800 -0.100482014
#> 8   0.043792414 -0.01042958  0.108649979
#> 9  -0.133718453  0.63159233  0.397958018
#> 10 -0.023686486  2.87812106  0.273714782
#> 11 -0.166012230  0.52125243  0.226192748
#> 12 -0.013988365  2.07997937 -0.196834484
#> 13 -0.032332344  0.02380426 -0.235622122
#> 14 -0.346707859  0.69196286  0.183071237
#> 15 -0.009334968  3.83158165  0.218532490
#> 16 -0.060962079 11.82432477 -0.265739185
#> 17  0.127298013 -1.93309437 -0.397074427
#> 18 -0.137506717  0.33324667  0.280837414
#> 19  0.151513540  2.52210574 -0.394340178
#> 20 -0.041823039  1.41503017 -0.174170467
#> 21  0.245476646  0.37095888 -0.188193109
#> 22  0.174710622 -6.75839613 -0.395417211
#> 23 -0.041418487  2.11210448  0.249029943
#> 24 -0.009660860  0.47608735  0.124693019
#> 25 -0.196004395  0.85127239  0.114436939
#> 26  0.032777226  1.33106322  0.121186506
#> 27 -0.229070521  0.81082716  0.207714291
#> 28 -0.289182724  0.54192828  0.440984345
#> 29 -0.101031641 -1.75404754  0.350756775
#> 30  0.356616582  0.59849623 -0.139612887
#> 31 -0.064294336  0.85744220  0.060687288
#> 32  0.224309396  2.90936683 -0.669870349
#> 33  0.026728357  7.11229929 -0.555003849
#> 34 -0.159402618  6.46652120  0.215878584
#> 35  0.019686974  1.13320968  0.046998644
#> 36 -0.448424863  0.17535258  1.040681037
#> 37 -0.326689899  0.28537075  0.307151687
#> 38 -0.053747404 -0.53408716 -0.214424106
#> 39 -0.239335594  0.58338696  0.436121934
#> 40 -0.112101355  0.98953473 -0.002776228
#> 41  0.128272895  2.08753929 -0.274078128
#> 42  0.139551720 -0.07475985 -0.427084702
#> 43 -0.215890538 18.42766766  0.505455053
#> 44  0.177976593  0.99910081 -0.000281083
#> 45  0.255713824 -1.49462158 -0.811589751
#> 46  0.081060030  0.23378689 -0.207942887
#> 47 -0.136681718  0.28479210  0.230997774
#> 48  0.145150923  0.85458517 -0.066889600
#> 49  0.245074795 -1.00574479 -0.526209468
#> 50  0.104677092  3.57604835  0.180127612
#> 51  0.142764135  0.77802211  0.013581027
#> 52  0.129318511 14.71502695  0.416411004
#> 53  0.041796297 17.04118722  0.165704382
#> 54 -0.101467990 -5.35082005  0.364030605
#> 55  0.030798114  2.78177681 -0.796257057
#> 56  0.248936660 -0.21081667 -0.521828894
#> 57 -0.188587355  0.97012217 -0.009495982
#> 58 -0.246213427  0.57977120  0.359533906
#> 59  0.253501560 -0.64835664 -0.407996869
#> 60 -0.072694360  1.73653115 -0.263000051
#> 61  0.266591564 -0.79871235 -0.805898861
#> 62 -0.101486679  0.16819091  0.195700904
#> 63 -0.213268648 -0.46045436  1.001103020
#> 64  0.217505454  0.54094773 -0.157242662
#> 65  0.130587353  1.63436335 -0.125657277
#> 66 -0.039573504  0.39124572  0.288008468
#> 67 -0.013259014  0.21368100 -0.273576691