Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeans.between.ABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeans.between.ABC(dat = dat, A.samples = 1:3, B.samples = 4:6, C.samples = 7:9)
#>               V1           V2          V3           V4           V5
#> 1  -0.1678017574 -0.128266571 -0.32332983 -0.729467875 -0.057395162
#> 2  -0.2936074936 -0.078382019  0.13308912  0.229939048  0.134898021
#> 3  -0.0282884901 -0.299747526 -0.10285340  0.151358147 -0.005573356
#> 4  -0.6474418404  0.247635530 -0.22429052 -0.430097070  0.487062975
#> 5  -0.4377611314 -0.153946750 -0.15183693 -0.058416706  0.575212835
#> 6   0.0597776699 -0.094125536 -0.18519858 -0.256083632 -0.109424647
#> 7   0.1207017381 -0.014228905 -0.37939463  0.126423867 -0.011145992
#> 8   0.1552563955 -0.391983376 -0.45805015 -0.563219443 -0.291058898
#> 9  -0.5459971390  0.117494377 -0.18354246  0.061833530  0.296817480
#> 10  0.3140133998 -0.042220345 -0.19155317 -0.389700281  0.312141934
#> 11 -0.2061570673 -0.333208022 -0.23654078 -0.262795710 -0.529921442
#> 12  0.0958286662  0.139778683  0.12427404 -0.095079992  0.195017231
#> 13 -0.0861324892  0.176572839  0.35287155 -0.065389728 -0.280668295
#> 14 -0.1367213171  0.025944309 -0.42149226  0.136419241 -0.256841476
#> 15 -0.0628344750  0.086938220  0.67260957  0.053848835  0.023524906
#> 16  0.2057131958  0.508741449 -0.11613967  0.305686592  0.133095706
#> 17  0.1672762778  0.119658664  0.32651985  0.455460141 -0.078606059
#> 18  0.0625925826  0.128064666  0.17613082 -0.499487405  0.797730453
#> 19  0.0004403398 -0.206860681 -0.12770345  0.066052862  0.014416559
#> 20 -0.0991364677 -0.234761744 -0.05311710  0.116096983 -0.071844518
#> 21 -0.2335777865  0.188368977 -0.40458064  0.511560738  0.118434141
#> 22 -0.0672327677  0.156752606 -0.20282421  0.115545921 -0.047781243
#> 23 -0.0246760213 -0.035667209 -0.23365380  0.247018899  0.036080367
#> 24  0.1867520983 -0.019219556 -0.29449990 -0.047175455 -0.280532425
#> 25 -0.1619048769  0.346522157 -0.65558132  0.129296358 -0.332493847
#> 26 -0.2292170682 -0.034449048  0.12156809  0.084222456 -0.327583792
#> 27 -0.1653022857  0.167471669  0.27614010 -0.135113563 -0.011517640
#> 28 -0.0450406408  0.340968691 -0.66627176 -0.248497446  0.098561316
#> 29  0.4663879853  0.298666802  0.11027263 -0.142470769  0.613614512
#> 30  0.1955319160 -0.201509331  0.47984877 -0.648487908  0.162448779
#> 31 -0.1642687182 -0.308182097 -0.13564073  0.297442203 -0.461628387
#> 32 -0.2469595244 -0.229615249  0.48230118  0.085985667 -0.075933579
#> 33  0.1970148675 -0.505392555 -0.25186611 -0.523639682  0.175017680
#> 34  0.0938689277 -0.625360367  0.27029036  0.431048052  0.216747350
#> 35 -0.1182338472  0.312670525  0.71045961  0.362542643  0.452146033
#> 36  0.0860775332  0.310295422  0.08750960  0.066658568 -0.381921942
#> 37  0.1285085618  0.021110561 -0.17990375 -0.303068172  0.226030547
#> 38  0.3349940873  0.113539129 -0.26182621 -0.312264636  0.278462250
#> 39  0.3565608454 -0.494068201 -0.30059308 -0.095879962  0.368582818
#> 40 -0.2139432451  0.413611983  0.05683751 -0.357337692  0.099521791
#> 41  0.1493061728 -0.434324782  0.11895705  0.146938274  0.144775050
#> 42  0.3474299272  0.123409566  0.32165739 -0.139680110  0.103449111
#> 43  0.5473607680 -0.047918895  0.09510933 -0.260844931  0.051477395
#> 44 -0.0987722808 -0.116447398 -0.17141185  0.020836072 -0.405609925
#> 45 -0.0674920396 -0.176862970 -0.36728453 -0.274089786  0.511143539
#> 46 -0.2813942751 -0.306938871  0.05548103 -0.268804981 -0.077907848
#> 47 -0.1321266836 -0.105357916 -0.35780176  0.355823414  0.309354191
#> 48  0.2113168754 -0.157563653 -0.21688518 -0.270462249  0.003462691
#> 49 -0.0327076679 -0.121292311  0.03831658  0.123416453  0.338651781
#> 50 -0.1789435175  0.204023070 -0.24699785  0.047550988 -0.024385637
#> 51  0.3807406191  0.084327509  0.02023899  0.050428155  0.215192471
#> 52  0.4580577495  0.644312348  0.12172716  0.242796292 -0.013719442
#> 53 -0.2491842633 -0.245235971 -0.72239034  0.001028592  0.134522349
#> 54 -0.5433340144 -0.153484998 -0.12664402 -0.312754397  0.162060746
#> 55 -0.2636960402  0.002203267  0.16558016 -0.301301332  0.320775678
#> 56  0.3210401381  0.204245709  0.17011537  0.112343338  0.094188472
#> 57  0.0944691685  0.755598507 -0.15659976 -0.251067171  0.090869071
#> 58 -0.0572278362 -0.173361481 -0.57486923  0.022352603  0.327639137
#> 59  0.2227007863  0.009116037 -0.26774609  0.648895976 -0.249549172
#> 60 -0.0030367910 -0.088848209 -0.35550647  0.102033119  0.395198504
#> 61  0.2448149928  0.074667531 -0.19390440 -0.303851151 -0.400449227
#> 62  0.3638684845  0.125945146  0.09367583 -0.015727890  0.019054016
#> 63  0.2806828827  0.399580328 -0.45445394  0.433066414  0.037132361
#> 64 -0.1100781275 -0.144241390  0.42714181 -0.971359879 -0.467063384
#> 65 -0.4061539230 -0.529300081  0.21702943 -0.078931728 -0.397799949
#> 66 -0.0948104421  0.319695325 -0.13335174 -0.529355860  0.179288320
#> 67  0.5548110249 -0.348681972 -0.14161230 -0.089087885 -0.307428497
#>               V6          V7           V8           V9       Mean.A
#> 1  -0.1124374983 -0.05471116 -0.187757168  0.034585228 -0.206466054
#> 2   0.5622296597  0.38844545 -0.151317633  0.129835277 -0.079633465
#> 3   0.0271961233 -0.21639804  0.412524185  0.172761269 -0.143629807
#> 4  -0.4171135163  0.16490037  0.482562124 -0.206099202 -0.208032277
#> 5  -0.5332378237 -0.05528235 -0.002579480  0.216606976 -0.247848270
#> 6  -0.1109933818 -0.11386829  0.390199620  0.075183275 -0.073182149
#> 7  -0.0635944812 -0.04107512  0.201581427 -0.550052649 -0.090973933
#> 8  -0.3200922492  0.31182736 -0.310780410  0.127319843 -0.231592375
#> 9   0.1228484030 -0.20510636  0.254380411 -0.194985437 -0.204015072
#> 10  0.4149531137  0.36417737 -0.069414426 -0.412277894  0.026746627
#> 11  0.0605614663 -0.42299187  0.160276881 -0.151349866 -0.258635290
#> 12 -0.4622706128 -0.20322551  0.482292092  0.151871375  0.119960463
#> 13  0.7548253601  0.13844716 -0.237214612  0.268804339  0.147770632
#> 14  0.2206037766 -0.85584095 -0.088350206 -0.212038893 -0.177423091
#> 15  0.4177654970  0.20096772  0.398340942 -0.505025900  0.232237770
#> 16 -0.0297211364  0.24174025  0.121987011  0.045522319  0.199438323
#> 17  0.5780820178  0.12608989 -0.197593151  0.067784364  0.204484930
#> 18  0.0250883063  0.14430626  0.039537795 -0.494477400  0.122262690
#> 19 -0.1679744518  0.20531416 -0.291569090 -0.130610127 -0.111374599
#> 20  0.0875652341  0.09257518 -0.047000165  0.195897151 -0.129005105
#> 21 -0.2269479321 -0.38813337  0.515487402 -0.456368909 -0.149929817
#> 22  0.0022150689  0.39918541 -0.391426804  0.140213072 -0.037768125
#> 23 -0.2541336935 -0.32069718  0.448429767  0.393784305 -0.097999011
#> 24 -0.4315555224  0.35641594 -0.265582918 -0.023318659 -0.042322452
#> 25 -0.1016416880  0.25560235 -0.022294150  0.482283903 -0.156988013
#> 26 -0.1940293556 -0.09441169 -0.268207668  0.231620405 -0.047366009
#> 27  0.3186404506  0.07266308 -0.020924189 -0.083763337  0.092769828
#> 28  0.0249400142  0.15459814  0.383850493 -0.163496895 -0.123447903
#> 29  0.0605790285 -0.35378683 -0.568206712 -0.134059794  0.291775806
#> 30 -0.5358263364  0.18125229  0.237804237 -0.154842825  0.157957118
#> 31  0.1609927529  0.16362766 -0.649151707  0.218013982 -0.202697182
#> 32 -0.0006020836  0.28026040 -0.018601934 -0.002327507  0.001908802
#> 33 -0.1540913024  0.23907730 -0.196308703  0.149178522 -0.186747933
#> 34  0.0404986040 -0.59143432 -0.010574778  0.208298704 -0.087067025
#> 35 -0.2795628807 -0.56035463  0.319836109 -0.185233730  0.301632098
#> 36  0.0814733699 -0.36025362  0.234083495  0.401547876  0.161294184
#> 37 -0.3078167882  0.27518504 -0.257859798  0.166854925 -0.010094876
#> 38 -0.3282257209  0.53060123 -0.038439668  0.183815665  0.062235670
#> 39  0.2622908704 -0.05728883  0.417173265 -0.470818319 -0.146033477
#> 40 -0.0603981329 -0.23224330 -0.039057700 -0.412667286  0.085502081
#> 41  0.2931137541 -0.28926306 -0.152164214  0.464865730 -0.055353854
#> 42  0.3301597471  0.45502515 -0.081213839 -0.427134325  0.264165629
#> 43 -0.2141632664  0.06120432  0.452906644 -0.499487985  0.198183735
#> 44  0.1671733061 -0.46117125  0.045563903 -0.345542403 -0.128877177
#> 45  0.1782144106  0.01741975  0.040628725  0.530311549 -0.203879848
#> 46 -0.2365415636 -0.03354265 -0.257719439  0.181036500 -0.177617371
#> 47 -0.3704447575 -0.08949423 -0.142810623  0.085024008 -0.198428787
#> 48  0.0324716501  0.15916760  0.331240923 -0.099317332 -0.054377319
#> 49 -0.1424443007 -0.09086845  0.463603398 -0.086454218 -0.038561132
#> 50  0.0254358306 -0.17673803  0.491283293  0.112539404 -0.073972765
#> 51 -0.3198583239 -0.03197112 -0.394268587  0.052992976  0.161769040
#> 52  0.5880648723  0.14139973  0.010469485 -0.200974813  0.408032420
#> 53 -0.4579009947  0.30297530  0.184821514  0.094869662 -0.405603525
#> 54 -0.4467122076 -0.38506673  0.002523152 -0.112352523 -0.274487678
#> 55 -0.0771686862  0.06296793 -0.545077561 -0.036377031 -0.031970872
#> 56  0.1356909005  0.32278304  0.533254988 -0.098042854  0.231800404
#> 57  0.1771793655 -0.51598244 -0.445776482  0.015010395  0.231155973
#> 58 -0.5453153521 -0.01252192 -0.109186837 -0.296430930 -0.268486182
#> 59  0.1883633348 -0.34858823  1.045669594  0.167265783 -0.011976423
#> 60  0.1576073418  0.43646641  0.082246562 -0.300045592 -0.149130492
#> 61  0.0285216500  0.12444005  0.660653536  0.470592964  0.041859373
#> 62 -0.2567205970  0.39899099 -0.100271825  0.333524210  0.194496485
#> 63 -0.0036582013 -0.49401728 -0.049326748  0.029546915  0.075269756
#> 64  0.2804999431  0.20462741 -0.037243971 -0.243001906  0.057607433
#> 65  0.1792878264  0.25871069 -0.039111558 -0.167801757 -0.239474858
#> 66  0.0982773665  0.05500437  0.232049509 -0.293607494  0.030511046
#> 67 -0.2358569082  0.41176348  0.136615278 -0.028288490  0.021505586
#>          Mean.B       Mean.C     FC.crude   logFC.crude
#> 1  -0.299766845 -0.069294366  -0.68017528 -3.267542e-01
#> 2   0.309022243  0.122321032   1.92447167  2.083067e-01
#> 3   0.057660305  0.122962473   0.75504854 -7.636759e-02
#> 4  -0.120049204  0.147121097   0.24773261 -3.168000e-01
#> 5  -0.005480565  0.052915050   0.80584197 -6.626625e-02
#> 6  -0.158833887  0.117171534  -0.44996102 -3.805746e-01
#> 7   0.017227798 -0.129848782  -2.78333504  2.054243e-01
#> 8  -0.391456863  0.042788930  -0.58263622 -6.011031e-01
#> 9   0.160499804 -0.048570462   2.34498240  2.399400e-01
#> 10  0.112464922 -0.039171649  -1.30037223  2.170291e-01
#> 11 -0.244051895 -0.138021618   0.12090996 -1.434015e-01
#> 12 -0.120777791  0.143645986 -10.16394097 -4.311034e-01
#> 13  0.136255779  0.056678961   0.12640951  1.210845e-01
#> 14  0.033393847 -0.385410017  -1.01360668  6.123846e-01
#> 15  0.165046413  0.031427586   0.33460134  2.230430e-01
#> 16  0.136353721  0.136416525   1.00099656 -9.670598e-05
#> 17  0.318312033 -0.001239633  -0.55329855  4.878140e-01
#> 18  0.107777118 -0.103544449   0.06415020  3.481840e-01
#> 19 -0.029168343 -0.072288351   2.10320152  5.866008e-02
#> 20  0.043939233  0.080490723   0.82552641 -4.427124e-02
#> 21  0.134348982 -0.109671624   7.06139007  3.040168e-01
#> 22  0.023326582  0.049323891   0.70149608 -3.492063e-02
#> 23  0.009655191  0.173838963   0.39602341 -1.994073e-01
#> 24 -0.253087801  0.022504789  -3.25118491 -4.320932e-01
#> 25 -0.101613059  0.238530701   0.14000590 -4.030458e-01
#> 26 -0.145796897 -0.043666318 -26.60516395 -1.548177e-01
#> 27  0.057336416 -0.010674814   0.34253502  1.054882e-01
#> 28 -0.041665372  0.124983911   0.32919508 -2.067065e-01
#> 29  0.177240924 -0.352017778   0.17790622  1.313722e+00
#> 30 -0.340621822  0.088071234   7.13418662 -8.913851e-01
#> 31 -0.001064477 -0.089170020   1.77607457  1.098592e-01
#> 32  0.003150002  0.086443651   0.01468270 -1.152868e-01
#> 33 -0.167571102  0.063982374   0.07648390 -2.953664e-01
#> 34  0.229431335 -0.131236797  -7.16549678  4.618794e-01
#> 35  0.178375265 -0.141917416   0.27788743  6.559010e-01
#> 36 -0.077930001  0.091792585   3.44199544 -2.905323e-01
#> 37 -0.128284805  0.061393390  -1.65327734 -2.810762e-01
#> 38 -0.120676036  0.225325742  -1.12153795 -5.093989e-01
#> 39  0.178331242 -0.036977960   2.97430821  2.559689e-01
#> 40 -0.106071345 -0.227989427   0.61109606  2.358392e-01
#> 41  0.194942359  0.007812818   3.96247269  2.339021e-01
#> 42  0.097976249 -0.017774338   0.58944952  2.156153e-01
#> 43 -0.141176934  0.004874325   1.75553104 -2.881526e-01
#> 44 -0.072533516 -0.253716584  -0.45132914  2.714596e-01
#> 45  0.138422721  0.196120008   0.85575673 -6.071677e-02
#> 46 -0.194418131 -0.036741864  -0.11925962 -2.145857e-01
#> 47  0.098244282 -0.049093614   1.98662554  1.740152e-01
#> 48 -0.078175969  0.130363731  -0.12882167 -2.793211e-01
#> 49  0.106541311  0.095426909   1.08295071  1.407125e-02
#> 50  0.016200394  0.142361556   0.41682318 -1.579825e-01
#> 51 -0.018079233 -0.124415578   0.62843445  2.003398e-01
#> 52  0.272380574 -0.016368533   0.31963134  5.865483e-01
#> 53 -0.107450018  0.194222157   0.49706692 -3.014537e-01
#> 54 -0.199135286 -0.164965368   0.68800952 -4.512917e-02
#> 55 -0.019231447 -0.172828885  -0.09044161  2.372945e-01
#> 56  0.114074237  0.252665059  -5.64237321 -2.104932e-01
#> 57  0.005660422 -0.315582842   0.41243743  7.729309e-01
#> 58 -0.065107871 -0.139379896   1.57527815  9.190896e-02
#> 59  0.195903380  0.288115716   0.69271992 -1.061370e-01
#> 60  0.218279655  0.072889126   1.65485443  1.621786e-01
#> 61 -0.225259576  0.418562184  -0.70909731 -9.095662e-01
#> 62 -0.084464823  0.210747792 -17.16546929 -4.951086e-01
#> 63  0.155513525 -0.171265705  -0.32548571  5.197454e-01
#> 64 -0.385974440 -0.025206157   5.35638987 -7.210456e-01
#> 65 -0.099147950  0.017265790   0.54657067 -1.402395e-01
#> 66 -0.083930058 -0.002184539   3.50020054 -1.273817e-01
#> 67 -0.210791097  0.173363423  -1.52969834 -5.853419e-01