Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeansBetweenABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeansBetweenABC(dat = dat, A.samples = 1:3, B.samples = 4:6,
 C.samples = 7:9)
#>             V1          V2           V3           V4           V5          V6
#> 1   0.07635019 -0.39136819 -0.009013909  0.150621641 -0.266068486 -0.33132816
#> 2   0.05345474 -0.32002978  0.539007155  0.168915023  0.585920103 -0.06417767
#> 3  -0.01449705  0.63045066 -0.184835211 -0.006226081 -0.471778309 -0.24396015
#> 4   0.19392539 -0.32533465  0.133045741 -0.216143754  0.445714890  0.06984971
#> 5   0.17678366 -0.30558548  0.385467883 -0.003919399 -0.003859276 -0.03904355
#> 6  -0.03238059  0.14717697 -0.056488470 -0.278684958 -0.171345971 -0.24749025
#> 7   0.06438629  0.16996237 -0.177947279  0.287847917 -0.319616090  0.57686227
#> 8   0.28226554 -0.08402790  0.285701312 -0.369902931 -0.235984096  0.06041132
#> 9  -0.02669560 -0.15100975 -0.794376037  0.274440362  0.026775378  0.42376393
#> 10  0.02945722  0.11290315  0.007767157 -0.186456456 -0.153934578 -0.08919653
#> 11 -0.20088692  0.38930205 -0.187402073 -0.353065768  0.417709626  0.29302558
#> 12  0.55519897 -0.23614448  0.166325005 -0.195277741  0.089009892  0.18333049
#> 13 -0.23435191 -0.39225692  0.130256617  0.336953176  0.352063484  0.24623175
#> 14  0.14313409  0.16517061  0.279670993  0.085972818 -0.124644060  0.49113244
#> 15  0.07032460 -0.23983258 -0.244016029  0.207295321  0.152882566 -0.19518690
#> 16 -0.47974415 -0.09521338 -0.123157318  0.255915852 -0.623297053  0.07066903
#> 17 -0.22374857 -0.39351411 -0.365782563  0.316540901  0.035954364  0.09796152
#> 18 -0.05415071  0.30868895 -0.117623097 -0.072906202 -0.300311231  0.34307116
#> 19  0.09831476  0.24882188  0.054827276  0.142730035 -0.204492236 -0.13885250
#> 20  0.09458233  0.10266402  0.226362483 -0.043835249 -0.419610509  0.01409191
#> 21 -0.16839399  0.85294916 -0.336170804  0.317810740  0.203957254 -0.07471059
#> 22 -0.10787698  0.30833596 -0.312836866 -0.201503235 -0.045170725  0.11173933
#> 23  0.11835900 -0.20091052  0.078812590  0.111645481 -0.210904356  0.24475596
#> 24  0.39814032  0.01181247 -0.122183844  0.093986434  0.051129272  0.22410799
#> 25  0.13130860  0.02328176 -0.446086064  0.300343836  0.056036457 -0.01798577
#> 26  0.09861175  0.34083762 -0.187144075 -0.336700334 -0.043882385 -0.07740615
#> 27 -0.43721233 -0.21693644  0.193096332  0.327273020  0.294445453  0.29185469
#> 28 -0.03563038 -0.23032165  0.465251901  0.392830716 -0.744407374  0.18869599
#> 29 -0.87194743  0.19153197 -0.161306684  0.247976904 -0.523593354  0.25910329
#> 30  0.31841168 -0.54501897 -0.096475878  0.343822162  0.099949307  0.27392399
#> 31 -0.82724567  0.03250257 -0.088463394 -0.222034003  0.573210933  0.04908653
#> 32 -0.16253146  0.01892923  0.123015429 -0.336442689 -0.073762588 -0.17581938
#> 33  0.19843375  0.21112986  0.611424345  0.515716092  0.070109889 -0.22816213
#> 34  0.07023916 -0.37219595 -0.502941218  0.525774087 -0.046696222 -0.33398649
#> 35 -0.13050847  0.30436881  0.139133314 -0.184611210 -0.544171617  0.02580604
#> 36  0.20551118  0.25426026 -0.127504645  0.311259672 -0.108102090 -0.32930976
#> 37 -0.04282574 -0.50896484 -0.058891879  0.499826862  0.240365403  0.02189434
#> 38 -0.18363284 -0.32197444  0.117025654  0.005893315  0.278997390  0.19851638
#> 39  0.44406846 -0.11235365 -0.698608629 -0.254117876 -0.271905327 -0.08095012
#> 40  0.15313721  0.55139904 -0.144157338  0.036063031  0.127708664  0.23353282
#> 41  0.55982355 -0.27226025 -0.170438658  0.137511000  0.380872033 -0.10825873
#> 42  0.05978560 -0.27644235 -0.169469706 -0.260546575  0.137379195 -0.40517026
#> 43  0.03830886 -0.08203999 -0.217375380  0.376121635  0.415971463 -0.09894028
#> 44  0.04599405 -0.18148800 -0.122863495 -0.102412340  0.018450937  0.21132546
#> 45 -0.19392426  0.35420286  0.206420894 -0.223621178 -0.042730559 -0.38459816
#> 46 -0.23675837 -0.01288233  0.017859610 -0.206709826 -0.226833383 -0.12174656
#> 47  0.64718286  0.08797393  0.212667062  0.025246530  0.124345212  0.65126519
#> 48  0.38553867 -0.23995827 -0.042281890  0.368880663  0.255684960  0.01584286
#> 49 -0.35482352 -0.07025191  0.068549189 -0.249116523  0.042325977 -0.39786061
#> 50 -0.21741134 -0.41173422 -0.235933875 -0.258297572 -0.518815564  0.42582847
#> 51 -0.21022396 -0.12587057  0.111730312 -0.682515770  0.207713917 -0.31458942
#> 52 -0.42106769  0.62459118  0.242497221 -0.424305121 -0.146875776 -0.19632356
#> 53 -0.59232656  0.07993753 -0.190317827 -0.070619477 -0.338481716  0.38670748
#> 54 -0.57221946  0.40212051  0.106419396  0.021226610  0.197312711 -0.02476889
#> 55 -0.22197851  0.13827092 -0.377744870 -0.369154036 -0.516563574  0.05925417
#> 56  0.33523923 -0.05841630 -0.387605165  0.021181658 -0.307304652 -0.38285172
#> 57 -0.44253610 -0.15718439 -0.354677621  0.289479001  0.184939157  0.54963038
#> 58 -0.26504126 -0.06600722 -0.030974064  0.069004324  0.175574863 -0.14049308
#> 59 -0.15194276  0.04195497  0.206457100 -0.617784319  0.020796330 -0.01790714
#> 60 -0.07906070 -0.20594262 -0.169993811  0.374809547  0.052118971  0.17711634
#> 61 -0.17087139 -0.11731958 -0.299139099 -0.054268439 -0.075029182  0.17782044
#> 62 -0.48939276 -0.20269042  0.249975877 -0.144630383  0.280405959  0.38708208
#> 63  0.48339770 -0.14831873 -0.368465232  0.690654021  0.155281964 -0.51837263
#> 64 -0.23101011 -0.20516781  0.297230297 -0.077108756  0.259948758  0.20304456
#> 65 -0.45359595  0.18868561  0.014674506  0.209322276 -0.014175888 -0.20067706
#> 66  0.22705883 -0.10119422  0.429187256 -0.382228756  0.310808333 -0.10710896
#> 67 -0.11647730  0.16728402  0.007899385 -0.231485396 -0.245882059 -0.79138494
#>              V7           V8           V9        Mean.A       Mean.B
#> 1  -0.285732341 -0.466613133  0.360883652 -0.1080106386 -0.148925001
#> 2   0.078419099 -0.088399049 -0.099889832  0.0908107038  0.230219151
#> 3   0.118177746  0.134418275  0.332333011  0.1437061339 -0.240654848
#> 4  -0.009795061 -0.395227591 -0.356709796  0.0005454942  0.099806949
#> 5  -0.047346138 -0.061105724 -0.437753732  0.0855553566 -0.015607409
#> 6  -0.300184156 -0.808983194  0.003591927  0.0194359701 -0.232507060
#> 7  -0.463814354 -0.029597566  0.248111520  0.0188004604  0.181698033
#> 8  -0.332773137  0.030946284 -0.324466751  0.1613129861 -0.181825237
#> 9  -0.076799881  0.412087690  0.305872291 -0.3240271280  0.241659889
#> 10  0.753774231  0.003489033  0.152112837  0.0500425104 -0.143195855
#> 11  0.224825437  0.383504503  0.032022720  0.0003376848  0.119223148
#> 12  0.196938096  0.478362396  0.077796542  0.1617931652  0.025687548
#> 13  0.294015596  0.055872373 -0.251172029 -0.1654507368  0.311749470
#> 14 -0.443068995  0.026569863  0.001300993  0.1959918976  0.150820401
#> 15  0.033002508 -0.614577422  0.222120424 -0.1378413365  0.054996997
#> 16  0.536289030  0.153391794  0.203108243 -0.2327049493 -0.098904056
#> 17  0.405239998 -0.296580311 -0.473355316 -0.3276817469  0.150152262
#> 18  0.174209929 -0.324921086 -0.096147541  0.0456383789 -0.010048759
#> 19 -0.145141280  0.080619290  0.103099389  0.1339879690 -0.066871565
#> 20  0.110456508  0.173691041  0.135950402  0.1412029463 -0.149784616
#> 21  0.093987874 -0.394370811  0.003086572  0.1161281239  0.149019133
#> 22  0.489297686  0.197254908  0.270265884 -0.0374592974 -0.044978209
#> 23 -0.253793239 -0.028246187 -0.149779188 -0.0012463104  0.048499029
#> 24 -0.412187611 -0.188675757 -0.042078216  0.0959229825  0.123074567
#> 25  0.289454401  0.001548362  0.167513264 -0.0971652349  0.112798176
#> 26 -0.535443347 -0.645418718 -0.032744347  0.0841017661 -0.152662957
#> 27 -0.235060871  0.152427347 -0.241725961 -0.1536841457  0.304524389
#> 28 -0.088331186  0.442633433  0.343447377  0.0664332936 -0.054293556
#> 29  0.041949931 -0.520912586  0.221511020 -0.2805740456 -0.005504386
#> 30  0.084395071  0.377556400  0.373521526 -0.1076943901  0.239231821
#> 31 -0.350753237 -0.006892884 -0.210546636 -0.2944021629  0.133421154
#> 32  0.185424022  0.329179985  0.103449243 -0.0068622674 -0.195341551
#> 33 -0.204438730  0.285859290 -0.427073979  0.3403293172  0.119221284
#> 34 -0.211975756 -0.108562212  0.721644055 -0.2682993378  0.048363791
#> 35 -0.568948050 -0.112473455  0.297903398  0.1043312177 -0.234325595
#> 36 -0.253832563 -0.136478918 -0.314659725  0.1107555981 -0.042050727
#> 37  0.217411912  0.285831820  0.068474410 -0.2035608214  0.254028870
#> 38 -0.209225204 -0.287449316 -0.107731018 -0.1295272094  0.161135695
#> 39 -0.090835513  0.142507253  0.149405044 -0.1222979423 -0.202324443
#> 40 -0.045139268 -0.404961202 -0.016167612  0.1867929714  0.132434838
#> 41 -0.188325958  0.037457502  0.228068178  0.0390415478  0.136708100
#> 42  0.732864158  0.028886177  0.041278479 -0.1287088184 -0.176112545
#> 43 -0.147269818  0.016120394  0.439677557 -0.0870355026  0.231050940
#> 44 -0.509545126 -0.002913621  0.197340192 -0.0861191473  0.042454685
#> 45 -0.109667770 -0.049975485  0.067824681  0.1222331625 -0.216983298
#> 46 -0.212523971  0.242466932  0.754568197 -0.0772603618 -0.185096590
#> 47  0.402777706 -0.056825408  0.406140786  0.3159412816  0.266952311
#> 48  0.390725903 -0.352821133 -0.026733466  0.0344328347  0.213469493
#> 49  0.041671610  0.039400169  0.214970845 -0.1188420808 -0.201550386
#> 50 -0.076405301  0.051075631  0.486965114 -0.2883598106 -0.117094888
#> 51 -0.062646356 -0.721018117  0.617243652 -0.0747880747 -0.263130423
#> 52 -0.248170119 -0.115697110 -0.130045751  0.1486735684 -0.255834820
#> 53  0.321213272 -0.122195654  0.476725552 -0.2342356188 -0.007464571
#> 54 -0.218859777 -0.033976502 -0.253048255 -0.0212265184  0.064590143
#> 55 -0.248433128  0.080486296  0.141315192 -0.1538174846 -0.275487815
#> 56 -0.019215905  0.191094673  0.082374379 -0.0369274128 -0.222991572
#> 57 -0.079154157  0.102875234 -0.557875531 -0.3181327034  0.341349512
#> 58  0.600938533 -0.669281572  0.243798503 -0.1206741828  0.034695369
#> 59 -0.620318294 -0.088271974 -0.083785609  0.0321564383 -0.204965044
#> 60  0.179238908  0.021802543 -0.020374441 -0.1516657112  0.201348287
#> 61 -0.673368758 -0.342247182 -0.246493891 -0.1957766878  0.016174275
#> 62 -0.077462911  0.195169605  0.527125298 -0.1473690977  0.174285885
#> 63  0.077693210 -0.217441830 -0.306625070 -0.0111287548  0.109187784
#> 64 -0.099635018  0.480375711  0.402180383 -0.0463158755  0.128628186
#> 65  0.254743668  0.431628565  0.076350187 -0.0834119427 -0.001843556
#> 66  0.055954839 -0.154606011  0.053454741  0.1850172903 -0.059509795
#> 67  0.375650441  0.214105422 -0.014497047  0.0195687020 -0.422917466
#>          Mean.C     FC.crude logFC.crude
#> 1  -0.130487274   1.82030636 -0.02747155
#> 2  -0.036623261  -1.09396618  0.38494878
#> 3   0.194976344  -7.49677018 -0.77197696
#> 4  -0.253910816  -0.39009233  0.56016976
#> 5  -0.182068531   0.37800350  0.29547515
#> 6  -0.368525141   0.64940279  0.28952482
#> 7  -0.081766800  -1.61978732  0.37063673
#> 8  -0.208764535   0.92720634  0.06041552
#> 9   0.213720033   1.05195723  0.02597753
#> 10  0.303125367  -0.76353795 -0.63526743
#> 11  0.213450887   0.55785123 -0.11665182
#> 12  0.251032345  -1.52517781 -0.33439393
#> 13  0.032905313   2.40577591  0.30180875
#> 14 -0.138399380   0.13508575  0.52056729
#> 15 -0.119818163  10.69946626  0.22862812
#> 16  0.297596356   0.25231108 -0.43264843
#> 17 -0.121565209   2.31827108  0.29311492
#> 18 -0.082286233   0.43531215  0.11481206
#> 19  0.012859133   1.65823053 -0.13720258
#> 20  0.140032650 248.64437808 -0.49442779
#> 21 -0.099098788  -0.15282015  0.39634056
#> 22  0.318939493  -0.02109690 -0.45066989
#> 23 -0.143939538  -0.34861738  0.29215594
#> 24 -0.214313861  -0.08751889  0.57447613
#> 25  0.152838676   0.83984050 -0.04696919
#> 26 -0.404535471   0.48454089  0.57778074
#> 27 -0.108119828  10.05630191  0.47991524
#> 28  0.232583208  -0.72661397 -0.40736994
#> 29 -0.085817212   1.41237488  0.09385905
#> 30  0.278490999   0.89834111 -0.04144940
#> 31 -0.189397585   4.07433015  0.36976512
#> 32  0.206017750  -0.88537800 -0.57973701
#> 33 -0.115217806   0.48536808  0.51661605
#> 34  0.133702029   0.78771655 -0.09060148
#> 35 -0.127839369   1.45865511 -0.21538671
#> 36 -0.234990402   0.44196122  0.37284098
#> 37  0.190572714   1.16100167  0.06421591
#> 38 -0.201468512  -4.04027856  0.47582429
#> 39  0.067025595  -0.42269705 -0.37047702
#> 40 -0.155422694   0.15884174  0.52367926
#> 41  0.025733240  -7.33876594  0.15376859
#> 42  0.267676271  -0.11959009 -0.55176005
#> 43  0.102842711   1.67521295  0.14763107
#> 44 -0.105039518  -6.79552388  0.20205866
#> 45 -0.030606191   2.21943140 -0.35845782
#> 46  0.261503719  -0.31832250 -0.58552128
#> 47  0.250697695   0.75086262  0.02487161
#> 48  0.003723768  -5.83009114  0.28260691
#> 49  0.098680875  -0.38022794 -0.40849649
#> 50  0.153878481   0.38726842 -0.30024212
#> 51 -0.055473607  -9.75136114 -0.32865596
#> 52 -0.164637660   1.29107530 -0.20557519
#> 53  0.225247724   0.49353486 -0.25059172
#> 54 -0.168628178  -0.58219604  0.34884237
#> 55 -0.008877214  -0.83945151 -0.38243789
#> 56  0.084751049  -1.52914621 -0.46267219
#> 57 -0.178051484   4.70785606  0.54159656
#> 58  0.058485155   0.86721437 -0.02940430
#> 59 -0.264125292   0.80032435  0.11645537
#> 60  0.060222337   1.66604017  0.15892033
#> 61 -0.420703277  -0.94231173  0.64492646
#> 62  0.214943997   0.88778183 -0.04371270
#> 63 -0.148791230  -0.87399663  0.37758185
#> 64  0.260973692   0.56931338 -0.15398666
#> 65  0.254240807   0.24157477 -0.30657876
#> 66 -0.015065477   1.22212966 -0.08247070
#> 67  0.191752939  -2.56984133 -1.07211984