Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeansBetweenABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeansBetweenABC(dat = dat, A.samples = 1:3, B.samples = 4:6,
 C.samples = 7:9)
#>              V1          V2           V3           V4           V5           V6
#> 1   0.409482460  0.03827492  0.049049814  0.382501058 -0.119676291  0.601969251
#> 2   0.459008437  0.36344355 -0.231611995  0.169995061 -0.373575622  0.063350177
#> 3   0.025457686  0.14338734  0.350042884  0.004231511 -0.336787579 -0.104769326
#> 4   0.042526366 -0.13051489 -0.331300666 -0.152313034  0.382205082  0.073625457
#> 5   0.431459558  0.06706657 -0.262938321  0.059645574 -0.598551899  0.262126155
#> 6   0.231541448 -0.01472675 -0.090138318 -0.125159544  0.068338792 -0.448114576
#> 7   0.021447476  0.22495076 -0.251058779  0.340376927 -0.495992550 -0.207555023
#> 8  -0.122228446  0.14124675  0.465874277  0.282598153  0.111634332  0.240155717
#> 9  -0.068046346  0.06359465 -0.281853837  0.680547019  0.388192790  0.256683722
#> 10  0.158796030  0.58302810 -0.321958072 -0.206163518 -0.258768468  0.647881455
#> 11  0.117945238  0.50244052 -0.107500121 -0.060813173  0.075718766  0.001917374
#> 12  1.110461805 -0.50640455 -0.257667489 -0.157501269 -0.074767689  0.267916341
#> 13 -0.275183753  0.07125275  0.038897396 -0.245371976  0.323150176  0.287563264
#> 14 -0.185100834  0.28968133  0.197388525 -0.054436282  0.129910165  0.064017255
#> 15 -0.108106191 -0.04505098  0.191157753 -0.029308371 -0.286247232 -0.143509888
#> 16  0.397029472  0.23906336  0.141757495 -0.189995639  0.071446483 -0.044510317
#> 17  0.110770349  0.03141022  0.042954608  0.115433774  0.052858559  0.263239415
#> 18  0.606613226  0.66000994 -0.532344354  0.605697294  0.046194419  0.409085157
#> 19 -0.025967114  0.28837914 -0.056958297  0.508407620  0.021060588 -0.036100751
#> 20  0.079203667  0.15786933 -0.244200239 -0.004735808 -0.065298286 -0.434937721
#> 21  0.627427783 -0.12509261 -0.329084286  0.125601280  0.046200115 -0.009993636
#> 22 -0.184530408  0.06787775  0.215843650  0.282581962 -0.092583170  0.006514934
#> 23 -0.006494043 -0.67690260  0.244663186 -0.024687761  0.053994659 -0.080076054
#> 24 -0.041467405 -0.02770839 -0.157753132 -0.083946710 -0.256920403  0.121744428
#> 25  0.243888388 -0.22953115 -0.476072936  0.123379916  0.007235013  0.264965006
#> 26 -0.624897308  0.11125254 -0.085423987  0.187022516  0.228302764 -0.190180342
#> 27  0.112864797  0.20245068 -0.082171933 -0.098282327 -0.415835178  0.101801577
#> 28 -0.244849581 -0.24927326 -0.268307390 -0.101090451  0.351669643  0.231260383
#> 29 -0.057839224  0.45688731  0.291587380  0.323275168 -0.608034928  0.082100596
#> 30  0.323113517 -0.04480830 -0.415820501  0.547904389  0.228809347 -0.240413425
#> 31  0.153071659  0.17979532 -0.360263826  0.148425911 -0.194946980 -0.371880669
#> 32  0.613531559 -0.19592116 -0.134121982 -0.122154754 -0.825621332 -0.162780414
#> 33  0.171414346  0.14446895  0.022664753 -0.502578879 -0.102433728 -0.381198989
#> 34  0.289066713  0.10672847  0.224956265 -0.019456476  0.207759342  0.128863459
#> 35 -0.083752373 -0.04708000  0.343501959 -0.028078382 -0.195100275  0.047955586
#> 36 -0.085657605  0.12139743 -0.420859489  0.048996382  0.365024739  0.155426650
#> 37  0.514929057 -0.26089309 -0.730976161 -0.139060745  0.117373932 -0.468908016
#> 38  0.597654155  0.25085241  0.306903880 -0.577962631 -0.110937056  0.583508883
#> 39 -0.032986371  0.33517030  0.480331199  0.266438124 -0.033430948 -0.111024448
#> 40 -0.449780407  0.01511205 -0.209484219 -0.290602545 -0.064403543  0.350026203
#> 41 -0.323521260  0.05101313  0.323641230  0.553000242 -0.213793669 -0.067449666
#> 42  0.210271295  0.40141147 -0.242797716  0.080634715  0.114870956 -0.552706015
#> 43 -0.106555927 -0.09648969 -0.087273293  0.157235741  0.126549595 -0.247762843
#> 44 -0.225283099  0.29570017  0.375814522  0.016085087  0.371535618 -0.208193555
#> 45 -0.398412018  0.30037485 -0.154313174 -0.215301377 -0.480069971 -0.705135258
#> 46 -0.894905821  0.17659841  0.201755688  0.299020275  0.743038027 -0.104785048
#> 47 -0.203105920  0.40282360  0.245043308  0.317726041 -0.095899105  0.222893969
#> 48  0.487539692  0.19305343 -1.060943552  0.380436694 -0.059881629  0.220794171
#> 49  0.062226531 -0.13602499 -0.020361392 -0.263137107 -0.234812158 -0.098817664
#> 50 -0.243984938  0.71569881  0.142538283  0.219652790 -0.346985436 -0.543913561
#> 51 -0.240126509 -0.01708441 -0.341044935  0.629411772  0.277444432  0.052621809
#> 52  0.144364530 -0.39488039  0.155189026  0.155437262  0.210841650 -0.038133104
#> 53  0.336056848  0.21970611  0.339925559  0.148387334 -0.045127391  0.012235907
#> 54 -0.140142060 -0.29281623  0.036103122 -0.466307316 -0.198199609  0.413687706
#> 55  0.166722199 -0.52048745  0.148915294 -0.219876439 -0.170681088  1.047637280
#> 56 -0.225363735  0.22733155 -0.311229758 -0.064467495 -0.140878353 -0.263235129
#> 57 -0.090536227  0.20219959  0.234458463  0.141908850  0.148227433  0.357399875
#> 58  0.069597446 -0.11556451 -0.355469162  0.652099217 -0.255903483 -0.157979352
#> 59 -0.229753484  0.30042653  0.505499412  0.204236686 -0.030652765 -0.039390705
#> 60 -0.011239593 -0.16150863  0.363464087 -0.232624300 -0.371150467 -0.381723680
#> 61 -0.069946557 -0.43792302  0.059861553  0.549096731 -0.163361279  0.662036805
#> 62  0.360618893 -0.26863479 -0.255952385  0.263696041  0.374131179 -0.026242722
#> 63 -0.319407339  0.09145578  0.002818232  0.191354717  0.494108976 -0.388689116
#> 64  0.004793190  0.20319480  0.146449914  0.079882582 -0.027702462  0.189546316
#> 65  0.115674769  0.25775986 -0.397196049 -0.218173312 -0.327665146  0.193957435
#> 66  0.097643015 -0.15161971  0.307745973 -0.203598666  0.137727642 -0.157647119
#> 67  0.024730352  0.25373879  0.262164429  0.509532276  0.061059633 -0.045392818
#>              V7           V8           V9       Mean.A       Mean.B
#> 1   0.196084085 -0.081903623 -0.155745740  0.165602399  0.288264673
#> 2   0.023859683  0.130994781  0.047859126  0.196946664 -0.046743461
#> 3  -0.089035652 -0.684111496  0.970052142  0.172962636 -0.145775131
#> 4   0.038209328  0.093540334  0.308790028 -0.139763064  0.101172501
#> 5   0.138157191 -0.320496386 -0.068237082  0.078529269 -0.092260057
#> 6   0.188476807 -0.541064609  0.337725521  0.042225461 -0.168311776
#> 7   0.524391379  0.049968014 -0.090960953 -0.001553514 -0.121056882
#> 8   0.274875771 -0.025853531  0.170189763  0.161630859  0.211462734
#> 9   0.120496379  0.030816646  0.233501048 -0.095435179  0.441807844
#> 10  0.520730525 -0.288590855 -0.103394869  0.139955355  0.060983156
#> 11 -0.166202563  0.183171820 -0.583054364  0.170961878  0.005607656
#> 12  0.036066342 -0.475906955  0.262222115  0.115463255  0.011882461
#> 13 -0.159734436  0.017996501  0.427767999 -0.055011202  0.121780488
#> 14 -0.493325029 -0.060130918 -0.564399713  0.100656341  0.046497046
#> 15  0.004976401 -0.146780078  0.047013670  0.012666859 -0.153021831
#> 16 -0.023037968  0.337326277  0.049833011  0.259283443 -0.054353158
#> 17 -0.500814690  0.004562871 -0.454196019  0.061711727  0.143843916
#> 18 -0.201649104 -0.428079238 -0.734043712  0.244759604  0.353658956
#> 19  0.448459838  0.139661728  0.350437605  0.068484577  0.164455819
#> 20  0.116523271  0.260895772  0.182989203 -0.002375747 -0.168323938
#> 21 -0.112271395  0.247181168  0.016308330  0.057750297  0.053935920
#> 22 -0.290964359 -0.179496558  0.388314219  0.033063662  0.065504575
#> 23  0.683181119 -0.579772989  0.083319937 -0.146244487 -0.016923052
#> 24  0.527811006  0.098506833  0.380350790 -0.075642974 -0.073040895
#> 25 -0.139464252  0.110428242 -0.034620259 -0.153905232  0.131859978
#> 26 -0.280008313 -0.425653694  0.174525058 -0.199689584  0.075048313
#> 27 -0.197361455  0.418598582  0.536330125  0.077714514 -0.137438642
#> 28 -0.086975040  0.447901727  0.019876713 -0.254143410  0.160613192
#> 29  0.005861559  0.497140769 -0.341449379  0.230211823 -0.067553055
#> 30  0.342178068  0.041569213 -0.369629115 -0.045838428  0.178766771
#> 31  0.256166691 -0.151973768 -0.261825563 -0.009132283 -0.139467246
#> 32 -0.135290970  0.127832234 -0.488547092  0.094496138 -0.370185500
#> 33  0.587160323 -0.172670424  0.432224380  0.112849351 -0.328737198
#> 34 -0.004567267  0.531274438 -0.044604115  0.206917148  0.105722108
#> 35  0.596965644 -0.212444099  0.499635307  0.070889862 -0.058407690
#> 36 -0.272903140 -0.655730870 -0.313639735 -0.128373222  0.189815924
#> 37  0.253023354  0.252681358 -0.220493265 -0.158980064 -0.163531609
#> 38  0.095204490  0.051346340  0.807290516  0.385136817 -0.035130268
#> 39  0.670710169  0.157134616  0.070378101  0.260838375  0.040660909
#> 40  0.279729030  0.541658263  0.017166250 -0.214717526 -0.001659962
#> 41 -0.001606500  0.215702203 -0.898615012  0.017044365  0.090585636
#> 42  0.163055284 -0.241715242  0.439235051  0.122961681 -0.119066782
#> 43 -0.330451549  0.075332735  0.122196519 -0.096772968  0.012007498
#> 44 -0.505052858  0.152235164  0.161124596  0.148743863  0.059809050
#> 45 -0.257680720  0.147125716  0.007929752 -0.084116780 -0.466835535
#> 46 -0.248698132 -0.360803712 -0.576327918 -0.172183908  0.312424418
#> 47 -0.054426725 -0.218077318  0.149276716  0.148253663  0.148240302
#> 48 -0.264222903  0.374297828 -0.126343975 -0.126783475  0.180449745
#> 49  0.179604150  0.399250337  0.458450839 -0.031386618 -0.198922309
#> 50  0.173777814  0.097348775  0.473904867  0.204750719 -0.223748735
#> 51  0.486816691 -0.517140487 -0.572366352 -0.199418619  0.319826004
#> 52 -0.229607304  0.057305557 -0.184941034 -0.031775611  0.109381936
#> 53 -0.306171049  0.101838161 -0.132297705  0.298562839  0.038498616
#> 54 -0.162560006 -0.118268081  0.070274335 -0.132285055 -0.083606406
#> 55  0.194609015 -0.142658236 -0.047371944 -0.068283318  0.219026584
#> 56 -0.532326801 -0.127818456 -0.794470210 -0.103087314 -0.156193659
#> 57 -0.368183971 -0.075332009 -0.103724608  0.115373942  0.215845386
#> 58  0.171866746 -0.087356531 -0.321344373 -0.133812075  0.079405460
#> 59  0.290791616  0.083258587  0.231636454  0.192057487  0.044731072
#> 60 -0.056498748 -0.210315989  0.090974771  0.063571953 -0.328499482
#> 61  0.377441414 -0.294928390 -0.618225456 -0.149336008  0.349257419
#> 62  0.079393249  0.139850692 -0.515290908 -0.054656094  0.203861499
#> 63  0.078488368 -0.228877265  0.198757739 -0.075044443  0.098924859
#> 64 -0.262141716  0.039633613 -0.229202083  0.118145968  0.080575479
#> 65 -0.010462966  0.066825726  0.409482460 -0.007920475 -0.117293674
#> 66  0.468276200  0.150947370  0.459008437  0.084589760 -0.074506048
#> 67  0.187531305 -0.182884424  0.025457686  0.180211192  0.175066363
#>          Mean.C      FC.crude  logFC.crude
#> 1  -0.013855093 -6.835172e-01  0.452274013
#> 2   0.067571196  1.883588e+00 -0.203073218
#> 3   0.065634998  2.969764e+00 -0.389920530
#> 4   0.146846563  8.406402e-01 -0.052146182
#> 5  -0.083525426  1.053899e+00 -0.015117391
#> 6  -0.004954094  4.462468e+00 -0.271333151
#> 7   0.161132813 -7.345631e-01 -0.401072545
#> 8   0.139737334 -2.276101e+00  0.102094878
#> 9   0.128271358  2.401553e+00  0.329087639
#> 10  0.042914934  8.138073e-01  0.028583298
#> 11 -0.188695035  4.597554e-01  0.382318963
#> 12 -0.059206166  5.930105e-01  0.119201465
#> 13  0.095343355  1.175832e+00  0.032780374
#> 14 -0.372618553  1.144352e-01  0.844547005
#> 15 -0.031596669  3.743233e+00 -0.196027071
#> 16  0.121373774  2.274218e+00 -0.328866422
#> 17 -0.316815946 -2.169780e-01  0.800114686
#> 18 -0.454590685 -1.557150e-01  1.882972952
#> 19  0.312853057  3.927317e-01 -0.183203813
#> 20  0.186802749 -8.772043e-01 -0.511756373
#> 21  0.050406034  5.193683e-01  0.005121126
#> 22 -0.027382233 -5.366934e-01  0.136011070
#> 23  0.062242689  6.202848e-01 -0.097745990
#> 24  0.335556210  6.328026e-03 -0.493172501
#> 25 -0.021218756  2.153688e+00  0.182878638
#> 26 -0.177045649  1.213296e+01  0.317896727
#> 27  0.252522418 -1.230798e+00 -0.581941820
#> 28  0.126934467  1.088378e+00  0.034759196
#> 29  0.053850983  1.688384e+00 -0.230058271
#> 30  0.004706056  4.443713e+00  0.221179461
#> 31 -0.052544213  3.002284e+00 -0.137437964
#> 32 -0.165335276  1.788397e+00 -0.467456766
#> 33  0.282238093 -2.606942e+00 -1.066349009
#> 34  0.160701018  2.189604e+00 -0.085654299
#> 35  0.294718951 -5.776620e-01 -0.491150411
#> 36 -0.414091248 -1.113647e+00  0.883991778
#> 37  0.095070482 -1.791591e-02 -0.333176981
#> 38  0.317947115  6.254933e+00 -0.686195318
#> 39  0.299407629 -5.708627e+00 -0.413379672
#> 40  0.279517848  4.310852e-01 -0.300759407
#> 41 -0.228173103 -2.999023e-01  0.508244715
#> 42  0.120191698  8.737541e+01 -0.395782634
#> 43 -0.044307432  2.073370e+00  0.075200753
#> 44 -0.063897699  4.182381e-01  0.210533724
#> 45 -0.034208417 -7.668429e+00 -0.766263551
#> 46 -0.395276587 -2.172229e+00  0.934267953
#> 47 -0.041075776  7.057140e-05  0.302793064
#> 48 -0.005423017  2.531576e+00  0.221266449
#> 49  0.345768442 -4.442091e-01 -0.726230684
#> 50  0.248343819 -9.829525e+00 -0.868732528
#> 51 -0.200896716 -3.512928e+02  0.605488206
#> 52 -0.119080927 -1.616826e+00  0.322293749
#> 53 -0.112210197  6.331093e-01  0.328576607
#> 54 -0.070184584  7.838692e-01 -0.018347593
#> 55  0.001526279  4.115622e+00  0.267005354
#> 56 -0.484871823  1.391003e-01  0.615092601
#> 57 -0.182413530 -3.373931e-01  0.648142069
#> 58 -0.078944719  3.886055e+00  0.201776657
#> 59  0.201895552 -1.497514e+01 -0.244058496
#> 60 -0.058613322  3.208827e+00 -0.530014663
#> 61 -0.178570811 -1.705479e+01  0.626414736
#> 62 -0.098682322 -5.871900e+00  0.396682443
#> 63  0.016122947  1.908240e+00  0.105522249
#> 64 -0.150570062  1.398148e-01  0.396249142
#> 65  0.155281740 -6.701698e-01 -0.385208991
#> 66  0.359410669 -5.789072e-01 -0.600281244
#> 67  0.010034856  3.023234e-02  0.261681718