Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeansBetweenABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeansBetweenABC(dat = dat, A.samples = 1:3, B.samples = 4:6,
 C.samples = 7:9)
#>              V1           V2           V3           V4            V5
#> 1  -0.554030303  0.139335830 -0.034548710 -0.148281154  0.3120386609
#> 2  -0.036249991  0.453040202 -0.032413743  0.567712464 -0.0899267651
#> 3  -0.542578890  0.274099808 -0.405883156 -0.469338659 -0.1009516502
#> 4  -0.440026277 -0.061078205 -0.694828260  0.215593735  0.0008074557
#> 5   0.463795822  0.031989509  0.194986107 -0.143033910 -0.3341764039
#> 6  -0.103996578  0.184135278  0.043515351 -0.070415350  0.1613099483
#> 7  -0.233323120  0.401693315  0.116086647 -0.310076001  0.5090862510
#> 8  -0.031046136  0.121995957 -0.064305733 -0.518634031  0.1841969416
#> 9   0.030570471 -0.218855002 -0.221909077  0.547879418 -0.1029840823
#> 10 -0.407771148 -0.306498602  0.051075079  0.371684665 -0.2268472844
#> 11  0.514536097 -0.389475809  0.116386944  0.113081654  0.3672596573
#> 12 -0.205961637  0.092167551 -0.272640199 -0.174235759  0.1301994166
#> 13 -0.371106459 -0.600692389 -0.043660870  0.476418554  0.0297048011
#> 14 -0.552628251  0.207221352 -0.176931656  0.116282876 -0.2111175190
#> 15  0.044078781 -0.784483568 -0.201064466  0.079099648 -0.2619328836
#> 16 -0.125924147 -0.123509049  0.395017916 -0.018384562 -0.0597020545
#> 17  0.136692218  0.206486479  0.230467073  0.449743052  0.4250884313
#> 18  0.028433152 -0.167189587 -0.164951890  0.126616289  0.2557473778
#> 19 -0.093253780 -0.705640718 -0.164398307 -0.022992500 -0.4553022713
#> 20  0.347685329 -0.355713375 -0.115745002 -0.276154132 -0.1897479984
#> 21  0.035290022 -0.037171642  0.050497481 -0.370826951  0.0485807463
#> 22 -0.563349828  0.035859895  0.422319145 -0.050020625  0.3328605808
#> 23  0.212946787 -0.250260823 -0.783909409  0.006056613 -0.0527057297
#> 24 -0.435185376  0.458030735  0.337088207 -0.485359978 -0.0478244208
#> 25  0.397618736 -0.128187778  0.494506773  0.081349169  0.0703601963
#> 26  0.387400037  0.250427062 -0.328821470  0.354684674 -0.0069116936
#> 27 -0.076766180 -0.281201726 -0.318643399 -0.129590962 -0.3451212547
#> 28  0.063584224 -0.172469412  0.010867220 -0.086420427 -0.0269693368
#> 29 -0.414150643 -0.001345238 -0.671672859 -0.148703068 -0.3083147121
#> 30  0.048834256 -0.188445621 -0.153761773  0.023526187  0.0904915665
#> 31 -0.302498720  0.037750989  0.296106573 -0.365473330  0.0992882591
#> 32  0.072313799 -0.456739147 -0.030389394 -0.167232119  0.2513541385
#> 33  0.280981101  0.342840550  0.482792930  0.038071907 -0.1260842413
#> 34 -0.260174726  0.272502846 -0.209610989 -0.432880822 -0.5189421155
#> 35  0.285218311 -0.197203724 -0.403469969 -0.420123934  0.0356941697
#> 36  0.148884063  0.168913179 -0.168776864  0.326408457 -0.1004279654
#> 37  0.096107622  0.317464928 -0.184026224  0.018599799 -0.6507698034
#> 38 -0.038125806  0.059448600 -0.375776893 -0.246977289 -0.3790016935
#> 39 -0.250211130  0.066350764  0.272154535  0.491855508 -0.3063431286
#> 40 -0.474178383 -0.427112730 -0.026109758 -0.168830688 -0.2983377779
#> 41  0.197588397 -0.055267270 -0.127339466  0.420502183  0.0365760155
#> 42 -0.607136228  0.460829915 -0.023748895 -0.240546549  0.1774810369
#> 43 -0.085293857  0.374776062  0.242967606  0.301265336 -0.0095157477
#> 44  0.088994825  0.116865369  0.053285152 -0.063896474  0.1759980725
#> 45  0.001298689 -0.760927064  0.151297229 -0.463607799  0.2464268158
#> 46  0.612084886 -0.028599983 -0.144348027 -0.218553322 -0.2911654657
#> 47  0.130951330  0.167748377 -0.196027913 -0.738567613  0.5062061660
#> 48  0.291164073 -0.434937784 -0.151057534  0.054066531 -0.3209583080
#> 49 -0.086551774  1.019312253 -0.114377431  0.051847045 -0.0839974235
#> 50  0.021810124  0.002360983  0.291199234  0.069202309 -0.3769701486
#> 51  0.533972395 -0.162652033 -0.257974905 -0.467153894  0.0626996910
#> 52 -0.483359796 -0.358341733 -0.047987989 -0.241115253  0.3613000359
#> 53  0.561859622 -0.047977891 -0.414556574 -0.230136574 -0.6501683323
#> 54 -0.041562828  0.386351807 -0.303415452 -0.268081668 -0.4998149930
#> 55 -0.272024677 -0.075620848 -0.027139755  0.288986002  0.8747081709
#> 56  0.712675070  0.584309411 -0.627491951  0.502187731  0.4898816295
#> 57  0.139272522  0.129349681 -0.249272428  0.292837117  0.2308938515
#> 58 -0.214975620 -0.020593446 -0.245237299 -0.436989105 -0.4494943045
#> 59  0.437987508  0.359514628 -0.182717418  0.088797746 -0.1157537837
#> 60 -0.339904058 -0.309153521  0.510029051 -0.049162906 -0.0049440186
#> 61 -0.159883318 -0.369673645  0.266536912 -0.346124483  0.3670403242
#> 62 -0.225874925 -0.517900195  0.172169485 -0.452583376  0.1186232432
#> 63 -0.152796328  0.294992848  0.006978765 -0.460658213  0.0404305034
#> 64  0.244010717  0.802194620  0.330018704  0.608380926  0.1213648864
#> 65  0.309720553 -0.276584073  0.055820499 -0.172118090  0.0645421550
#> 66  0.356610179 -0.051994264  0.083082265 -0.720787599 -0.0542841509
#> 67 -0.011254404  0.175002536 -0.488121206 -0.080184290  0.1453960960
#>               V6           V7           V8            V9       Mean.A
#> 1   0.6085392522  0.135030737  0.402028929  0.0300999885 -0.149747728
#> 2   0.0088303467 -0.433948523  0.205595950  0.6351956269  0.128125489
#> 3   0.6723199147  0.779376829  0.018472127  0.0542946547 -0.224787413
#> 4   0.3915143360  0.300226774  0.501501022 -0.0561979306 -0.398644248
#> 5  -0.3919392097  0.245957763  0.014074743  0.0773902070  0.230257146
#> 6   0.0462256372 -0.219290677  0.578664336  0.1147128895  0.041218017
#> 7  -0.1018957214 -0.262889016 -0.133666344  0.6843493264  0.094818947
#> 8   0.3744985382 -0.312002551 -0.258878807 -0.0103981804  0.008881363
#> 9   0.2136102464  0.127028362 -0.711114666  0.4249181839 -0.136731203
#> 10 -0.1442185910 -0.178509667  0.310969493 -0.0295388675 -0.221064890
#> 11 -0.1021624008  0.264942699 -0.153553335  0.0728340615  0.080482411
#> 12 -0.1759422059  0.169296158  0.376713457  0.3889498078 -0.128811428
#> 13  0.3333212333  0.172675991  0.128203097 -0.1873415206 -0.338486573
#> 14  0.1839845973  0.337987614  0.294163877  0.3086210570 -0.174112852
#> 15 -0.8042382313  0.251147268 -0.292969668 -0.3646857675 -0.313823084
#> 16  0.0414739660 -0.203491579 -0.024395258 -0.0318414977  0.048528240
#> 17 -0.2315332815 -0.021387953  0.003560481  0.1989226494  0.191215257
#> 18  0.0123869504  0.254325846  0.120181062 -0.2621404777 -0.101236108
#> 19 -0.3940039544  0.273048518 -0.232989165 -0.1355894567 -0.321097602
#> 20 -0.1438474441  0.029430198 -0.037308810 -0.0089446062 -0.041257683
#> 21  0.2807610703 -0.067064077  0.234492696  0.0467908602  0.016205287
#> 22 -0.2418223950 -0.523940095  0.357083934 -0.0914450951 -0.035056929
#> 23 -0.1799829976 -0.193495289  0.124023006 -0.2460796722 -0.273741148
#> 24  0.3261713292 -0.321736677  0.353215441 -0.0168421424  0.119977856
#> 25  0.2772291330 -0.015992611  0.316682594 -0.3611105302  0.254645910
#> 26  0.3081108455 -0.154819651 -0.296751420  0.3535188894  0.103001876
#> 27  0.0439724915  0.022989505 -0.510483031  0.0581308859 -0.225537101
#> 28  0.2886963008  0.294151095 -0.078669045  0.3933429061 -0.032672656
#> 29  0.0341306747  0.132956723  0.002347445  0.4538405130 -0.362389580
#> 30  0.2436110326  0.036478186 -0.198740904 -0.2780979125 -0.097791046
#> 31  0.1731590112  0.080448877  0.060444183  0.4160264904  0.010452947
#> 32  0.0685005887  0.153223424  0.009979399 -0.0087369798 -0.138271581
#> 33 -0.5826220661 -0.555164601  0.289197680  0.1444153751  0.368871527
#> 34  0.2111018213  0.015083541 -0.555517563  0.1548056526 -0.065760956
#> 35  0.2913279701 -0.126357814  0.333202788  0.2952837635 -0.105151794
#> 36 -0.3025561682 -0.224544823  0.280973653 -0.7476776323  0.049673459
#> 37 -0.0190220334 -0.193404134  0.442390142 -0.1354061052  0.076515442
#> 38  0.0279663598 -0.007808324  0.118859490  0.0196911778 -0.118151366
#> 39 -0.3480120213  0.486485779 -0.124671185 -0.0333211910  0.029431390
#> 40 -0.1808074345 -0.103582743 -0.124923081 -0.3063383679 -0.309133624
#> 41 -0.3586646914 -0.017532343 -0.403660100  0.2982583357  0.004993887
#> 42 -0.5089630568  0.762438348 -0.058854654  0.0196639238 -0.056685070
#> 43 -0.4285649628 -0.370774064 -0.316892787  0.7099490213  0.177483270
#> 44  0.2483291052  0.253395825 -0.113993157  0.1894494353  0.086381782
#> 45 -0.0218262869 -0.260179263 -0.165649016 -0.1559124557 -0.202777049
#> 46  0.5216951878 -0.297219801  0.022000890  0.2679558283  0.146378959
#> 47 -0.2609038265 -0.081012169  0.137678561  0.0103733946  0.034223931
#> 48  0.0277577074 -0.327940246  0.345278483 -0.3995563813 -0.098277082
#> 49  0.4708043446 -0.512884759  0.308206191 -0.0623064365  0.272794349
#> 50  0.2387440193 -0.254050827 -0.392597779  0.3445925605  0.105123447
#> 51  0.2271471518 -0.389287099  0.348635606 -0.0762562996  0.037781819
#> 52 -0.0130438283  0.228539683  0.004195598 -0.2383588628 -0.296563173
#> 53  0.3853140285 -0.078846494  0.350036809 -0.1794410391  0.033108386
#> 54 -0.3892041116 -0.389662937  0.206158548  0.1519952498  0.013791176
#> 55 -0.0567185718 -0.029360398  0.417792689  0.0075574451 -0.124928427
#> 56 -0.2234004473  0.400558994 -0.242141792  0.0159485335  0.223164176
#> 57  0.0890861791 -0.264668814 -0.079667017 -0.0221292927  0.006449925
#> 58  0.2422854045  0.624289728  0.271185847 -0.3441582471 -0.160268788
#> 59  0.0489984552  0.681519041 -0.640298147 -0.0918939134  0.204928239
#> 60 -0.1887303301 -0.002917345 -0.392230704 -0.1777917036 -0.046342843
#> 61 -0.2668298443 -0.098010423 -0.195314931  0.0399221122 -0.087673351
#> 62 -0.2655305460  0.305364989  0.194453924  0.0479491694 -0.190535211
#> 63 -0.1518128983 -0.122401484 -0.195609196  0.0005049832  0.049725095
#> 64 -0.0440793973  0.135426566 -0.165989849  0.0280848826  0.458741347
#> 65  0.3798516606 -0.584744115  0.301223349 -0.5540303028  0.029652326
#> 66  0.1687864857 -0.260876927 -0.321956968 -0.0362499913  0.129232727
#> 67  0.0003283316 -0.106916313  0.006512939 -0.5425788902 -0.108124358
#>           Mean.B        Mean.C     FC.crude   logFC.crude
#> 1   0.2574322530  0.1890532182   1.20182658  0.0718653881
#> 2   0.1622053486  0.1356143514   4.55073941  0.0375837159
#> 3   0.0340098685  0.2840478704   0.50860719 -0.2613893502
#> 4   0.2026385089  0.2485099551   0.92911821 -0.0407475399
#> 5  -0.2897165077  0.1124742375   4.41467832 -0.8780201204
#> 6   0.0457067451  0.1580288496   0.03842733 -0.1529234653
#> 7   0.0323715096  0.0959313222 -56.13883738 -0.0946324457
#> 8   0.0133538162 -0.1937598460  -0.02207080  0.3331370486
#> 9   0.2195018606 -0.0530560400   4.25733337  0.3236727550
#> 10  0.0002062633  0.0343069861   0.86646641 -0.0397312430
#> 11  0.1260596368  0.0614078084  -2.38941944  0.0920842985
#> 12 -0.0733261827  0.3116531408   0.12596983 -0.4486277636
#> 13  0.2798148627  0.0378458558   1.64296613  0.2336513893
#> 14  0.0297166514  0.3135908493   0.41793717 -0.3054561408
#> 15 -0.3290238224 -0.1355027223  -0.08524398 -0.2588302471
#> 16 -0.0122042167 -0.0865761117   0.44952258  0.1190100786
#> 17  0.2144327339  0.0603650591  -0.17743555  0.2354360523
#> 18  0.1315835390  0.0374554768   1.67868618  0.1145847137
#> 19 -0.2907662420 -0.0318433679   0.10486056 -0.3234283934
#> 20 -0.2032498582 -0.0056077394  -4.54396725 -0.3055008252
#> 21 -0.0138283782  0.0714064930  -0.54407625 -0.1215115377
#> 22  0.0136725203 -0.0861004187  -0.95466533  0.1442286634
#> 23 -0.0755440383 -0.1051839851   1.17584507  0.0361369497
#> 24 -0.0690043564  0.0048788739   1.64191037 -0.1257825324
#> 25  0.1429794994 -0.0201401821   0.40637577  0.2926949503
#> 26  0.2186279420 -0.0326840606  -0.85215954  0.3682260069
#> 27 -0.1435799085 -0.1431208802   0.99443037 -0.0006119441
#> 28  0.0584355122  0.2029416521   0.38668351 -0.1794343502
#> 29 -0.1409623685  0.1963815605   0.39627532 -0.3518412347
#> 30  0.1192095953 -0.1467868769  -4.42896133  0.3558063576
#> 31 -0.0310086865  0.1856398503  -0.23667085 -0.2939821397
#> 32  0.0508742029  0.0514886142   0.99676217 -0.0007452235
#> 33 -0.2235448001 -0.0405171820   1.44707539 -0.5351126765
#> 34 -0.2469070386 -0.1285427896   2.88532641 -0.1947788189
#> 35 -0.0310339314  0.1673762460   0.27196417 -0.2445451378
#> 36 -0.0255252255 -0.2304162674   0.26848070  0.3613263435
#> 37 -0.2170640125  0.0378599676   7.59477048 -0.4445262522
#> 38 -0.1993375410  0.0435807812  -0.50197920 -0.3384330049
#> 39 -0.0541665474  0.1094978011  -1.04410746 -0.2370674126
#> 40 -0.2159919669 -0.1782813970   0.71180796 -0.0489300510
#> 41  0.0328045023 -0.0409780358  -0.60494784  0.1074708269
#> 42 -0.1906761897  0.2410825395  -0.44998555 -0.5835783401
#> 43 -0.0456051247  0.0074273901   1.31185346 -0.0952637412
#> 44  0.1201435677  0.1096173677   1.45302065  0.0147654337
#> 45 -0.0796690899 -0.1939135781  13.88936305  0.1547656657
#> 46  0.0039921332 -0.0024210277   0.95690080  0.0108288945
#> 47 -0.1644217579  0.0223465954  16.72476819 -0.3022498902
#> 48 -0.0797113566 -0.1274060483  -0.63736297  0.0691875045
#> 49  0.1462179888 -0.0889950017   0.34986204  0.4526488975
#> 50 -0.0230079400 -0.1006853483   0.62257488  0.1346243594
#> 51 -0.0591023503 -0.0389692643   1.26231664 -0.0318086296
#> 52  0.0357136515 -0.0018745274   1.12755218  0.0412886893
#> 53 -0.1649969593  0.0305830919  78.44842593 -0.3148675361
#> 54 -0.3857002575 -0.0105030465  16.44388643 -0.7002613443
#> 55  0.3689918671  0.1319965790   1.92242982  0.2492046029
#> 56  0.2562229711  0.0581219118  -0.20030502  0.3071472880
#> 57  0.2042723825 -0.1221550413  -1.53821787  0.4590152902
#> 58 -0.2147326684  0.1837724427  -0.15830626 -0.5073729204
#> 59  0.0073474724 -0.0168910065   0.89072869  0.0442508859
#> 60 -0.0809457516 -0.1909799175   0.23923955  0.1745858033
#> 61 -0.0819713343 -0.0844677471   1.77876530  0.0035855946
#> 62 -0.1998302263  0.1825893609  -0.02491129 -0.4709350985
#> 63 -0.1906802025 -0.1058352322   1.54541522 -0.1527645159
#> 64  0.2285554715 -0.0008261333   0.50087503  0.5103956529
#> 65  0.0907585750 -0.2791836895  -0.19785985  0.6184691736
#> 66 -0.2020950881 -0.2063612954   0.98728759  0.0092340528
#> 67  0.0218467125 -0.2143274213  -1.22379776  0.3382668319