Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeansBetweenABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeansBetweenABC(dat = dat, A.samples = 1:3, B.samples = 4:6,
 C.samples = 7:9)
#>             V1          V2           V3           V4           V5           V6
#> 1  -0.66427340  0.10975690 -0.251085282  0.035086941 -0.652460460  0.108754124
#> 2  -0.42419205 -0.21657822 -0.012345443  0.048729245  0.306528062  0.253259598
#> 3   0.11223517 -0.46524440  0.194005973 -0.163217774 -0.302770998  0.775664847
#> 4  -0.23497745 -0.38217469 -0.160455226 -0.447463268 -0.568668630 -0.246818331
#> 5   0.06208699 -0.32674936  0.196244612  0.003325169 -0.248130497  0.333621656
#> 6  -0.17081092 -0.04646548 -0.382514220  0.513207479  0.142750596  0.080626294
#> 7   0.23565277  0.20991854  0.012401022  0.356778277  0.041910912  0.825061398
#> 8  -0.41994435 -0.19722707  0.470715515 -0.331877400 -0.104818079 -0.271633262
#> 9   0.15055578  0.55324791  0.117404416  0.045978450  0.246606918 -0.060212087
#> 10 -0.07621769  0.05982294 -0.189599365  0.190008577 -0.447673466 -0.078264881
#> 11 -0.50418652  0.45799165  0.089232373 -0.665373502 -0.021380967  0.263210269
#> 12 -0.68596139  0.34737657 -0.382365781 -0.124960423  0.206958717 -0.320973194
#> 13 -0.75808154  0.26195107  0.055598870  0.197788609 -0.146409137 -0.505793350
#> 14  0.28731152 -0.44153647  0.167346621  0.104863341 -0.215949270  0.046195351
#> 15 -0.02093090  0.41786619 -0.163236130  0.038449735  0.192919829  0.047226183
#> 16 -0.14425977 -0.10788495  0.393437724  0.145298234  0.217826667  0.393732481
#> 17 -0.10406921 -0.01816259 -0.425897678 -0.358015333  0.180758660  0.394179467
#> 18 -0.07741946 -0.83515648  0.127117045  0.103459706  0.185919137 -0.245222482
#> 19  0.20742122  0.60108190 -0.076890730  0.261269542  0.065112330 -0.264680485
#> 20  0.52558074 -0.24529919 -0.296300032 -0.349730189 -0.220430296  0.164088993
#> 21  0.34189362 -0.20038968  0.007367026  0.452270116 -0.621752259 -0.021818546
#> 22  0.31055256  0.07226791 -0.143332512  0.022216236 -0.412658775  0.247942779
#> 23  0.22724405 -0.01202110  0.018324439 -0.471340846 -0.204684982 -0.581664402
#> 24 -0.03030553  0.23218497  0.471903334 -0.262124586 -0.238697194  0.245916772
#> 25  0.07309502  0.14262769 -0.530019763  0.371559377 -0.193795718 -0.104555647
#> 26 -0.04629941 -0.53578318 -0.093474955  0.037965315 -0.088639378 -0.210856111
#> 27 -0.44033181 -0.31157437  0.474022223 -0.045655975 -0.315116322 -0.135751164
#> 28 -0.25221323 -0.70095036 -0.100267639 -0.115316721 -0.419988383 -0.469622690
#> 29 -0.26055047 -0.08825134 -0.215198995 -0.309657895 -0.646474020  0.236860518
#> 30  0.26645143  0.38420304 -0.152029633  0.115105640 -0.323883413  0.079149151
#> 31  0.28236476 -0.79787998 -0.476953240 -0.264715078 -0.324075749  0.025801611
#> 32 -0.11077894 -0.50604470 -0.241834622 -0.409829323 -0.078478002  0.124035593
#> 33  0.12687991  0.26001320 -0.048374270 -0.096586708 -0.192459975  0.238924434
#> 34 -0.10939587  0.47147313  0.248852379 -0.135414258  0.133463061  0.193920390
#> 35 -0.49699156  0.10406728  0.442999334 -0.482811469 -0.071522156  0.133423271
#> 36  0.03929513  0.04439492 -0.340603717 -0.072780050  0.172043614 -0.081847819
#> 37  0.06721275 -0.40005708 -0.020895874 -0.761396927  0.116898152  0.192880056
#> 38 -0.14684905  0.03531316  0.332825416 -0.278918328 -0.109682221 -0.307898696
#> 39  0.12803407 -0.04393230  0.395305802  0.061548882 -0.069815395 -0.308528517
#> 40  0.37910767  0.22110122  0.286945649  0.341135526 -0.109558750  0.341986946
#> 41 -0.19383935 -0.13359729  0.121814023 -0.329007206 -0.199430599 -0.376126256
#> 42  0.23403248  0.37502686  0.513781945  0.491332627  0.055429161 -0.307217066
#> 43 -0.46208252  0.26848897  0.313389145  0.161054883  0.039038235  0.347899133
#> 44 -0.24722240 -0.01829475 -0.035578499  0.149265615  0.095031931 -0.108866920
#> 45 -0.02244314 -0.40957508  0.011266521 -0.170352028 -0.534585487  0.199142031
#> 46  0.21062556 -0.61885780 -0.098493136  0.246125344  0.543486890  0.346036105
#> 47  0.10273683  0.42324244 -0.215252874 -0.025312540 -0.533472576  0.027594523
#> 48 -0.10809970  0.33087768 -0.327294359 -0.404751557 -0.009125331  0.540683100
#> 49  0.61339930 -0.29531194 -0.085794087 -0.121302178 -0.095612008 -0.362320637
#> 50 -0.28757743  0.23743528  0.442972798 -0.120268198  0.489828996 -0.208927478
#> 51  0.30923638 -0.33708153  0.051721781 -0.231599184 -0.139626083 -0.159555118
#> 52 -0.26820748  0.20611683 -0.149460557 -0.102189043  0.064761254  0.137528754
#> 53  0.29722832  0.22924206  0.097423585 -0.207473281 -0.053306506  0.101321985
#> 54 -0.19974566 -0.41099875  0.192276607 -0.253244715  0.122226837 -0.330289319
#> 55 -0.22056856  0.15620511  0.202145412 -0.522505378  0.181076362  0.132387874
#> 56  0.35043677  0.21743830  0.083609414 -0.151687763 -0.046035322  0.164300657
#> 57  0.14891933 -0.47462420  0.159801658 -0.321381246  0.493323465  0.004647886
#> 58 -0.01342783  0.35289419  0.253506803  0.206526256 -0.331965098 -0.121864149
#> 59 -0.18350406  0.28605136  0.341078989 -0.142648099  0.084549786  0.803793429
#> 60 -0.23765201  0.11297802 -0.056649636  0.134846064 -0.033786899 -0.005794789
#> 61 -0.18037799  0.11442627  0.031674669  0.487548257  0.141182569 -0.353646936
#> 62 -0.52669086  0.35367276  0.119456559 -0.094654080 -0.163029671 -0.589554432
#> 63  0.08629150 -0.27123233 -0.025336976  0.276209077  0.084904072  0.326277558
#> 64 -0.03352074 -0.45833465 -0.157013559  0.389149931 -0.509624885 -0.104635310
#> 65  0.30330606  0.30563346  0.087068365  0.119715611  0.071707296  0.448918268
#> 66  0.09602616 -0.36799347 -0.615952925 -0.258072014  0.294431763 -0.066217910
#> 67  0.06687340 -0.45820319 -0.405716110  0.578312429 -0.094519015  0.059449819
#>             V7          V8           V9       Mean.A       Mean.B       Mean.C
#> 1  -0.82018599  0.41778170  0.168882612 -0.268533926 -0.169539798 -0.077840556
#> 2  -0.16117862 -0.37198242  0.041621745 -0.217705238  0.202838968 -0.163846434
#> 3   0.38565420 -0.08985008  0.101993373 -0.053001086  0.103225358  0.132599164
#> 4  -0.05854277  0.10988162 -0.660213712 -0.259202458 -0.420983410 -0.202958286
#> 5   0.07613925  0.34988621  0.309017952 -0.022805920  0.029605443  0.245014471
#> 6  -0.26465438  0.35397945 -0.103948013 -0.199930204  0.245528123 -0.004874316
#> 7   0.14937455 -0.51190264  0.006762914  0.152657443  0.407916863 -0.118588391
#> 8  -0.67367943 -0.26856509  0.156676403 -0.048818637 -0.236109580 -0.261856039
#> 9   0.06200646  0.22292320 -0.342681417  0.273736037  0.077457760 -0.019250586
#> 10  0.15006571  0.41806290  0.210376243 -0.068664705 -0.111976590  0.259501619
#> 11 -0.42071708  0.75589036  0.744647716  0.014345835 -0.141181400  0.359940331
#> 12  0.13611524 -0.21753601 -0.146804860 -0.240316869 -0.079658300 -0.076075209
#> 13  0.17527569  0.05928301  0.144320598 -0.146843866 -0.151471292  0.126293102
#> 14  0.29040405 -0.49169199  0.194987152  0.004373889 -0.021630193 -0.002100264
#> 15 -0.40403565 -0.10721002 -0.093297486  0.077899718  0.092865249 -0.201514385
#> 16 -0.03646451 -0.05225464  0.450442161  0.047097667  0.252285794  0.120574338
#> 17  0.12560498  0.31169482  0.088281164 -0.182709829  0.072307598  0.175193655
#> 18  0.14593518 -0.02160251  0.268016922 -0.261819633  0.014718787  0.130783197
#> 19  0.41803572  0.23907513  0.485769801  0.243870796  0.020567129  0.380960218
#> 20  0.08551598 -0.17242322  0.409276823 -0.005339495 -0.135357164  0.107456525
#> 21 -0.28151997  0.02050801 -0.061218015  0.049623657 -0.063766897 -0.107409989
#> 22 -0.16714503 -0.18510142  0.221173735  0.079829321 -0.047499920 -0.043690903
#> 23 -0.05393622  0.02656677  0.159078144  0.077849130 -0.419230077  0.043902896
#> 24 -0.57819568 -0.07356083  0.152049258  0.224594257 -0.084968336 -0.166569082
#> 25 -0.03893349 -0.22051236 -0.262168412 -0.104765687  0.024402671 -0.173871419
#> 26 -0.08062958 -0.24407031 -0.302646143 -0.225185848 -0.087176725 -0.209115342
#> 27 -0.43816615  0.21913738 -0.116368642 -0.092627988 -0.165507821 -0.111799136
#> 28 -0.55709685 -0.15547940 -0.129224224 -0.351143744 -0.334975932 -0.280600155
#> 29 -0.53254193  0.25966492  0.466802614 -0.188000269 -0.239757132  0.064641865
#> 30  0.39470834  0.04703992 -0.153627255  0.166208280 -0.043209541  0.096040334
#> 31 -0.46706516 -0.24578200 -0.059550891 -0.330822820 -0.187663072 -0.257466014
#> 32  0.03302158  0.07268206  0.249535511 -0.286219421 -0.121423911  0.118413049
#> 33 -0.16384250 -0.19070823 -0.354423845  0.112839611 -0.016707416 -0.236324859
#> 34  0.08373365  0.27538507 -0.395343229  0.203643212  0.063989731 -0.012074838
#> 35 -0.14601840 -0.16184501  0.237715264  0.016691684 -0.140303451 -0.023382715
#> 36  0.08212139 -0.68818302  0.163142771 -0.085637891  0.005805248 -0.147639621
#> 37  0.38666636 -0.18777121  0.431495737 -0.117913399 -0.150539573  0.210130295
#> 38 -0.22288286 -0.48133550  0.454830881  0.073763175 -0.232166415 -0.083129161
#> 39  0.27025241 -0.52793127  0.061349800  0.159802524 -0.105598343 -0.065443019
#> 40 -0.27799622  0.05494790  0.073193021  0.295718180  0.191187907 -0.049951765
#> 41 -0.17697583 -0.40020972  0.413363901 -0.068540869 -0.301521354 -0.054607216
#> 42  0.20916197  0.07411921  0.775878643  0.374280428  0.079848241  0.353053274
#> 43 -0.04119387 -0.45003778 -0.043292116  0.039931866  0.182664084 -0.178174590
#> 44  0.02530145  0.08447881  0.075000554 -0.100365217  0.045143542  0.061593605
#> 45 -0.18227209 -0.45771235 -0.152492887 -0.140250567 -0.168598495 -0.264159109
#> 46  0.42282402 -0.03251594 -0.494680759 -0.168908460  0.378549446 -0.034790894
#> 47 -0.02924364  0.27164140 -0.506996077  0.103575463 -0.177063531 -0.088199439
#> 48 -0.47959332  0.43252183 -0.238256650 -0.034838791  0.042268737 -0.095109378
#> 49  0.29952500  0.11129143  0.114715548  0.077431091 -0.193078274  0.175177324
#> 50 -0.15370189  0.40931644 -0.233292031  0.130943550  0.053544440  0.007440838
#> 51  0.12830162  0.50824728 -0.170503569  0.007958879 -0.176926795  0.155348446
#> 52  0.13278027  0.10716306  0.293133589 -0.070517069  0.033366988  0.177692306
#> 53  0.09568716 -0.21940041  0.371197233  0.207964654 -0.053152601  0.082494661
#> 54  0.02813889  0.23669800  0.580075913 -0.139489268 -0.153769065  0.281637601
#> 55  0.21942684  0.32234941  0.315875663  0.045927321 -0.069680381  0.285883972
#> 56  0.04765713  0.23710639 -0.910828605  0.217161495 -0.011140809 -0.208688362
#> 57  0.09708753 -0.20054191  0.380337632 -0.055301071  0.058863368  0.092294416
#> 58  0.16818442 -0.51543975 -0.360767662  0.197657719 -0.082434330 -0.236007665
#> 59  0.22606504  0.07004279 -0.093659998  0.147875431  0.248565039  0.067482612
#> 60  0.57184562  0.02136291 -0.748831362 -0.060441211  0.031754792 -0.051874278
#> 61 -0.15124793  0.11539782  0.088551810 -0.011425684  0.091694630  0.017567231
#> 62 -0.26390415 -0.02831453 -0.602322693 -0.017853847 -0.282412728 -0.298180459
#> 63  0.28309088  0.22348365 -0.650672349 -0.070092600  0.229130235 -0.048032607
#> 64 -0.40407471  0.39517623  0.356628216 -0.216289650 -0.075036754  0.115909913
#> 65  0.12908638  0.07708990 -0.664273399  0.232002628  0.213447059 -0.152699039
#> 66  0.19823240  0.49128229 -0.424192048 -0.295973414 -0.009952720  0.088440879
#> 67  0.83857996 -0.01101835  0.112235170 -0.265681968  0.181081077  0.313265595
#>        FC.crude  logFC.crude
#> 1    0.51912727 -0.115618257
#> 2    7.80827229  0.430762140
#> 3    0.84173618 -0.036193683
#> 4   -2.87640383 -0.333544170
#> 5    0.19569594 -0.268651647
#> 6    2.28374714  0.274448927
#> 7   -0.94106301  0.784481420
#> 8    0.87914583  0.046443901
#> 9    0.66992232  0.184958563
#> 10  -0.13198150 -0.473315244
#> 11  -0.45002810 -0.672120909
#> 12   0.97818403 -0.004446910
#> 13  -0.01694178 -0.355079096
#> 14   4.01659956 -0.028641774
#> 15  -0.05356040  0.494188411
#> 16   2.79256157  0.166967527
#> 17   0.71253128 -0.113673543
#> 18   0.70437195 -0.125546901
#> 19  -1.62889058 -0.549923154
#> 20  -1.15267959 -0.355131159
#> 21   0.72207808  0.072823686
#> 22   1.03083719 -0.006283345
#> 23  14.64313262 -0.941772282
#> 24   0.79138959  0.181455398
#> 25  -1.86914101  0.278571387
#> 26   8.58772714  0.163511608
#> 27   3.80153714 -0.081245074
#> 28   0.22918896 -0.075204866
#> 29  -0.20486236 -0.401645382
#> 30   2.98452261 -0.234054725
#> 31   1.95155372  0.090897286
#> 32   0.40727209 -0.270115991
#> 33   0.37102007  0.419473345
#> 34   0.64738895  0.133545448
#> 35   3.91759182 -0.187381638
#> 36  -1.47484821  0.218579805
#> 37  -0.09945679 -0.457157205
#> 38   1.94993330 -0.280634849
#> 39   1.17827356 -0.076781957
#> 40   0.30239908  0.452626122
#> 41 -16.72070352 -0.402628062
#> 42  13.87054456 -0.472189329
#> 43  -0.65441538  0.547443268
#> 44   0.89843058 -0.020570453
#> 45   0.22878106  0.149358323
#> 46   4.08192545  0.448329964
#> 47   1.46337707 -0.168041194
#> 48  -1.27935585  0.196844972
#> 49  -2.76746590 -0.589583223
#> 50   0.62669968  0.073957058
#> 51  -1.25440137 -0.493280978
#> 52   0.41853398 -0.177271301
#> 53   2.08111316 -0.243162448
#> 54  -0.03390854 -0.527785255
#> 55  -0.48178578 -0.487531318
#> 56   0.53610985  0.426607740
#> 57   0.77349546 -0.042652025
#> 58   0.64587136  0.346157737
#> 59  -1.25247019  0.259318052
#> 60  10.76184502  0.114924952
#> 61   3.55674188  0.100357101
#> 62   0.94375229  0.031267530
#> 63  13.56404947  0.346169013
#> 64   0.42520494 -0.223191694
#> 65   0.04823366  0.673620500
#> 66   0.74404282 -0.106361884
#> 67   0.77168136 -0.126134607