Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeansBetweenABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeansBetweenABC(dat = dat, A.samples = 1:3, B.samples = 4:6,
 C.samples = 7:9)
#>              V1          V2           V3           V4            V5
#> 1   0.010724964 -0.33367554  0.205806365  1.014500919 -0.1322484040
#> 2   0.607295992  0.40372947  0.050480144 -0.222444415 -0.0207004137
#> 3  -0.926795515 -0.16587008  0.428456336  0.025977074 -0.0524162873
#> 4  -0.164636599 -0.25757498 -0.026941739  0.594337283 -0.1218728368
#> 5  -0.083729140  0.35861806 -0.070287829  0.007809283 -0.2505911623
#> 6  -0.403841421  0.09628306 -0.218701789  0.256099552  0.0381096321
#> 7  -0.666136645  0.01979520  0.259656472  0.251075061 -0.2089254415
#> 8   0.203749703  0.60762655 -0.276674706 -0.327633038 -0.0630265220
#> 9   0.405565587 -0.04444878  0.084011547  0.175689632 -0.5325737586
#> 10 -0.116496038  0.04735266 -0.415667025 -0.393706447 -0.1357835588
#> 11 -0.196755710  0.13836436 -0.587195348  0.129873569 -0.4038510577
#> 12 -0.045032030  0.14555323 -0.046547957 -0.159799818  0.1607902645
#> 13  0.035256306 -0.41543217  0.312384933 -0.833523498 -0.1758917671
#> 14  0.988593420  0.25729711 -0.174792502 -0.323846001  0.2226986720
#> 15 -0.110382962 -0.05330461 -0.494362062  0.104361797 -0.6737647009
#> 16  0.016823986 -0.35861071 -0.006429479 -0.523407064 -0.3615905282
#> 17 -0.149505286 -0.13809081 -0.204432554  0.029144326 -0.3430765874
#> 18 -0.016563889  0.02897639 -0.497728170 -0.126775266  0.0912651570
#> 19 -0.108355183  0.04565318  0.584364400 -0.144944336  0.1204575814
#> 20 -0.131359304  0.13226839 -0.124334543  0.212308128  0.1954689574
#> 21  0.140993712 -0.52185041 -0.355621831  0.063031367  0.2377395796
#> 22  0.230527974 -0.41484817 -0.558541385 -0.582043323 -0.4070053648
#> 23  0.681334392  0.20908382  0.091736822 -0.188575221  0.1669170536
#> 24 -0.005848180 -0.03654283  0.043936517 -0.318441017  0.1509197693
#> 25 -0.164080327  0.02636982  0.359676989  0.448441261 -0.1776107612
#> 26  0.052475526  0.43953927 -0.029370046  0.373769295 -0.0553399810
#> 27  0.516352454 -0.15482127  0.496143336  0.255141703  0.1332776852
#> 28 -0.202334759  0.41749257 -0.112231970  0.163487456 -0.8652643917
#> 29 -0.293466083 -0.50364082  0.191028721 -0.160450062  0.4339300021
#> 30  0.027017608 -0.45579508  0.139502601 -0.026786662 -0.0439486443
#> 31  0.096426928 -0.25383144  0.178562160  0.173603167 -0.3740600980
#> 32  0.182233731  0.20282454 -0.073836639 -0.036557915 -0.0881892968
#> 33 -0.211341439 -0.45227588  0.282875109  0.056399557 -0.0946639414
#> 34 -0.343027312 -0.21753097  0.138136866 -0.820907854 -0.0307291093
#> 35  0.710638520 -0.01464496  0.168206463 -0.311358774 -0.1925090308
#> 36  0.607671716  0.12476693 -0.336473038 -0.052898736 -0.3085541369
#> 37  0.183728236  0.26861569 -0.431997559  0.206015368  0.2818317405
#> 38  0.540867105 -0.45753732 -0.316387278  0.148278446 -0.0119914079
#> 39 -0.567193407  0.09451045  0.153776394  0.084398992 -0.1467812286
#> 40 -0.232806002 -0.02751240  0.505911399 -0.163030444  0.1593958422
#> 41  0.294120446 -0.09700271  0.198589059 -0.238318157 -0.4768991726
#> 42  0.117085614 -0.10984352  0.128116640  0.039947672  0.0380973723
#> 43 -0.658080976 -0.68108324  0.331743089 -0.335233808 -0.2056367252
#> 44 -0.501248501  0.03730972 -0.249324619 -0.266433487 -0.2402423420
#> 45 -0.023792469  0.12012691  0.049548011  0.317057540  0.0147022594
#> 46  0.172426651 -0.23978601 -0.086760521 -0.075925861 -0.2246793714
#> 47  0.636574552 -0.04417139  0.155184840 -0.243601641 -0.3366703027
#> 48 -0.153438101  0.03491255 -0.572679025 -0.187803312  0.1987817617
#> 49  0.032691106 -0.16082081 -0.191919916  0.027375826 -0.1690645212
#> 50 -0.456706747  0.32369503  0.679805729  0.092997427 -0.1699056489
#> 51  0.076783408  0.46414700 -0.736258904  0.407458350  0.4941938144
#> 52 -0.076847797 -0.43838148 -0.115119664  0.540556494 -0.0002115297
#> 53  0.020882779  0.09023787 -0.332105986 -0.381969628  0.4041285759
#> 54  0.098914825 -0.03150902  0.093376430 -0.071053617  0.4491113727
#> 55  0.188041683  0.33552689 -0.136612150 -0.134531513  0.2938047963
#> 56  0.005776522  0.01620024 -0.203213334  0.086237805  0.7385814386
#> 57 -0.035948362  0.08495578  0.221337328 -0.576362582  0.1249711388
#> 58 -0.195764564 -0.25463815  0.359771222  0.138794121  0.1939239678
#> 59  0.008214208  0.51865670 -0.125681607  0.035912320 -0.1717747111
#> 60  0.046288225  0.19378766  0.357663504 -0.006194035 -0.4274962568
#> 61  0.265161826  0.35540750  0.179566705  0.222170376  0.0816688744
#> 62  0.118963696 -0.37905263 -0.074571480 -0.041347397  0.4429358599
#> 63 -0.120295196 -0.41836440  0.326688660 -0.172867292  0.3007280018
#> 64 -0.136181970 -0.01388069  0.189397592  0.131203912  0.4290329671
#> 65 -0.254675293 -0.41137013  0.043952360 -0.105756212  0.8415768119
#> 66  0.509970365 -0.17758091  0.791789449 -0.447770135  0.0298719727
#> 67 -0.479000170 -0.09597171  0.085277412 -0.184673405  0.2267226089
#>              V6           V7           V8           V9        Mean.A
#> 1  -0.022218145 -0.952168819 -0.006434895  0.786123037 -0.0390480718
#> 2  -0.057250695 -0.127993666 -0.705523573  0.186529208  0.3538352008
#> 3   0.090406376 -0.264593250  0.301443619  0.085143931 -0.2214030875
#> 4   0.152790701 -0.097266922 -0.405617823  0.336421954 -0.1497177733
#> 5   0.149821408  0.549432395 -0.651408954 -0.180860416  0.0682003642
#> 6   0.125933863  0.132330483  0.441999858 -0.183791285 -0.1754200494
#> 7  -0.088976681  0.184175113 -0.475521964 -0.078570789 -0.1288949922
#> 8  -0.304193954 -0.036050623 -0.096846519  0.124031854  0.1782338498
#> 9  -0.266308757 -0.022439352  0.270301703 -0.226255027  0.1483761178
#> 10  0.310472955 -0.019923603  0.419107309 -0.333699812 -0.1616034687
#> 11 -0.212206198  0.796588537  0.248005844  0.150976973 -0.2151955672
#> 12 -0.009451519 -0.030997605 -0.286792603 -0.283503424  0.0179910811
#> 13  0.017791331  0.292569706  0.523493997  0.108821507 -0.0225969772
#> 14 -0.096105903  0.019851981  0.213622995  0.317649297  0.3570326759
#> 15  0.326242743  0.218322495  0.368990885  0.263791216 -0.2193498789
#> 16  0.122533819 -0.197281918 -0.124262179 -0.075882137 -0.1160720678
#> 17 -0.415195206 -0.146162974 -0.141130216 -0.289014161 -0.1640095499
#> 18  0.285432690 -0.120522803  0.156013898 -0.147761088 -0.1617718893
#> 19 -0.146110581  0.581207469  0.596168747  0.362352156  0.1738874658
#> 20 -0.020884991 -0.043676180  0.124316515  0.157003410 -0.0411418193
#> 21 -0.091762421  0.269125889 -0.104110235  0.107553288 -0.2454928417
#> 22  0.641000165 -0.066816400 -0.751398781 -0.807162110 -0.2476205270
#> 23 -0.065402558 -0.002134151 -0.676491663 -0.510257656  0.3273850110
#> 24  0.566516302  0.167238776  0.053114689 -0.442013729  0.0005151685
#> 25 -0.413646429  0.368637759 -0.236417242 -0.341555723  0.0739888276
#> 26 -0.294693970  0.492272614  0.217470242 -0.072056121  0.1542149182
#> 27 -0.163617068 -0.171586860  0.185344008  0.121516739  0.2858915078
#> 28 -0.202038007 -0.358437706  0.442516287 -0.200651742  0.0343086133
#> 29  0.636237784 -0.197471279 -0.183247897  0.657779967 -0.2020260596
#> 30  0.442828121  0.044771564  0.234786238 -0.637965578 -0.0964249584
#> 31 -0.430697021  0.001775245 -0.295468750  0.243815479  0.0070525494
#> 32 -0.186425106 -0.106173966  0.466549257  0.120695411  0.1037405452
#> 33 -0.258663189 -0.063333446 -0.082863787 -0.077771790 -0.1269140696
#> 34 -0.020187836  0.041075200  0.502657209 -0.663754038 -0.1408071380
#> 35  0.361207254 -0.500328032  0.107139730 -0.143920579  0.2880666729
#> 36 -0.258379532  0.205682265 -0.052517555 -0.230456109  0.1319885373
#> 37  0.076302868  0.405275269  0.060683582  0.354357520  0.0067821225
#> 38 -0.157205978  0.127662534 -0.123204652  0.298525014 -0.0776858318
#> 39  0.185407863  0.057068316 -0.149860009  0.080485960 -0.1063021876
#> 40  0.347238192  0.334197082 -0.126129444 -0.041533293  0.0818643316
#> 41 -0.003828723  0.219174222 -0.632310720  0.122990956  0.1319022661
#> 42 -0.028089913 -0.079689466 -0.165077413  0.077299290  0.0451195788
#> 43 -0.077463646 -0.324537345 -0.151446308 -0.134750723 -0.3358070434
#> 44 -0.086440006  0.068465627  0.507208144  0.588940127 -0.2377544659
#> 45 -0.202401458 -0.015066987 -0.023374199 -0.792837014  0.0486274851
#> 46  0.779105713  0.341368573 -0.188905221 -0.163959222 -0.0513732937
#> 47  0.101724147  0.385237668 -0.102956436  0.425267839  0.2491960011
#> 48  0.322984456 -0.052546314  0.293977643  0.332494094 -0.2304015267
#> 49 -0.252204296  0.307898838 -0.367221811 -0.063186595 -0.1066832081
#> 50 -0.059232278 -0.214232351  0.415932853  0.199081988  0.1822646711
#> 51  0.393890373 -0.424966221  0.274054775  0.438756527 -0.0651094975
#> 52  0.124954785  0.085302205  0.198145530  0.376881382 -0.2101163123
#> 53 -0.320772051  0.345563369  0.370905133 -0.261168688 -0.0736617801
#> 54 -0.407181279 -0.282567671 -0.150053713 -0.259983595  0.0535940784
#> 55  0.287669094 -0.108747561 -0.065979657  0.125465356  0.1289854731
#> 56 -0.451656793  0.297831633  0.711125841 -0.046663991 -0.0604121900
#> 57 -0.424781548  0.169359054  0.034940591  0.005084506  0.0901149149
#> 58 -0.071322799  0.237893632  0.211496559  0.068787348 -0.0302104957
#> 59 -0.002792604 -0.085340045 -0.307727837 -0.309862869  0.1337297659
#> 60 -0.175188243 -0.250194592  0.069599082 -0.138965011  0.1992464622
#> 61  0.027131354 -0.151795112  0.131901892  0.129879274  0.2667120100
#> 62  0.126096811 -0.418121893 -0.025077948 -0.741718135 -0.1115534721
#> 63  0.082436445  0.135595191 -0.003475200 -0.161024257 -0.0706569796
#> 64  0.114256270 -0.006747011  0.428318263 -0.613852270  0.0131116444
#> 65 -0.143925350 -0.274141418  0.192075214  0.010724964 -0.2073643541
#> 66  0.321657994 -0.311544357  0.473697589  0.607295992  0.3747263020
#> 67 -0.271753336  0.483969884  0.555894286 -0.926795515 -0.1632314884
#>          Mean.B       Mean.C     FC.crude  logFC.crude
#> 1   0.286678123 -0.057493559 -17.65885551  0.433642545
#> 2  -0.100131841 -0.215662677   0.79713562  0.342967684
#> 3   0.021322387  0.040664767   0.92619324 -0.022281846
#> 4   0.208418382 -0.055487597   3.80065252  0.311711873
#> 5  -0.030986824 -0.094278992   0.61046025  0.105102594
#> 6   0.140047682  0.130179686   1.03229059  0.010863190
#> 7  -0.015609020 -0.123305880  20.26904559  0.146783293
#> 8  -0.231617838 -0.002955096   2.26201265 -0.472453011
#> 9  -0.207730961  0.007202441   2.52247506 -0.415545622
#> 10 -0.073005684  0.021827965   0.48300219 -0.120505115
#> 11 -0.162061229  0.398523785   0.08657758 -0.615700217
#> 12 -0.002820357 -0.200431211   0.09528074  0.325197385
#> 13 -0.330541312  0.308295070  -0.93064894 -0.943433563
#> 14 -0.065751077  0.183708091   2.43926015 -0.518209075
#> 15 -0.081053387  0.283701532   0.27491523 -0.401017971
#> 16 -0.254154591 -0.132475411   8.41794999 -0.190517084
#> 17 -0.243042489 -0.192102450   2.81327092 -0.077668862
#> 18  0.083307527 -0.037423331   1.97090678  0.147148413
#> 19 -0.056865779  0.513242791  -0.67997532 -0.800020393
#> 20  0.128964031  0.079214582   1.41335109  0.062681299
#> 21  0.069669509  0.090856314   0.93700949 -0.023056094
#> 22 -0.116016174 -0.541792430  -0.44737227  0.680980866
#> 23 -0.029020242 -0.396294490   0.49249046  1.219809959
#> 24  0.132998351 -0.073886755  -1.78064192  0.291031854
#> 25 -0.047605310 -0.069778402   0.84577089  0.036884630
#> 26  0.007911781  0.212562245  -2.50745226 -0.310017383
#> 27  0.074934107  0.045091296   0.87606817  0.055623602
#> 28 -0.301271648 -0.038857720   4.58653925 -0.480215515
#> 29  0.303239241  0.092353597   1.71637302  0.217756944
#> 30  0.124030939 -0.119469259  -9.56661276  0.321055113
#> 31 -0.210384651 -0.016626009   9.18287335 -0.319149684
#> 32 -0.103724106  0.160356901  -3.66439432 -0.414904458
#> 33 -0.098975858 -0.074656341   0.53462354 -0.033734557
#> 34 -0.290608267 -0.040007210  -1.48612336 -0.372680032
#> 35 -0.047553517 -0.179036293   0.71851436  0.318151430
#> 36 -0.206610802 -0.025763800   2.14639824 -0.348720033
#> 37  0.188049992  0.273438790   0.67978000 -0.100689376
#> 38 -0.006972980  0.100994298   0.39575106 -0.138600629
#> 39  0.041008542 -0.004101911   1.44139267  0.057869779
#> 40  0.114534530  0.055511448  -1.23972007  0.084908688
#> 41 -0.239682018 -0.096715181   1.62535401 -0.295727254
#> 42  0.016651711 -0.055822530   0.28202173  0.111847695
#> 43 -0.206111393 -0.203578125   0.98084180 -0.003231527
#> 44 -0.197705278  0.388204632   0.06398052 -0.644639206
#> 45  0.043119447 -0.277092733   0.01691033  0.560612372
#> 46  0.159500160 -0.003831957   4.43558107  0.209040924
#> 47 -0.159515932  0.235849690  30.62358911 -0.738682633
#> 48  0.111320969  0.191308474   0.81032580 -0.083540902
#> 49 -0.131297664 -0.040836523  -0.37381465 -0.127955435
#> 50 -0.045380167  0.133594163   4.67726446 -0.300680690
#> 51  0.431847512  0.095948361   3.08558065  0.366592924
#> 52  0.221766583  0.220109706   1.00385118  0.001670355
#> 53 -0.099537701  0.151766605  -0.11478555 -0.331108722
#> 54 -0.009707841 -0.230868326   0.22253176  0.388556539
#> 55  0.148980792 -0.016420621  -0.13751363  0.255251597
#> 56  0.124387483  0.320764494   0.48481369 -0.221254737
#> 57 -0.292057664  0.069794717  18.80752221 -0.665106392
#> 58  0.087131763  0.172725846   0.57822201 -0.106489126
#> 59 -0.046218331 -0.234310250   0.48893623  0.375882014
#> 60 -0.202959512 -0.106520174   1.31540177 -0.215772221
#> 61  0.110323535  0.036662018   0.67980213  0.131814028
#> 62  0.175895091 -0.394972659  -1.01421702  0.845313474
#> 63  0.070099052 -0.009634756   2.30663555  0.104535406
#> 64  0.224831050 -0.064093672  -2.74229048  0.392974059
#> 65  0.197298417 -0.023780413   2.20423839  0.247061785
#> 66 -0.032080056  0.256483075   3.44041994 -0.571877617
#> 67 -0.076568044  0.037689552   0.43133086 -0.144236112