Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeansBetweenABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeansBetweenABC(dat = dat, A.samples = 1:3, B.samples = 4:6,
 C.samples = 7:9)
#>              V1           V2          V3          V4            V5           V6
#> 1   0.425656939 -0.106230003 -0.12809464 -0.03052199  0.0683323904  0.528646290
#> 2   0.369532165  0.513292869 -0.32704605  0.20305850 -0.2152395045 -0.302790016
#> 3   0.254261324  0.296683972  0.38064693  0.05743322 -0.1571694174 -0.064034477
#> 4   0.217463775  0.232868125  0.13871095  0.18673156  0.5568991736 -0.143531766
#> 5   0.239953763 -0.599523313 -0.27380733 -0.05698395  0.4820039561  0.052405463
#> 6  -0.513674390 -0.014369759 -0.26382140  0.32793368  0.3768826609  0.149744796
#> 7  -0.411178365  0.154603004  0.17536628  0.21233147  0.6553824124 -0.111901465
#> 8  -0.225974921  0.001258371 -0.10605901  0.29408143  0.1354500117  0.303145598
#> 9   0.085821898  0.169936930  0.01105230  0.13325160 -0.3651078851  0.090181535
#> 10  0.530727734 -0.047399976 -0.19495954  0.16067165  0.2212784560 -0.041677669
#> 11 -0.021624418  0.376508534  0.16197994  0.21582320  0.2107656265  0.317018444
#> 12  0.138303510 -0.049826863 -0.37747866 -0.09434866  0.0685894542  0.285041690
#> 13  0.038146596 -0.596214805  0.34374734  0.11813952 -0.1776584172 -0.264267723
#> 14  0.191656581 -0.116019156 -0.30382763  0.35050175  0.3452048707  0.057149186
#> 15  0.074631989 -0.323421215 -0.16246259  0.12886446 -0.0555217945 -0.135967342
#> 16 -0.413843006 -0.196594905  0.04432107  0.13046602  0.1939559463 -0.558921731
#> 17  0.270228983  0.505725957  0.33738589  0.07144265  0.3868733949 -0.012127753
#> 18  0.199550440  0.404582209 -0.31732984  0.30060736 -0.0197655036  0.059324326
#> 19 -0.188868996 -0.121298351  0.16704209 -0.01348544  0.0569805969  0.007665886
#> 20  0.066786974  0.071912774  0.18078743  0.01437108 -0.2807361628  0.157160661
#> 21 -0.497312084 -0.099101734 -0.18247218 -0.33717179 -0.0169265256 -0.112375051
#> 22  0.347465206  0.325858397 -0.43375441 -0.68225492  0.7351195852  0.416912990
#> 23  0.487968532 -0.090647744  0.08648806  0.20835552  0.1241007338  0.228585242
#> 24 -0.004774962 -0.266585897  0.75592656  0.12302656  0.3735993190 -0.333846411
#> 25  0.059647837 -0.093060223 -0.02916433 -0.35425150 -0.3525830371 -0.327652947
#> 26  0.130082379 -0.231240478 -0.25592514  0.10929772 -0.2513136902  0.618669173
#> 27  0.428598015 -0.342541360 -0.01030156 -0.55535510 -0.0932059243 -0.098613325
#> 28 -0.085277593  0.136113896  0.52029573 -0.02455286  0.0727714284  0.162922606
#> 29  0.098861296 -0.117125149  0.03912920  0.17684555  0.2083284758 -0.563932438
#> 30 -0.236883093  0.207387247  0.28559292  0.12547626 -0.0483337287  0.318548435
#> 31  0.578936368 -0.367636129  0.02386607  0.08233889  0.6834536679  0.645256278
#> 32  0.261379154 -0.068590163 -0.28768631  0.24467006  0.0073109370 -0.531154862
#> 33 -0.328247257 -0.211284045 -0.20863276  0.27067663 -0.3645617379 -0.493092180
#> 34 -0.050917077 -0.058598497  0.29540172  0.45410938  0.0002602562  0.103891961
#> 35  0.090009906 -0.048522970 -0.37147473 -0.19470342 -0.5068470824 -0.052013705
#> 36  0.423661401  0.327973621 -0.07563329 -0.06838575 -0.2018700236  0.160720621
#> 37  0.096710152  0.009115318 -0.56318263  0.19570915 -0.2719892883 -0.285325543
#> 38 -0.438313753  0.067360165 -0.12284919 -0.72842439 -0.2082791299 -0.300550515
#> 39  0.264041426 -0.042135030  0.88572515  0.13078986 -0.3763306629 -0.343364591
#> 40  0.327004237  0.130866815  0.08557048  0.06999439 -0.0173580228 -0.341932081
#> 41 -0.864763896  0.234803749 -0.24382448 -0.31187362  0.6854391209  0.405278334
#> 42 -0.078000489 -0.183980678 -0.49807473  0.06471602  0.3707305541 -0.844180649
#> 43  0.012207219  0.031221938 -0.20647581 -0.27321688 -0.2492235742 -0.203864294
#> 44  0.277137410  0.219791491 -0.48555325  0.01857352  0.0584094005  0.143353996
#> 45 -0.231703835  0.017495227  0.06892084  0.48137618  0.1753651608 -0.613650000
#> 46  0.505912891  0.237292155  0.08075147  0.51755738  0.0985583559 -0.070249449
#> 47 -0.119102910 -0.050334926 -0.61621563 -0.02188855 -0.2404203990 -0.014070090
#> 48  0.551782824 -0.145284802  0.47865989  0.34860452 -0.1291589619 -0.448190776
#> 49  0.050976323 -0.236231754  0.17071615  0.06153337 -0.2705645772 -0.607742829
#> 50  0.186437378 -0.208643531  0.34052967  0.17197000 -0.2007042330  0.184339768
#> 51  0.003481078 -0.076550302 -0.53055819 -0.16991701 -0.0574260770 -0.269056209
#> 52  0.043765960 -0.219062857 -0.32998683  0.04715034 -0.4274986392  0.276252650
#> 53  0.248611679 -0.344463860 -0.51601634  0.04209198  0.4084043328 -0.174821660
#> 54  0.235208176  0.075909887 -0.09797556 -0.43647600  0.1548361491  0.816198460
#> 55 -0.205376210  0.149521448  0.07591149 -0.03142510 -0.1194508103 -0.376405883
#> 56 -0.029979355 -0.110345826  0.14559339 -0.02965113  0.0032147749  0.082564410
#> 57 -0.149674588 -0.135778749 -0.04090927 -0.45467440 -0.2367248804  0.169849809
#> 58 -0.553532395 -0.353463977 -0.36467235 -0.02176230  0.2739093199  0.236379718
#> 59 -0.239888703 -0.364469441 -0.64353475  0.40889445  0.2866251320  0.242199826
#> 60  0.098191606 -0.264196039  0.41477494 -0.21747399  0.1582929532 -0.202770315
#> 61 -0.125213547 -0.333564209  0.02533154 -0.42898493  0.2378270573  0.414130835
#> 62  0.251900672 -0.618050141 -0.06839500 -0.07334780 -0.0199440581 -0.180214079
#> 63  0.206690928 -0.036693428 -0.01101906  0.22260205  0.1638264543 -0.094473738
#> 64 -0.366240254  0.021091683 -0.00771866  0.13251195 -0.7681424270  0.560536908
#> 65 -0.164005432 -0.116719854 -0.07673142 -0.15603923 -0.2391354231  0.139544547
#> 66 -0.047893254 -0.240130369 -0.59805347 -0.07830540 -0.1460821069 -0.144416456
#> 67  0.215373021  0.065737995  0.39365628  0.42591717 -0.0955302487  0.099697275
#>              V7            V8           V9       Mean.A      Mean.B
#> 1  -0.348952280  0.2036919954  0.040933299  0.063777433  0.18881890
#> 2   0.689574356  0.2761570891 -0.078506684  0.185259661 -0.10499034
#> 3   0.246681062  0.2793553580 -0.017288926  0.310530741 -0.05459023
#> 4   0.354472622 -0.5052197329  0.094210035  0.196347618  0.20003299
#> 5  -0.300088622 -0.0860050631 -0.028881606 -0.211125626  0.15914182
#> 6  -0.103445307  0.3227345064 -0.083986384 -0.263955184  0.28485371
#> 7   0.484479934 -0.3486673808  0.247252170 -0.027069692  0.25193747
#> 8   0.037590282 -0.3491246456  0.134004954 -0.110258521  0.24422568
#> 9  -0.001363898  0.6261247162 -0.290783023  0.088937041 -0.04722492
#> 10  0.105226204  0.1390986095  0.157318945  0.096122740  0.11342415
#> 11  0.174201411 -0.3572742984  0.003024238  0.172288018  0.24786909
#> 12  0.016477116  0.0077433822  0.258112088 -0.096334006  0.08642749
#> 13  0.330836414 -0.2653583327  0.141072456 -0.071440290 -0.10792887
#> 14  0.178454071  0.0606121827  0.142916946 -0.076063402  0.25095194
#> 15 -0.306306354  0.0598087710 -0.278904572 -0.137083940 -0.02087489
#> 16  0.093040667 -0.1162633655 -0.244341694 -0.188705614 -0.07816659
#> 17 -0.086210507  0.1017309110  0.182143310  0.371113609  0.14872943
#> 18  0.086374628  0.1353376422 -0.128196184  0.095600937  0.11338873
#> 19 -0.030237015  0.0077360253 -0.163038629 -0.047708420  0.01705368
#> 20 -0.026364085 -0.2568826479  0.567028979  0.106495726 -0.03640147
#> 21  0.194803883  0.1046814733 -0.154820097 -0.259628665 -0.15549112
#> 22 -0.252367772  0.0768885486 -0.407018757  0.079856397  0.15659255
#> 23  0.275703241 -0.3552115831  0.059311325  0.161269618  0.18701383
#> 24  0.497717174  0.4854075158  0.150381821  0.161521900  0.05425982
#> 25 -0.043148069 -0.2143319104 -0.053957817 -0.020858905 -0.34482916
#> 26  0.249674106  0.0266430902  0.567154316 -0.119027746  0.15888440
#> 27  0.536584372 -0.2286418705  0.213472850  0.025251698 -0.24905812
#> 28 -0.223662675 -0.3467794075  0.140528605  0.190377344  0.07038039
#> 29 -0.094249073  0.0840915751  0.405307942  0.006955116 -0.05958614
#> 30  0.251534690  0.3129573344  0.106699929  0.085365692  0.13189699
#> 31 -0.084172418 -0.1622413295 -0.044754066  0.078388771  0.47034961
#> 32  0.516946828  0.1372660580  0.237598380 -0.031632440 -0.09305795
#> 33 -0.247865346 -0.1416954181 -0.019416264 -0.249388020 -0.19565909
#> 34  0.018160900  0.0556548654 -0.498579200  0.061962047  0.18608720
#> 35  0.107633406  0.2729212153 -0.131499271 -0.109995932 -0.25118807
#> 36  0.342630604  0.1070909081 -0.045316709  0.225333910 -0.03651172
#> 37 -0.248998486  0.1298752595  0.127486944 -0.152452386 -0.12053523
#> 38  0.019183192 -0.1630321437 -0.399285284 -0.164600926 -0.41241801
#> 39 -0.009205969 -0.0410759402  0.100935052  0.369210517 -0.19630180
#> 40  0.245907467  0.1370201340 -0.207284405  0.181147178 -0.09643191
#> 41  0.180348488 -0.0695795886  0.496350766 -0.291261541  0.25961461
#> 42  0.046940433 -0.3663607768  0.024615010 -0.253351968 -0.13624469
#> 43 -0.229694709 -0.2277714996 -0.575450690 -0.054348883 -0.24210158
#> 44 -0.583218537  0.1240719571  0.340705340  0.003791884  0.07344564
#> 45  0.526770571  0.0035283134  0.161957519 -0.048429257  0.01436378
#> 46 -0.429266564  0.3868932028 -0.317521969  0.274652173  0.18195543
#> 47 -0.003534793 -0.4835202494  0.013853871 -0.261884489 -0.09212635
#> 48  0.185268086  0.1073333708  0.910975475  0.295052637 -0.07624841
#> 49  0.084113515 -0.1744792740 -0.622830207 -0.004846427 -0.27225801
#> 50  0.410882083  0.4476763867 -0.029415984  0.106107838  0.05186851
#> 51  0.049398959 -0.2551018726  0.063734119 -0.201209138 -0.16546643
#> 52 -0.352716397  0.0705866703 -0.161234240 -0.168427908 -0.03469855
#> 53 -0.020070016 -0.2093975745  0.006994086 -0.203956172  0.09189155
#> 54 -0.022424585 -0.1255465508 -0.008684020  0.071047502  0.17818620
#> 55 -0.059631932  0.6427908350  0.284512161  0.006685576 -0.17576060
#> 56 -0.357724036  0.6538321561  0.468120272  0.001756069  0.01870935
#> 57 -0.345215766 -0.0506924276 -0.009484065 -0.108787534 -0.17384982
#> 58  0.504396146  0.1653708620  0.253159406 -0.423889573  0.16284225
#> 59  0.111226006 -0.0046907120 -0.656676084 -0.415964297  0.31257314
#> 60  0.167447508  0.0073251832 -0.214413850  0.082923503 -0.08731712
#> 61 -0.285923207 -0.0008338179  0.063694028 -0.144482072  0.07432432
#> 62 -0.066626657  0.0339517662 -0.658488245 -0.144848155 -0.09116865
#> 63 -0.109359845 -0.2345701346 -0.786110128  0.052992813  0.09731825
#> 64 -0.004470270  0.2595422533  0.182665882 -0.117622410 -0.02503119
#> 65  0.331675701  0.0737777953  0.425656939 -0.119152237 -0.08521003
#> 66  0.262044051 -0.0366983473  0.369532165 -0.295359031 -0.12293466
#> 67  0.059770117  0.3071680351  0.254261324  0.224922431  0.14336140
#>          Mean.C    FC.crude  logFC.crude
#> 1  -0.034775662 -1.26877256  0.319663748
#> 2   0.295741587 -2.62712656 -0.645803064
#> 3   0.169582498  2.59046129 -0.436263322
#> 4  -0.018845692 -0.01712586  0.354897847
#> 5  -0.138325097  5.08605440  0.353075652
#> 6   0.045100938  1.77575804  0.242632193
#> 7   0.127688241  1.80286178  0.147433887
#> 8  -0.059176470  6.93950606  0.365868271
#> 9   0.111325932 -6.08167517 -0.243111306
#> 10  0.133881253  0.45821210 -0.028723581
#> 11 -0.060016216 -0.32535383  0.486509677
#> 12  0.094110862  0.95965568 -0.009341619
#> 13  0.068850179 -0.26009310 -0.243027706
#> 14  0.127327733  1.60781510  0.141079412
#> 15 -0.175134052 -3.05410532  0.214573600
#> 16 -0.089188131  1.11074981  0.014389548
#> 17  0.065887905  0.72858929  0.162513190
#> 18  0.031172029 -0.27608399  0.121517600
#> 19 -0.061846539 -4.58067289  0.111073684
#> 20  0.094594082 12.00650907 -0.205186406
#> 21  0.048221753  0.33827319 -0.244277647
#> 22 -0.194165993 -0.28003607  0.568667812
#> 23 -0.006732339 -0.15323759  0.302018977
#> 24  0.377835503 -0.49586376 -0.446206631
#> 25 -0.103812599  3.90543494 -0.439907860
#> 26  0.281157171  0.69445932 -0.131828713
#> 27  0.173805117 -1.84653990 -0.662392295
#> 28 -0.143304492  0.35961487  0.401297300
#> 29  0.131716815 -0.53334681 -0.268961203
#> 30  0.223730651  0.33629393 -0.121347704
#> 31 -0.097055938 -2.23409895  0.755430481
#> 32  0.297270422 -0.18675883 -0.501692504
#> 33 -0.136325676  0.47521503 -0.079030625
#> 34 -0.141587812 -0.60980220  0.497146709
#> 35  0.083018450 -0.73151096 -0.474204131
#> 36  0.134801601  2.89228928 -0.301099822
#> 37  0.002787906  0.20559843 -0.162865807
#> 38 -0.181044745 15.07053081 -0.386923926
#> 39  0.016884381  1.60508193 -0.575952236
#> 40  0.058547732  2.26411369 -0.280395954
#> 41  0.202373222  1.11595899  0.054256073
#> 42 -0.098268445  0.75512392 -0.048229439
#> 43 -0.344305633  0.64751967  0.194012140
#> 44 -0.039480413 -1.60966158  0.160963597
#> 45  0.230752134  0.22491841 -0.267360177
#> 46 -0.119965110  0.23490290  0.583737349
#> 47 -0.157733724  1.62992699  0.083273095
#> 48  0.401192311 -3.49823048 -0.815092294
#> 49 -0.237731989  1.14825317 -0.066438927
#> 50  0.276380828 -0.31854333 -0.307298048
#> 51 -0.047322932  0.23226712 -0.155835294
#> 52 -0.147787989  6.47916110  0.151602312
#> 53 -0.074157835  2.27928746  0.197830911
#> 54 -0.052218385 -0.86916748  0.336624685
#> 55  0.289223688 -0.64574005 -0.649616075
#> 56  0.254742797  0.06701254 -0.301117729
#> 57 -0.135130753  2.46979272 -0.058543055
#> 58  0.307642138  0.80205931 -0.125990497
#> 59 -0.183380263  3.13236219  0.487865839
#> 60 -0.013213720  1.77080859 -0.123410729
#> 61 -0.074354332  3.12011184  0.187685951
#> 62 -0.230387712 -0.62754019  0.204443450
#> 63 -0.376680036 -0.10316091  0.872709745
#> 64  0.145912622  0.35134312 -0.209711927
#> 65  0.277036812  0.08567173 -0.433338759
#> 66  0.198292623  0.34928350 -0.349348857
#> 67  0.207066492  4.56772561 -0.096750800