Log of Means between three samples

Computes the mean of each sample, fold change, and log fold change.

logofMeansBetweenABC(dat, A.samples, B.samples, C.samples)

Arguments

dat

a dataframe. Differential analysis results from "DESeqCalculation"

A.samples

First sample type

B.samples

Second sample type

C.samples

Third sample type

Value

log2 fold change of means between three given samples

Examples

dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
#> Warning: data length [600] is not a sub-multiple or multiple of the number of rows [67]
dat <- as.data.frame(dat)
logofMeansBetweenABC(dat = dat, A.samples = 1:3, B.samples = 4:6,
 C.samples = 7:9)
#>              V1          V2            V3           V4            V5
#> 1  -0.019838880 -0.60111648  1.269824e-01  0.741205374 -0.0300744864
#> 2   0.025009573  0.80682650  5.032353e-01  0.250907244 -0.4442829900
#> 3   0.215204363  0.31023165 -4.593745e-01  0.242019641 -0.0282020528
#> 4  -0.616612555 -0.35722125 -3.495989e-01 -0.521392611 -0.2213954028
#> 5   0.279910986  0.49499458 -1.532440e-01 -0.016128642  0.0767297571
#> 6  -0.124659675 -0.38199346  3.280719e-02  0.053779824 -0.1791115084
#> 7   0.026111436  0.13480356  1.367654e-02  0.004430873 -0.4676333470
#> 8   0.079480846 -0.29305673 -6.023707e-02  0.697055066 -0.1478594385
#> 9   0.080245341  0.06581809  7.201148e-02 -0.095874675 -0.0282531494
#> 10  0.002711520  0.01369546 -5.094892e-02 -0.191147882  0.3704001779
#> 11  0.927179529 -0.10892392 -1.131965e-01  0.237061301 -0.7649311588
#> 12  0.376378902  0.01120415 -2.416048e-01  0.256890761 -0.4906802181
#> 13  0.388886292  0.44979766  2.193512e-02  0.309934763  0.5044203576
#> 14 -0.020368520 -0.05706348  4.800192e-01  0.278429935  0.0611983662
#> 15  0.063712633  0.37346193  4.506589e-01 -0.123678769 -0.1585843544
#> 16  0.275949713 -0.08556350  1.760118e-01 -0.011058273  0.5173692565
#> 17 -0.123148358  0.61745384  9.885434e-02 -0.724312003  0.0799085016
#> 18  0.438710776  0.24731716 -1.328335e-01  0.233760482 -0.2404813401
#> 19 -0.525069575  0.52365099  3.898894e-02  0.286512561 -0.0782428902
#> 20  0.249430575 -0.04556544  3.165216e-01 -0.511302206 -0.0145734439
#> 21 -0.180044761  0.01908531  5.453827e-03  0.255704805  0.2923027847
#> 22 -0.076980129 -0.10224267  5.412565e-02 -0.108436605 -0.2139666411
#> 23 -0.461170922  0.03190893  2.935591e-01 -0.228146055 -0.1558431579
#> 24  0.223912956 -0.14125930  3.241010e-01  0.129421382 -0.2567523418
#> 25  0.242863961 -0.36665819  1.928767e-01  0.399290217 -0.0006313113
#> 26 -0.235599505  0.05943189 -5.071424e-01  0.366557322 -0.2074111629
#> 27 -0.040767254  0.06378232 -2.213595e-01  0.202749400  0.4877936704
#> 28  0.169000015 -0.03730246 -4.120650e-01 -0.456093737 -0.0007400531
#> 29  0.179128929  0.08102289 -4.440996e-01  0.260850513  0.2046996071
#> 30 -0.119394946  0.16456518 -1.315019e-01 -0.090250419 -0.0668324869
#> 31 -0.256519873  0.06359322 -3.937369e-01  0.025295342 -0.2829740355
#> 32  0.227616852 -0.13732091  1.754347e-02  0.111746280  0.1037484749
#> 33  0.195667344  0.04085729 -4.883696e-01  0.135706683  0.0322499825
#> 34 -0.160435001  0.40347589 -4.496414e-01 -0.287618294 -0.1387933289
#> 35  0.604384092 -0.11205114  2.554086e-01 -0.534990297 -0.6030175642
#> 36  0.438497961  0.08924501 -4.431839e-02 -0.468031922 -0.2062662729
#> 37 -0.421798321  0.01321137 -2.421413e-01  0.747417117 -0.1918006594
#> 38  0.269439241  0.21360125  3.484546e-02 -0.388796779 -0.0316151146
#> 39  0.488646273 -0.04585343  3.767909e-01 -0.243399074  0.3177165510
#> 40  0.052372921 -0.12823796 -5.216027e-02  0.158688660 -0.1559291916
#> 41  0.157677431  0.20737150 -1.456668e-01  0.128759462  0.4375278496
#> 42  0.150896125  0.19561358  5.819344e-01  0.502881211 -0.2412834227
#> 43  0.274313208  0.43339180  7.764131e-02 -0.204934540 -0.3874925491
#> 44 -0.146483784 -0.04196473 -4.515374e-01 -0.258548929  0.1594244662
#> 45  0.025571340 -0.15212186  3.525812e-01  0.214205864 -0.0139014603
#> 46 -0.805638442 -0.33672599  1.803818e-01 -0.151287210 -0.3469771325
#> 47 -0.128783219  0.35517039  2.975564e-01 -0.272669729 -0.4025083788
#> 48  0.554224015 -0.08742204  6.254376e-01  0.328551384  0.3488426682
#> 49 -0.454740675  0.58849079 -3.132830e-01 -0.143989272  0.4689351303
#> 50 -0.480638304  0.32235008 -8.244164e-02  0.248158924  0.1302409726
#> 51  0.092026568 -0.21388978 -3.981899e-03 -0.211545825 -0.3675453307
#> 52  0.002294561  0.40169870  1.625022e-01 -0.021367599 -0.2993808433
#> 53 -0.065194441 -0.61898410  9.107501e-05  0.007492345 -0.2010056176
#> 54  0.052554778  0.53482666  4.834310e-01 -0.185738753  0.1208398362
#> 55 -0.104255814 -0.20953906 -3.167229e-02 -0.064806571 -0.2560903458
#> 56 -0.586852932  0.18168509 -2.966503e-01  0.203500375  0.4708018593
#> 57 -0.377210495  0.23337006  9.215899e-02 -0.485034785  0.2633184632
#> 58  0.205132443  0.01270755 -7.645298e-02  0.113222893  0.4245641203
#> 59 -0.462770020  0.13131880  6.440823e-02 -0.180366032  0.2210156506
#> 60  0.045783152 -0.50910096 -2.033132e-01  0.051277535  0.4723151793
#> 61  0.063155621 -0.22459112  1.039534e-01 -0.025576911 -0.4001736352
#> 62  0.033456461 -0.02619132  3.814048e-01  0.298250806  0.3250874613
#> 63 -0.303317852 -0.11447191 -1.160986e-01  0.032168045 -0.1065586009
#> 64  0.200004034  0.03066116  2.271633e-01 -0.026817082  0.1533194949
#> 65 -0.087099561  0.27478304 -1.558984e-01  0.375992868 -0.2395090543
#> 66 -0.034967406 -0.71165389 -1.234394e-01 -0.350528458 -0.3159751895
#> 67 -0.004842528 -0.48627959  4.081785e-01  0.148381266  0.1512058074
#>             V6          V7            V8          V9      Mean.A        Mean.B
#> 1   0.32038605 -0.14731277  5.240070e-02  0.12681831 -0.16465767  0.3438389793
#> 2  -0.30596540 -0.25966235 -3.528477e-01 -0.01901708  0.44502380 -0.1664470495
#> 3   0.06946311 -0.32543890  5.454341e-01 -0.24893095  0.02202050  0.0944269000
#> 4  -0.20653881  0.33018774 -5.557584e-02 -0.26203456 -0.44114423 -0.3164422761
#> 5   0.25946574  0.05843031  2.346827e-01 -0.49521757  0.20722052  0.1066889510
#> 6  -0.56257158  0.24202475  3.791208e-01  0.06144104 -0.15794865 -0.2293010869
#> 7   0.22229852  0.09820072  5.099946e-03  0.29729914  0.05819718 -0.0803013169
#> 8  -0.19502649  0.15300894  6.276618e-01 -0.20125683 -0.09127099  0.1180563795
#> 9  -0.43618362 -0.13179989 -2.077330e-01 -0.03967147  0.07269164 -0.1867704828
#> 10  0.31321645 -0.32651555 -1.102461e-01  0.20207524 -0.01151398  0.1641562487
#> 11 -0.26322384 -0.02355599  1.628176e-01 -0.05627671  0.23501969 -0.2636978982
#> 12  0.23955749 -0.27707614 -3.013152e-01  0.48557956  0.04865941  0.0019226761
#> 13  0.26527045 -0.15315348 -8.649358e-02  0.06305200  0.28687303  0.3598751909
#> 14  0.46808923 -0.04295326 -4.111372e-05 -0.07940181  0.13419573  0.2692391766
#> 15 -0.28817343 -0.11845359  3.532341e-02 -0.05207191  0.29594449 -0.1901455166
#> 16  0.33215860  0.01985209 -2.834652e-02  0.16835068  0.12213265  0.2794898611
#> 17  0.17410543 -0.17336852  2.368770e-01 -0.28843108  0.19771994 -0.1567660235
#> 18 -0.19510519 -0.39442194  6.643497e-02  0.47096566  0.18439815 -0.0672753496
#> 19 -0.62017849 -0.45711304 -4.992943e-02  0.22041826  0.01252345 -0.1373029383
#> 20 -0.09582159 -0.42258557 -3.327187e-02  0.02436265  0.17346224 -0.2072324149
#> 21 -0.06100847 -0.10115864  2.623697e-01  0.40262128 -0.05183521  0.1623330396
#> 22  1.07723159  0.06745663 -4.101189e-01 -0.07802177 -0.04169905  0.2516094487
#> 23 -0.05854455  0.02717260 -1.733259e-01  0.15057983 -0.04523429 -0.1475112537
#> 24  0.28023188 -0.22915435 -3.293863e-01 -0.07895183  0.13558487  0.0509669727
#> 25 -0.02743671 -0.16795208 -8.304704e-03  0.19924233  0.02302748  0.1237407317
#> 26  0.03102331 -0.04703070 -2.144337e-01  0.80930004 -0.22777002  0.0633898228
#> 27  0.20861117  0.14225072  3.198834e-01  0.09636319 -0.06611482  0.2997180789
#> 28  0.51377363 -0.24245275 -3.997433e-02  0.29321259 -0.09345582  0.0189799451
#> 29 -0.14839091 -0.56370801 -4.928510e-01 -0.12250728 -0.06131592  0.1057197371
#> 30  0.01618080 -0.08780986 -2.948816e-01 -0.03778210 -0.02877724 -0.0469673679
#> 31 -0.09102331 -0.45743259  1.181419e-01  0.19796755 -0.19555451 -0.1162340014
#> 32 -0.11878334  0.52282765  2.186076e-01  0.15672745  0.03594647  0.0322371391
#> 33  0.26933556  0.20687731 -6.831803e-02  0.07950821 -0.08394832  0.1457640753
#> 34  0.32916108  0.03914713  5.968842e-01  0.30606326 -0.06886682 -0.0324168474
#> 35 -0.18002901 -0.01731813 -3.718866e-02  0.13278249  0.24924719 -0.4393456239
#> 36  0.42720777 -0.57020242  5.653611e-02  0.63125355  0.16114153 -0.0823634757
#> 37 -0.67185688  0.35469234 -2.123092e-01  0.03798140 -0.21690942 -0.0387468063
#> 38 -0.07098298  0.29531667  3.232821e-01  0.44646197  0.17262865 -0.1637982917
#> 39 -0.02506718  0.23416149  1.828315e-01 -0.06080438  0.27319458  0.0164167669
#> 40  0.15348325 -0.30823865  1.698047e-01 -0.04955428 -0.04267510  0.0520809055
#> 41 -0.17393419 -0.07774500 -7.981164e-01  0.16902623  0.07312739  0.1307843726
#> 42 -0.35088946  0.01769548 -9.785767e-02  0.01597991  0.30948138 -0.0297638914
#> 43 -0.08658923  0.11069730 -2.033387e-02  0.26779741  0.26178211 -0.2263387732
#> 44  0.11189841 -0.10132400 -6.769697e-01  0.67564534 -0.21332864  0.0042579820
#> 45 -0.12704023  0.19024173  2.835032e-02  0.36011124  0.07534357  0.0244213929
#> 46 -0.08166755  0.67713180 -1.307341e-01 -0.11986850 -0.32066088 -0.1933106308
#> 47 -0.52949650  0.19395105  3.726871e-01 -0.06684548  0.17464785 -0.4015582021
#> 48  0.31471796 -0.01647094  2.591974e-03 -0.57645130  0.36407986  0.3307040034
#> 49 -0.32764446 -0.52693656 -4.699834e-01 -0.09026403 -0.05984431 -0.0008995323
#> 50 -0.39728256 -0.23440017 -1.810026e-02  0.05991146 -0.08024329 -0.0062942207
#> 51  0.06516365 -0.02724968 -6.011076e-01 -0.38352576 -0.04194837 -0.1713091692
#> 52  0.14956353  0.37142947  7.872681e-01 -0.33182416  0.18883182 -0.0570616376
#> 53  0.66055628  0.06934396 -8.575866e-02 -0.18242352 -0.22802915  0.1556810010
#> 54  0.10086251 -0.65320218  3.081956e-01 -0.12781567  0.35693749  0.0119878640
#> 55  0.10048206  0.23121338  2.007208e-01 -0.39865197 -0.11515572 -0.0734716186
#> 56 -0.27883387 -0.43846996 -1.852544e-01  0.12452888 -0.23393937  0.1318227873
#> 57  0.53955280  0.06754144 -1.102402e-01  0.03561937 -0.01722715  0.1059454931
#> 58  0.19764629 -0.30455395 -3.138136e-01  0.60370087  0.04712900  0.2451444354
#> 59  0.20310367  0.10755843  1.093753e-01  0.33664125 -0.08901433  0.0812510954
#> 60 -0.04638737  0.22036927 -9.407603e-02  0.51568532 -0.22221033  0.1590684491
#> 61 -0.07473019 -0.43715378 -1.537893e-01  0.11541233 -0.01916070 -0.1668269131
#> 62  0.48066097 -0.41414210  7.457439e-01  0.35980453  0.12955666  0.3679997444
#> 63  0.24866461  0.08671280 -1.308508e-01 -0.02912449 -0.17796277  0.0580913498
#> 64  0.03757874 -0.10340470  4.990053e-01  0.24675340  0.15260951  0.0546937175
#> 65  0.25126553 -0.31152347 -8.305763e-03 -0.01983888  0.01059503  0.1292497798
#> 66 -0.04215662 -0.05346711 -2.548692e-01  0.02500957 -0.29002025 -0.2362200886
#> 67 -0.17377269  0.02430435 -2.175413e-01  0.21520436 -0.02764786  0.0419381291
#>          Mean.C    FC.crude logFC.crude
#> 1   0.010635412  2.90083696  0.36009093
#> 2  -0.210509050  0.93278445  0.17365644
#> 3  -0.009645239 -2.28658464  0.14727469
#> 4   0.004192445  0.28001725 -0.36186285
#> 5  -0.067368200  0.36611689  0.31027347
#> 6   0.227528868 -0.18510143 -0.57718022
#> 7   0.133533270 -1.83840822 -0.31986242
#> 8   0.193137961  0.73600837 -0.08689978
#> 9  -0.126401448  1.30322015 -0.11306104
#> 10 -0.078228811 -2.63315112  0.33309357
#> 11  0.027661625  2.40510339 -0.66104568
#> 12 -0.030937246  0.58716955  0.05060848
#> 13 -0.058865019 -0.21114878  0.71371270
#> 14 -0.040798729 -0.77170124  0.46027181
#> 15 -0.045067365  1.42543433 -0.35873681
#> 16  0.053285416 -2.28559945  0.31374443
#> 17 -0.074974211  1.29993974 -0.17211371
#> 18  0.047659566  1.84054485 -0.20612964
#> 19 -0.095541403  1.38644882 -0.06918133
#> 20 -0.143831596  1.19981737 -0.14061388
#> 21  0.187944125  0.89318894 -0.03011501
#> 22 -0.140228001 -2.97687629  0.52071338
#> 23  0.001475506 -2.18962577 -0.22151918
#> 24 -0.212497511  0.24309733  0.48968442
#> 25  0.007661847 -6.55444756  0.16077868
#> 26  0.182611891  0.70948508 -0.12741826
#> 27  0.186165776  1.45010320  0.12522313
#> 28  0.003595170  1.15852261  0.02009145
#> 29 -0.393022082 -0.50356511  0.80429417
#> 30 -0.140157870  0.16331501  0.14387808
#> 31 -0.047107716  0.53433629 -0.08956068
#> 32  0.299387554 -0.01408031 -0.34271977
#> 33  0.072689164  1.46652252  0.08838418
#> 34  0.314031541  0.09519491 -0.41604464
#> 35  0.026091902  3.08571135 -1.31882402
#> 36  0.039195745  1.99683006 -0.21497948
#> 37  0.060121506  0.64311454 -0.11625478
#> 38  0.355020264 -1.84453075 -0.83338072
#> 39  0.118729525  1.66236833 -0.18607081
#> 40 -0.062662729 -4.74073280  0.15973750
#> 41 -0.235611714 -0.18674984  0.61356959
#> 42 -0.021394096  1.02529591 -0.01815988
#> 43  0.119386949  3.42793169 -0.74450989
#> 44 -0.034216129  1.21480415  0.04632303
#> 45  0.192901097 -0.43316810 -0.23575080
#> 46  0.142176404  0.27515124 -0.37583350
#> 47  0.166597550 71.57572548 -1.22690398
#> 48 -0.196776755  0.05950872  1.13826296
#> 49 -0.362394663 -0.19482633  0.60246639
#> 50 -0.064196322  4.60828946  0.07995848
#> 51 -0.337294335  0.43799751  0.30515987
#> 52  0.275624462 -2.83311398 -0.52723643
#> 53 -0.066279407  2.37224579  0.25224244
#> 54 -0.157607423  0.67039750  0.43226803
#> 55  0.011094056  0.33017169 -0.11260898
#> 56 -0.166398481  5.41541832  0.35541494
#> 57 -0.002359809  8.28478096  0.14628855
#> 58 -0.004888880 -3.80667977  0.33771474
#> 59  0.184524979  0.62245323 -0.12200771
#> 60  0.213992853  0.87408528 -0.05625534
#> 61 -0.158510261  1.05968194 -0.01400887
#> 62  0.230468790  2.36287834  0.16982824
#> 63 -0.024420820  1.53739169  0.09967144
#> 64  0.214117992 -1.59190708 -0.23478188
#> 65 -0.113222706 -0.95830176  0.35246193
#> 66 -0.094442259  0.27508288 -0.18210694
#> 67  0.007322462  1.98985834  0.04746307